### 2026-03-04 #10141 @geoHeil
So far there are no scientifc publications out there. JOSS is the first time we submit to a journal. There are not yet papers citing it.

Metaxy is used:
- https://anam.ai/blog/metaxy
- https://georgheiler.com/2026/02/22/metaxy-dagster-slurm-multimodal/

Will add countries soon.

How is Metaxy used?
- Anam is using it to speed up Genai data preparation and Model training
- ASCII is using it for speed up entity extraction and document preprocessing for webcrawl data
- An unnamed enterprise is in the process of adapting it for an audio processing usecase to save on cost but enable experimentation
