### 2026-03-04 #10141 @geoHeil
Addressed in https://github.com/anam-org/metaxy/pull/998
