### 2026-03-05 6656776
fix/externalize-slide-building (#1002)
