### 2026-03-05 f440ffc
use bun only - no npm (#1004)
