+++ .github/workflows/_qa.yml
@@ -329,6 +329,10 @@ jobs:
         uses: actions/checkout@v5
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+        with:
+          bun-version: 1.3.1
       - name: Build Docs
         id: docs
         env:
