### 2026-03-05 06214a8
[chore] pin slides deps (#1008)
