### 2026-03-06 0b23ed1
fix: duckdb memory mode has no readonly (#1017)
resolves: https://github.com/anam-org/metaxy/issues/1016
