### 2026-03-06 07163b1
[cli] add --stale-if argument to `metadata status` (#1006)
