### 2026-03-06 89db2b8
[core] add staleness_predicates (#1005)
This PR adds a new `staleness_predicates` argument to `MetadataStore.resolve_update` which can be used to treat samples as stale based on arbitrary Narwhals expressions
