+++ tests/metadata_stores/test_resolve_update_staleness_predicates.py
@@ -1,0 +1,383 @@
+"""Test resolve_update with staleness_predicates parameter.
+
+Records matching any staleness predicate are treated as stale by resolve_update,
+regardless of version. Predicates are OR'd and applied against existing store records.
+
+skip_comparison=True takes precedence (no predicate evaluation needed).
+
+Tests use downstream features with a resolve-then-write pattern to establish
+properly computed provenance before testing predicate behavior.
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+import narwhals as nw
+import polars as pl
+from metaxy_testing import add_metaxy_provenance_column
+from metaxy_testing.models import SampleFeature, SampleFeatureSpec
+
+from metaxy import FeatureDep, FeatureGraph
+from metaxy.metadata_store.base import MetadataStore
+from metaxy.models.constants import METAXY_PROVENANCE_BY_FIELD
+from metaxy.models.field import FieldSpec
+from metaxy.models.types import FeatureKey, FieldKey
+from metaxy.versioning.types import LazyIncrement
+
+
+def _setup_upstream_and_downstream(
+    store: MetadataStore,
+    graph: FeatureGraph,
+    downstream_cls: Any,
+    upstream_cls: Any,
+    upstream_provenance: list[dict[str, str]],
+    downstream_extra_columns: dict[str, list[Any]] | None = None,
+) -> None:
+    """Write upstream, resolve downstream, augment with extra columns, and write downstream."""
+    sample_uids = list(range(1, len(upstream_provenance) + 1))
+
+    upstream_data = pl.DataFrame(
+        {
+            "sample_uid": sample_uids,
+            METAXY_PROVENANCE_BY_FIELD: upstream_provenance,
+        }
+    )
+    upstream_data = add_metaxy_provenance_column(upstream_data, upstream_cls)
+    store.write(upstream_cls, upstream_data)
+
+    # Resolve downstream to get correctly computed provenance
+    inc = store.resolve_update(downstream_cls)
+    downstream_df = inc.new.to_polars()
+
+    # Add extra user columns by joining on sample_uid (order-independent)
+    if downstream_extra_columns:
+        schema_overrides = {
+            col: pl.Utf8 if all(v is None for v in values) else None for col, values in downstream_extra_columns.items()
+        }
+        extra_df = pl.DataFrame(
+            {"sample_uid": sample_uids, **downstream_extra_columns},
+            schema_overrides={k: v for k, v in schema_overrides.items() if v is not None},
+        )
+        downstream_df = downstream_df.join(extra_df, on="sample_uid", how="left")
+
+    store.write(downstream_cls, downstream_df)
+
+
+def _make_features(graph: FeatureGraph) -> tuple[Any, Any]:
+    """Define a simple upstream/downstream feature pair within the given graph."""
+    with graph.use():
+
+        class UpstreamFeature(
+            SampleFeature,
+            spec=SampleFeatureSpec(
+                key=FeatureKey(["upstream"]),
+                fields=[FieldSpec(key=FieldKey(["value"]), code_version="1")],
+            ),
+        ):
+            pass
+
+        class DownstreamFeature(
+            SampleFeature,
+            spec=SampleFeatureSpec(
+                key=FeatureKey(["downstream"]),
+                deps=[FeatureDep(feature=UpstreamFeature)],
+                fields=[FieldSpec(key=FieldKey(["result"]), code_version="1")],
+            ),
+        ):
+            pass
+
+    return UpstreamFeature, DownstreamFeature
+
+
+def test_predicate_marks_matching_records_as_stale(any_store: MetadataStore) -> None:
+    """Records matching a staleness predicate should appear in inc.stale."""
+    graph = FeatureGraph()
[…]
