### 2026-03-06 a8b7f62
[ray] add staleness_predictates to MetaxyDatasource (#1007)
