### 2026-03-06 34fbc8e
[ci] prevent changing devenv.lock in CI (#1021)
