### 2026-03-06 4407fa2
[ci] always restore devenv.lock in CI (#1022)
