+++ main/slides/2026-introducing-metaxy/dist/assets/modules/shiki-D_M5EFjr.js
@@ -1,1 +1,0 @@
-import{r as He,c as Gt,h as Yt,a as ae,o as j,b as Te,w as we,d as gt,n as Ve,e as le,t as Xt,m as wt,f as Pe,g as J,F as Ut,i as ve,j as Ye,k as ze,l as Kt,p as Zt,q as Qt,s as Jt,u as eo,v as to,x as oo}from"./vue-aalH_SXY.js";const io=["top","right","bottom","left"],Xe=["start","end"],Ue=io.reduce((e,t)=>e.concat(t,t+"-"+Xe[0],t+"-"+Xe[1]),[]),de=Math.min,Q=Math.max,no={left:"right",right:"left",bottom:"top",top:"bottom"};function Re(e,t,o){return Q(e,de(t,o))}function ee(e,t){return typeof e=="function"?e(t):e}function F(e){return e.split("-")[0]}function M(e){return e.split("-")[1]}function vt(e){return e==="x"?"y":"x"}function je(e){return e==="y"?"height":"width"}function V(e){const t=e[0];return t==="t"||t==="b"?"y":"x"}function We(e){return vt(V(e))}function yt(e,t,o){o===void 0&&(o=!1);const i=M(e),n=We(e),s=je(n);let r=n==="x"?i===(o?"end":"start")?"right":"left":i==="start"?"bottom":"top";return t.reference[s]>t.floating[s]&&(r=$e(r)),[r,$e(r)]}function so(e){const t=$e(e);return[ye(e),t,ye(t)]}function ye(e){return e.includes("start")?e.replace("start","end"):e.replace("end","start")}const Ke=["left","right"],Ze=["right","left"],ro=["top","bottom"],ao=["bottom","top"];function lo(e,t,o){switch(e){case"top":case"bottom":return o?t?Ze:Ke:t?Ke:Ze;case"left":case"right":return t?ro:ao;default:return[]}}function po(e,t,o,i){const n=M(e);let s=lo(F(e),o==="start",i);return n&&(s=s.map(r=>r+"-"+n),t&&(s=s.concat(s.map(ye)))),s}function $e(e){const t=F(e);return no[t]+e.slice(t.length)}function co(e){return{top:0,right:0,bottom:0,left:0,...e}}function $t(e){return typeof e!="number"?co(e):{top:e,right:e,bottom:e,left:e}}function ie(e){const{x:t,y:o,width:i,height:n}=e;return{width:i,height:n,top:o,left:t,right:t+i,bottom:o+n,x:t,y:o}}function Qe(e,t,o){let{reference:i,floating:n}=e;const s=V(t),r=We(t),a=je(r),l=F(t),d=s==="y",c=i.x+i.width/2-n.width/2,p=i.y+i.height/2-n.height/2,h=i[a]/2-n[a]/2;let u;switch(l){case"top":u={x:c,y:i.y-n.height};break;case"bottom":u={x:c,y:i.y+i.height};break;case"right":u={x:i.x+i.width,y:p};break;case"left":u={x:i.x-n.width,y:p};break;default:u={x:i.x,y:i.y}}switch(M(t)){case"start":u[r]-=h*(o&&d?-1:1);break;case"end":u[r]+=h*(o&&d?-1:1);break}return u}async function ho(e,t){var o;t===void 0&&(t={});const{x:i,y:n,platform:s,rects:r,elements:a,strategy:l}=e,{boundary:d="clippingAncestors",rootBoundary:c="viewport",elementContext:p="floating",altBoundary:h=!1,padding:u=0}=ee(t,e),f=$t(u),w=a[h?p==="floati
[…]
