### 2026-03-06 ea184da
[core] add rebases (#1000)
