+++ src/metaxy/metadata_store/base.py
@@ -290,7 +290,7 @@ class MetadataStore(ABC):
 
                 !!! info "Required for root features"
                     Metaxy doesn't know how to populate input metadata for root features,
-                    so `samples` argument for **must** be provided for them.
+                    so `samples` argument **must** be provided for them.
 
                 !!! tip
                     For non-root features, use `samples` to customize the automatic upstream loading and field provenance calculation.
@@ -925,6 +925,7 @@ class MetadataStore(ABC):
         feature: CoercibleToFeatureKey,
         df: IntoFrame,
         materialization_id: str | None = None,
+        preserve_feature_version: bool = False,
     ) -> None:
         """
         Write metadata for a feature (append-only by design).
@@ -939,6 +940,10 @@ class MetadataStore(ABC):
             materialization_id: Optional external orchestration ID for this write.
                 Overrides the store's default `materialization_id` if provided.
                 Useful for tracking which orchestration run produced this metadata.
+            preserve_feature_version: If `True`, keep the `metaxy_feature_version`
+                and `metaxy_project_version` already present in `df` instead of
+                overwriting them with the current graph's versions.
+                Is useful when writing [rebased](/guide/concepts/metadata-stores.md#rebases) metadata.
 
         Raises:
             MetadataSchemaError: If DataFrame schema is invalid
@@ -951,6 +956,8 @@ class MetadataStore(ABC):
             - Fallback stores are never used for writes.
 
         """
+        from contextlib import nullcontext
+
         self._check_write_mode()
 
         feature_key = self._resolve_feature_key(feature)
@@ -974,11 +981,13 @@ class MetadataStore(ABC):
 
             raise MetadataSchemaError(f"DataFrame must have '{METAXY_PROVENANCE_BY_FIELD}' column")
 
-        # Add all required system columns
-        # warning: for dataframes that do not match the native MetadataStore implementation
-        # and are missing the METAXY_DATA_VERSION column, this call will lead to materializing the equivalent Polars DataFrame
-        # while calculating the missing METAXY_DATA_VERSION column
-        df_nw = self._add_system_columns(df_nw, feature, materialization_id=materialization_id)
+        ctx = allow_feature_version_override() if preserve_feature_version else nullcontext()
+        with ctx:
+            # Add all required system columns
+            # warning: for dataframes that do not match the native MetadataStore implementation
+            # and are missing the METAXY_DATA_VERSION column, this call will lead to materializing the equivalent Polars DataFrame
+            # while calculating the missing METAXY_DATA_VERSION column
+            df_nw = self._add_system_columns(df_nw, feature, materialization_id=materialization_id)
 
         self._validate_schema(df_nw)
         self._write_feature(feature_key, df_nw)
@@ -1583,20 +1592,38 @@ class MetadataStore(ABC):
             nw.lit(materialization_id or self._materialization_id, dtype=nw.String).alias(METAXY_MATERIALIZATION_ID)
         )
 
-        # Check for missing data_version columns (should come from resolve_update but it's acceptable to just use provenance columns if they are missing)
+        # data_version defaults to provenance but can be user-overridden.
+        # When the column is missing, add it from provenance.
+        # When the column is present, fill null rows from provenance (partial override).
         # Re-fetch columns since df may have been modified
         columns = df.collect_schema().names()
 
         if METAXY_DATA_VERSION_BY_FIELD not in columns:
             df = df.with_columns(nw.col(METAXY_PROVENANCE_BY_FIELD).alias(METAXY_DATA_VERSION_BY_FIELD))
             df = df.with_columns(nw.col(METAXY_PROVENANCE).alias(METAXY_DATA_VERSION))
-        elif METAXY_DATA_VERSION not in columns:
-            df = self.hash_struct_version_column(
-                plan=self._resolve_feature_plan(feature_key),
[…]
