### 2026-03-06 e20c3c2
[docs] re-arrange metadata store docs (#1023)
