### 2026-03-06 0d19ee9
fix: refine duckdb testing (#1019)
minor refactorings for duckdb testing
