+++ tests/ext/dagster/test_io_manager.py
@@ -168,6 +168,43 @@ class TestMetaxyIOManagerHandleOutput:
         # Should have 2 rows
         assert metadata["dagster/row_count"].value == 2
 
+    def test_handle_output_with_none_in_memory_duckdb_does_not_fail(
+        self,
+        upstream_feature: type[mx.BaseFeature],
+        duckdb_in_memory_resources: dict[str, Any],
+        instance: dg.DagsterInstance,
+    ):
+        """Regression test for issue #1016 with in-memory DuckDB resources."""
+
+        @dg.asset(
+            metadata={"metaxy/feature": "features/upstream"},
+            io_manager_key="metaxy_io_manager",
+        )
+        def my_asset(store: dg.ResourceParam[mx.MetadataStore]):
+            with store.open("w"):
+                store.write(
+                    upstream_feature,
+                    pl.DataFrame(
+                        {
+                            "id": ["in_memory_1", "in_memory_2"],
+                            "metaxy_provenance_by_field": [
+                                {"value": "v1"},
+                                {"value": "v2"},
+                            ],
+                        }
+                    ),
+                )
+            return None
+
+        result = dg.materialize(
+            [my_asset],
+            resources=duckdb_in_memory_resources,
+            instance=instance,
+        )
+
+        assert result.success
+        assert len(result.get_asset_materialization_events()) == 1
+
     def test_handle_output_with_metaxy_output_writes_data(
         self,
         upstream_feature: type[mx.BaseFeature],
@@ -216,44 +253,6 @@ class TestMetaxyIOManagerHandleOutput:
         )
         assert result2.success
         assert captured_data["rows"] == 3
-
-    def test_handle_output_with_none_in_memory_duckdb_does_not_fail(
-        self,
-        upstream_feature: type[mx.BaseFeature],
-        duckdb_in_memory_resources: dict[str, Any],
-        instance: dg.DagsterInstance,
-    ):
-        """Regression test for issue #1016 with in-memory DuckDB in read mode."""
-
-        @dg.asset(
-            metadata={"metaxy/feature": "features/upstream"},
-            io_manager_key="metaxy_io_manager",
-        )
-        def my_asset(store: dg.ResourceParam[mx.MetadataStore]):
-            # Write data inside the asset and return None to trigger metadata logging path.
-            with store.open("w"):
-                store.write(
-                    upstream_feature,
-                    pl.DataFrame(
-                        {
-                            "id": ["in_memory_1", "in_memory_2"],
-                            "metaxy_provenance_by_field": [
-                                {"value": "v1"},
-                                {"value": "v2"},
-                            ],
-                        }
-                    ),
-                )
-            return None
-
-        result = dg.materialize(
-            [my_asset],
-            resources=duckdb_in_memory_resources,
-            instance=instance,
-        )
-
-        assert result.success
-        assert len(result.get_asset_materialization_events()) == 1
 
 
 class TestMetaxyIOManagerLoadInput:

[…]
