+++ .claude-plugin/skills/metaxy/examples/testing.md
@@ -1,6 +1,6 @@
 # Testing Metaxy Code
 
-See full documentation: https://anam-org.github.io/metaxy/guide/concepts/testing/
+See full documentation: https://docs.metaxy.io/latest/guide/concepts/lifecycle/testing/
 
 ## Graph Isolation
 
@@ -25,19 +25,22 @@ def test_my_feature(isolated_graph):
         pass
 
     # Feature is registered to isolated graph, not global
-    assert mx.FeatureGraph.get_active().get_feature("test/feature") is not None
+    assert mx.FeatureGraph.get().get_feature("test/feature") is not None
 ```
 
 ## Configuration Isolation
 
 ```python
 import metaxy as mx
-from metaxy.ext.metadata_stores.delta import DeltaMetadataStore
 
 
 def test_with_custom_config(tmp_path):
-    store = DeltaMetadataStore(root_path=tmp_path / "delta_test")
-    with mx.MetaxyConfig(stores={"test": store}).use() as config:
+    with mx.MetaxyConfig(
+        stores={"test": mx.StoreConfig(
+            type="metaxy.ext.metadata_stores.delta.DeltaMetadataStore",
+            config={"root_path": str(tmp_path / "delta_test")},
+        )}
+    ).use() as config:
         store = config.get_store("test")
         # test with isolated config and store
 ```
@@ -47,14 +50,17 @@ def test_with_custom_config(tmp_path):
 ```python
 import pytest
 import metaxy as mx
-from metaxy.ext.metadata_stores.delta import DeltaMetadataStore
 
 
 @pytest.fixture
 def metaxy_env(tmp_path):
     with mx.FeatureGraph().use():
-        store = DeltaMetadataStore(root_path=tmp_path / "delta_test")
-        with mx.MetaxyConfig(stores={"test": store}).use() as config:
+        with mx.MetaxyConfig(
+            stores={"test": mx.StoreConfig(
+                type="metaxy.ext.metadata_stores.delta.DeltaMetadataStore",
+                config={"root_path": str(tmp_path / "delta_test")},
+            )}
+        ).use() as config:
             yield config
 
 
