### 2026-03-06 8e3c7ed
Deployed dfff3415 to 0.1.2 with MkDocs 1.6.1 and mike 2.1.3
