+++ 0.1.2/slides/2026-introducing-metaxy/dist/assets/md-Cj42_pAF.js
@@ -1,0 +1,1 @@
+import{d as v,e as x,f as b}from"./modules/unplugin-icons-BaqmoMqW.js";import{d as h,B as y,f as k,o as u,g as o,j as w,k as C,b as $,w as B,D as m,z as a,e as l,v as P,x as z}from"./modules/vue-D2_2azlL.js";import{u as p,f as S}from"./slidev/context-CXJvKEzn.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";const A=""+new URL("../img/metaxy.svg",import.meta.url).href;function d(t){return t.startsWith("/")?"./"+t.slice(1):t}function E(t,n=!1){const s=t&&["#","rgb","hsl"].some(i=>t.indexOf(i)===0),r={background:s?t:void 0,color:t&&!s?"white":void 0,backgroundImage:s?void 0:t?n?`linear-gradient(#0005, #0008), url(${d(t)})`:`url("${d(t)}")`:void 0,backgroundRepeat:"no-repeat",backgroundPosition:"center",backgroundSize:"cover"};return r.background||delete r.background,r}const I={class:"my-auto w-full"},N=h({__name:"cover",props:{background:{default:""}},setup(t){p();const n=t,s=y(()=>E(n.background,!0));return(r,i)=>(u(),k("div",{class:"slidev-layout cover",style:C(s.value)},[o("div",I,[w(r.$slots,"default")])],4))}}),R={class:"abs-br m-6 text-xl"},V={href:"https://github.com/anam-org/metaxy/tree/main/slides/2026-introducing-metaxy/introducing-metaxy.md",target:"_blank",class:"slidev-icon-btn"},F={__name:"introducing-metaxy.md__slidev_1",setup(t){const{$slidev:n,$clicksContext:s,$frontmatter:r}=p();return s.setup(),(i,e)=>{const _=v,g=x,f=b;return u(),$(N,P(z(a(S)(a(r),0))),{default:B(()=>[e[3]||(e[3]=o("img",{src:A,class:"mx-auto mb-4 w-24 h-24"},null,-1)),e[4]||(e[4]=o("h1",null,[m("Perfecting the art of doing "),o("span",{style:{color:"#696FCC"}},"nothing")],-1)),o("div",{onClick:e[0]||(e[0]=(...c)=>a(n).nav.next&&a(n).nav.next(...c)),class:"mt-12 py-1","hover:bg":"white op-10"},[e[2]||(e[2]=m(" Accelerating multimodal AI experimentation without massive cost. ",-1)),l(_)]),o("div",R,[o("button",{onClick:e[1]||(e[1]=c=>a(n).nav.openInEditor()),title:"Open in Editor",class:"slidev-icon-btn"},[l(g)]),o("a",V,[l(f)])])]),_:1},16)}}};export{F as default};
