+++ 0.1.2/slides/2026-introducing-metaxy/dist/assets/md-BMrbsxsW.js
@@ -1,0 +1,1 @@
+import{_ as d}from"./slidev/VClick-CMerNFb9.js";import{_ as c}from"./slidev/VClicks-kWLYTJrV.js";import{b as p,o as m,w as n,g as t,e as o,D as e,v as f,x as _,z as r}from"./modules/vue-D2_2azlL.js";import{I as g}from"./slidev/center-CLDVn-hs.js";import{u as h,f as v}from"./slidev/context-CXJvKEzn.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";const B={__name:"introducing-metaxy.md__slidev_3",setup(x){const{$clicksContext:a,$frontmatter:s}=h();return a.setup(),(y,l)=>{const u=c,i=d;return m(),p(g,f(_(r(v)(r(s),2))),{default:n(()=>[l[2]||(l[2]=t("h1",null,"Why Existing Tools Fall Short",-1)),o(u,null,{default:n(()=>[...l[0]||(l[0]=[t("table",null,[t("thead",null,[t("tr",null,[t("th",null,"Tool"),t("th",null,"What it does"),t("th",null,"The gap")])]),t("tbody",null,[t("tr",null,[t("td",null,[t("strong",null,"Dagster / Airflow")]),t("td",null,"Orchestrate at the asset/table level"),t("td",null,[e("Can’t see which "),t("em",null,"fields"),e(" changed; partitions don’t scale to millions of samples")])]),t("tr",null,[t("td",null,"DVC"),t("td",null,"Version files as opaque artifacts"),t("td",null,"No record-level or field-level tracking")]),t("tr",null,[t("td",null,"Feast"),t("td",null,"Serve precomputed features at inference"),t("td",null,"No dependency graph, no change detection")]),t("tr",null,[t("td",null,"Hamilton"),t("td",null,"Column-level lineage via Python functions"),t("td",null,"Table granularity, not record granularity")]),t("tr",null,[t("td",null,[t("strong",null,"LanceDB")]),t("td",null,"Vector storage with versioned datasets"),t("td",null,"No cross-feature dependency tracking")])])],-1)])]),_:1}),o(i,null,{default:n(()=>[...l[1]||(l[1]=[t("div",{"mt-6":"","text-center":"","text-lg":""},[t("p",null,[e("All of these operate at "),t("strong",null,"table or file granularity"),e(". None track dependencies at the "),t("strong",null,"field level per record"),e(".")])],-1)])]),_:1})]),_:1},16)}}};export{B as default};
