+++ 0.1.2/slides/2026-introducing-metaxy/dist/assets/flowDiagram-TD5NMDKU-DmKdvmny.js
@@ -1,0 +1,162 @@
+import{g as q1}from"./chunk-SCLVSU6P-VHFMvI2b.js";import{a as m,o as O1,l as $,d as bt,e as Ct,p as H1,q as X1,u as i1,c as Q1,s as J1,t as Z1,b as $1,g as te,v as ee,m as se,x as ie,J as re,y as ae,z as s1,A as ne,B as ue,C as oe,D as le}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";import{g as ce}from"./chunk-VH2OSJNQ-CWsncCGH.js";import{s as he}from"./chunk-WDWNAAEG-C-FBSnpL.js";import"./modules/vue-D2_2azlL.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";import"./lz-string-BCxBxReP.js";import"./modules/file-saver-B7oFTzqn.js";var de="flowchart-",Gt,pe=(Gt=class{constructor(){this.vertexCounter=0,this.config=bt(),this.vertices=new Map,this.edges=[],this.classes=new Map,this.subGraphs=[],this.subGraphLookup=new Map,this.tooltips=new Map,this.subCount=0,this.firstGraphFlag=!0,this.secCount=-1,this.posCrossRef=[],this.funs=[],this.setAccTitle=Q1,this.setAccDescription=J1,this.setDiagramTitle=Z1,this.getAccTitle=$1,this.getAccDescription=te,this.getDiagramTitle=ee,this.funs.push(this.setupToolTips.bind(this)),this.addVertex=this.addVertex.bind(this),this.firstGraph=this.firstGraph.bind(this),this.setDirection=this.setDirection.bind(this),this.addSubGraph=this.addSubGraph.bind(this),this.addLink=this.addLink.bind(this),this.setLink=this.setLink.bind(this),this.updateLink=this.updateLink.bind(this),this.addClass=this.addClass.bind(this),this.setClass=this.setClass.bind(this),this.destructLink=this.destructLink.bind(this),this.setClickEvent=this.setClickEvent.bind(this),this.setTooltip=this.setTooltip.bind(this),this.updateLinkInterpolate=this.updateLinkInterpolate.bind(this),this.setClickFun=this.setClickFun.bind(this),this.bindFunctions=this.bindFunctions.bind(this),this.lex={firstGraph:this.firstGraph.bind(this)},this.clear(),this.setGen("gen-2")}sanitizeText(i){return se.sanitizeText(i,this.config)}lookUpDomId(i){for(const r of this.vertices.values())if(r.id===i)return r.domId;return i}addVertex(i,r,a,n,l,g,c={},b){if(!i||i.trim().length===0)return;let u;if(b!==void 0){let k;b.includes(`
+`)?k=b+`
+`:k=`{
+`+b+`
[…]
