+++ 0.1.2/slides/2026-introducing-metaxy/dist/assets/slidev/context-CXJvKEzn.js
@@ -1,0 +1,1 @@
+import{N as s,O as t,r as d,B as m,P as $}from"../modules/vue-D2_2azlL.js";import{i as f,b as C,d as x,e as S,f as p,h as v,j as R,k as T,H as E,F}from"../index-B2pMgQ-r.js";function D(){const n=t(S),o=s(n,"nav"),e=t(f).value,c=s(e,"current"),i=t(T),a=t(v),r=t(p,{}),l=t(R,void 0),u=t(C,d(1)),j=t(x,m(()=>1));return{$slidev:n,$nav:o,$clicksContext:e,$clicks:c,$page:i,$route:l,$renderContext:a,$frontmatter:r,$scale:u,$zoom:j}}function L(n,o){return{...$(n,o===0?E:F),frontmatter:n}}export{L as f,D as u};
