+++ 0.1.2/slides/2026-introducing-metaxy/dist/assets/slidev/IconButton.vue_vue_type_script_setup_true_lang-DOdjJVYn.js
@@ -1,0 +1,1 @@
+import{d as n,B as l,b as c,w as i,C as r,o as u,g as o,j as d,t as m,n as p}from"../modules/vue-D2_2azlL.js";const f={class:"sr-only"},v=n({__name:"IconButton",props:{title:{},icon:{},as:{},to:{}},setup(t){const e=t,s=l(()=>e.as||(e.to?"router-link":"button"));return(a,B)=>(u(),c(r(s.value),{class:"slidev-icon-btn",title:t.title,to:t.to},{default:i(()=>[o("span",f,m(t.title),1),d(a.$slots,"default",{},()=>[o("div",{class:p(t.icon)},null,2)])]),_:3},8,["title","to"]))}});export{v as _};

[…]
