### 2026-03-07 8cb0b5e
refine ducklake example
