### 2026-03-07 79b98ed
[docs] fix inheritance docs (#1026)
