### 2026-03-09 203c4c7
[cli] move `metaxy graph history` to `metaxy history` (#1029)
