+++ src/metaxy/cli/app.py
@@ -101,6 +101,7 @@ app.command("metaxy.cli.config:app", name="config")
 app.command("metaxy.cli.describe:app", name="describe")
 app.command("metaxy.cli.migrations:app", name="migrations")
 app.command("metaxy.cli.graph:app", name="graph")
+app.command("metaxy.cli.history:app", name="history")
 app.command("metaxy.cli.graph_diff:app", name="graph-diff")
 app.command("metaxy.cli.list:app", name="list")
 app.command("metaxy.cli.metadata:app", name="metadata")
