### 2026-03-10 2873f0a
[mcp] add some info to the MCP server (#1028)
