### 2026-03-10 3d199a9
fix/sybil-snippet-expansion
