+++ docs/conftest.py
@@ -3,11 +3,12 @@
 Configures Sybil to parse and test markdown code blocks (```python and ```py)
 found in the docs/ directory.
 
-Code blocks using --8<-- snippet includes are automatically skipped since
-they are tested via their source files.
+Code blocks using --8<-- snippet includes are expanded before evaluation so
+docs execute the same example source that MkDocs renders.
 """
 
 import re
+from pathlib import Path
 
 from metaxy.models.feature import FeatureGraph
 from sybil import Sybil
@@ -21,6 +22,13 @@ from sybil.parsers.markdown.skip import SkipParser
 from sybil.typing import Evaluator
 
 import metaxy as mx
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+SNIPPET_BASE_PATHS = (
+    REPO_ROOT / "examples",
+    REPO_ROOT / "docs" / "snippets",
+    REPO_ROOT / "docs",
+)
 
 
 # Workaround for pytest-cases compatibility issue
@@ -124,17 +132,138 @@ class CodeBlockWithAttributesParser(AbstractCodeBlockParser):
         )
 
 
-class SnippetSkippingEvaluator(PythonEvaluator):
-    """PythonEvaluator that skips code blocks containing --8<-- snippet includes."""
+class SnippetExpandingEvaluator(PythonEvaluator):
+    """PythonEvaluator that expands MkDocs snippet includes before execution."""
 
-    # Pattern to detect snippet includes
-    SNIPPET_PATTERN = re.compile(r'--8<--\s*"')
+    SNIPPET_PATTERN = re.compile(r'^(?P<indent>\s*)--8<--\s*"(?P<target>[^"]+)"\s*$')
+    SECTION_START_PATTERN = re.compile(r"#\s*--8<--\s*\[start:(?P<name>[^\]]+)\]\s*$")
+    SECTION_END_PATTERN = re.compile(r"#\s*--8<--\s*\[end:(?P<name>[^\]]+)\]\s*$")
+
+    def _resolve_snippet_path(self, target_path: str, *, document_path: Path) -> Path:
+        candidate_paths = []
+
+        if Path(target_path).is_absolute():
+            candidate_paths.append(Path(target_path))
+        else:
+            candidate_paths.append((document_path.parent / target_path).resolve())
+            candidate_paths.extend(
+                (base_path / target_path).resolve() for base_path in SNIPPET_BASE_PATHS
+            )
+
+        for candidate_path in candidate_paths:
+            if candidate_path.exists():
+                return candidate_path
+
+        searched = ", ".join(str(path) for path in candidate_paths)
+        raise FileNotFoundError(
+            f"Unable to resolve snippet path {target_path!r}; searched: {searched}"
+        )
+
+    def _extract_section(
+        self, source_lines: list[str], section_name: str, snippet_path: Path
+    ) -> list[str]:
+        start_index: int | None = None
+
+        for index, line in enumerate(source_lines):
+            start_match = self.SECTION_START_PATTERN.match(line)
+            if start_match and start_match.group("name") == section_name:
+                start_index = index + 1
+                break
+
+        if start_index is None:
+            raise ValueError(
+                f"Snippet section {section_name!r} not found in {snippet_path}"
+            )
+
+        for index in range(start_index, len(source_lines)):
+            end_match = self.SECTION_END_PATTERN.match(source_lines[index])
+            if end_match and end_match.group("name") == section_name:
+                return source_lines[start_index:index]
+
+        raise ValueError(
+            f"Snippet section {section_name!r} was not closed in {snippet_path}"
+        )
+
+    def _extract_range(
+        self,
+        source_lines: list[str],
+        start_line: int,
[…]
