### 2026-03-10 c0b41a0
[iceberg] Add Apache Iceberg metadata store backend
Closes #1032

Made-with: Cursor
