+++ mkdocs.yml
@@ -261,6 +261,7 @@ nav:
         - PostgreSQL: integrations/metadata-stores/databases/postgresql.md
       - Storage:
         - integrations/metadata-stores/storage/index.md
+        - Apache Iceberg: integrations/metadata-stores/storage/iceberg.md
         - Delta Lake: integrations/metadata-stores/storage/delta.md
         - DuckLake: integrations/metadata-stores/storage/ducklake.md
     - Compute:
