### 2026-03-10 27609fe
[iceberg] Add Apache Iceberg metadata store backend
Closes #1032

Made-with: Cursor
