+++ src/metaxy/ext/metadata_stores/iceberg.py
@@ -1,0 +1,269 @@
+"""Apache Iceberg metadata store implemented with PyIceberg."""
+
+from __future__ import annotations
+
+from collections.abc import Iterator, Sequence
+from contextlib import contextmanager
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Literal, overload
+
+if TYPE_CHECKING:
+    from pyiceberg.catalog import Catalog
+
+import narwhals as nw
+import polars as pl
+from narwhals.typing import Frame
+from pydantic import Field
+
+from metaxy._decorators import public
+from metaxy._utils import collect_to_polars
+from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
+from metaxy.metadata_store.types import AccessMode
+from metaxy.metadata_store.utils import is_local_path
+from metaxy.models.plan import FeaturePlan
+from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.versioning.polars import PolarsVersioningEngine
+from metaxy.versioning.types import HashAlgorithm
+
+
+@public
+class IcebergMetadataStoreConfig(MetadataStoreConfig):
+    """Configuration for IcebergMetadataStore.
+
+    Example:
+        ```toml title="metaxy.toml"
+        [stores.dev]
+        type = "metaxy.ext.metadata_stores.iceberg.IcebergMetadataStore"
+
+        [stores.dev.config]
+        warehouse = "/path/to/warehouse"
+        namespace = "metaxy"
+
+        [stores.dev.config.catalog_properties]
+        type = "sql"
+        ```
+    """
+
+    warehouse: str | Path = Field(
+        description="Warehouse directory or URI where Iceberg tables are stored.",
+    )
+    namespace: str = Field(
+        default="metaxy",
+        description="Iceberg namespace for feature tables.",
+    )
+    catalog_name: str = Field(
+        default="metaxy",
+        description="Name of the Iceberg catalog.",
+    )
+    catalog_properties: dict[str, str] | None = Field(
+        default=None,
+        description="Properties passed to pyiceberg.catalog.load_catalog.",
+    )
+    layout: Literal["flat", "nested"] = Field(
+        default="nested",
+        description="Table naming layout ('nested' or 'flat').",
+    )
+
+
+@public
+class IcebergMetadataStore(MetadataStore):
+    """Apache Iceberg metadata store backed by [PyIceberg](https://py.iceberg.apache.org/).
+
+    Stores feature metadata in Iceberg tables managed by a PyIceberg catalog.
+    Uses the Polars versioning engine for provenance calculations.
+
+    Example:
+
+        ```py
+        from metaxy.ext.metadata_stores.iceberg import IcebergMetadataStore
+
+        store = IcebergMetadataStore(
+            warehouse="/path/to/warehouse",
+            namespace="metaxy",
+        )
+        ```
+    """
+
+    _should_warn_auto_create_tables = False
+    versioning_engine_cls = PolarsVersioningEngine
+
+    def __init__(
+        self,
+        warehouse: str | Path,
+        *,
+        namespace: str = "metaxy",
+        catalog_name: str = "metaxy",
+        catalog_properties: dict[str, str] | None = None,
+        fallback_stores: list[MetadataStore] | None = None,
+        layout: Literal["flat", "nested"] = "nested",
+        **kwargs: Any,
+    ) -> None:
+        if layout not in ("flat", "nested"):
+            raise ValueError(f"Invalid layout: {layout}. Must be 'flat' or 'nested'.")
+
+        self.namespace = namespace
+        self.catalog_name = catalog_name
+        self.layout = layout
+        self._catalog: Catalog | None = None
+
[…]
