### 2026-03-10 29cd69e
[chore] update trove classifiers (#1035)
