### 2026-03-10 b536777
[cli] ignore packages that are not dependencies in `metaxy lock` (#1036)
## Motivation

This fix removes the need to run `metaxy lock` with `uv run --isolated` in `uv` workspaces.

## Changelog

`metaxy lock` now avoids loading feature definitions from `metaxy.project` entrypoints of Python packages that aren't dependencies (direct or transitive) of the current Python package when creating the lock file.
