### 2026-03-11 1b6f4f8
fix/duckdb-extension-loading-from-community-repo
