+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "metaxy"
-version = "0.1.2"
+version = "0.1.3"
 description = "Pluggable versioned metadata management for incremental multimodal pipelines."
 readme = "README.md"
 authors = [
