### 2026-03-11 f45f33f
[chore] remove allow_feature_version_override (#1040)
Resolve #1001
