+++ 0.1.3/slides/2026-introducing-metaxy/dist/assets/dagre-ESHLYXRB-BOvW1GRl.js
@@ -1,0 +1,4 @@
+import{l as Y}from"./chunk-LXPT3ST6-hWA4TcXo.js";import{G as k}from"./chunk-4QEEASX4-EybeeKIx.js";import{a as g,bx as H,by as V,bz as z,bA as U,l as i,d as $,bB as q,bC as K,bf as Q,bl as W,bg as P,be as Z,bD as I,bE as L,bF as ee}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";import{i as N,B as ne,b as A}from"./chunk-ZZTYOBSU-CHxtuk5s.js";import"./modules/vue-D2_2azlL.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";import"./lz-string-BCxBxReP.js";import"./modules/file-saver-B7oFTzqn.js";function h(e){var t={options:{directed:e.isDirected(),multigraph:e.isMultigraph(),compound:e.isCompound()},nodes:J(e),edges:G(e)};return N(e.graph())||(t.value=ne(e.graph())),t}g(h,"write");function J(e){return A(e.nodes(),function(t){var n=e.node(t),a=e.parent(t),r={v:t};return N(n)||(r.value=n),N(a)||(r.parent=a),r})}g(J,"writeNodes");function G(e){return A(e.edges(),function(t){var n=e.edge(t),a={v:t.v,w:t.w};return N(t.name)||(a.name=t.name),N(n)||(a.value=n),a})}g(G,"writeEdges");var d=new Map,p=new Map,R=new Map,te=g(()=>{p.clear(),R.clear(),d.clear()},"clear"),D=g((e,t)=>{const n=p.get(t)||[];return i.trace("In isDescendant",t," ",e," = ",n.includes(e)),n.includes(e)},"isDescendant"),se=g((e,t)=>{const n=p.get(t)||[];return i.info("Descendants of ",t," is ",n),i.info("Edge is ",e),e.v===t||e.w===t?!1:n?n.includes(e.v)||D(e.v,t)||D(e.w,t)||n.includes(e.w):(i.debug("Tilt, ",t,",not in descendants"),!1)},"edgeInCluster"),T=g((e,t,n,a)=>{i.warn("Copying children of ",e,"root",a,"data",t.node(e),a);const r=t.children(e)||[];e!==a&&r.push(e),i.warn("Copying (nodes) clusterId",e,"nodes",r),r.forEach(o=>{if(t.children(o).length>0)T(o,t,n,a);else{const l=t.node(o);i.info("cp ",o," to ",a," with parent ",e),n.setNode(o,l),a!==t.parent(o)&&(i.warn("Setting parent",o,t.parent(o)),n.setParent(o,t.parent(o))),e!==a&&o!==e?(i.debug("Setting parent",o,e),n.setParent(o,e)):(i.info("In copy ",e,"root",a,"data",t.node(e),a),i.debug("Not Setting parent for node=",o,"cluster!==rootId",e!==a,"node!==clusterId",o!==e));const u=t.edges(o);i.debug("Copying Edges",u),u.forEach(c=>{i.info("Edge",c);const m=t.edge(c.v,c.w,c.name);i.info("Edge data",m,a);try{se(c,a)?(i.info("Copying as ",c.v,c.w,m,c.name),n.setEdge(c.v,c.w,m,c.name),i.info("newGraph edges ",n.edges(),n.edge(n.edges()[0]))):i.info("Skipping copy of edge ",c.v,"-->",c.w," rootId: ",a," clusterId:",e)}catch(v){i.error(v)}})}i.debug("Removing node",o),t.removeNode(o)})},"copy"),M=g
[…]
