+++ 0.1.3/slides/2026-introducing-metaxy/dist/assets/slidev/Modal.vue_vue_type_script_setup_true_lang-Cjs_vcHA.js
@@ -1,0 +1,1 @@
+import{d as r,b as m,f as c,i as u,z as f,A as p,g as t,j as b,n as k,K as v,o as s}from"../modules/vue-D2_2azlL.js";const g={key:0,class:"fixed top-0 bottom-0 left-0 right-0 grid z-modal"},z=r({__name:"Modal",props:{modelValue:{default:!1},class:{default:""}},emits:["update:modelValue"],setup(l,{emit:n}){const e=l,a=p(e,"modelValue",n);function d(){a.value=!1}return(i,o)=>(s(),m(v,null,[f(a)?(s(),c("div",g,[t("div",{bg:"black opacity-80",class:"absolute top-0 bottom-0 left-0 right-0 -z-1",onClick:o[0]||(o[0]=_=>d())}),t("div",{class:k(["m-auto rounded-md bg-main shadow",e.class]),"dark:border":"~ main"},[b(i.$slots,"default")],2)])):u("v-if",!0)],1024))}});export{z as _};

[…]
