### 2026-03-11 0cc93df
refine ducklake example (#1025)
Redo DuckLake example to showcase reading and writing data
