### 2026-03-11 0d1a0fa
generalize
