### 2026-03-11 634b28b
generalize fix
