+++ examples/example-ducklake/.example.yaml
@@ -15,12 +15,12 @@ scenarios:
         timeout: 30.0
 
       - type: "assert_output"
-        description: "Verify the output includes successful Metaxy writes and DuckLake tables"
+        description: "Verify the output includes successful Metaxy writes and native Metaxy reads"
         returncode: 0
         contains:
           - "DuckLake pipeline"
           - "Wrote 2 rows"
-          - "SHOW ALL TABLES after Metaxy wrote metadata:"
-          - "ducklake_column"
+          - "Discovered Metaxy features:"
+          - "examples/ducklake_demo"
           - "Created feature table: examples__ducklake_demo"
           - "clip_001: s3://demo-bucket/processed/clip_001.parquet"
