+++ 0.1.4/slides/2026-introducing-metaxy/dist/assets/md-JqQiAHai.js
@@ -1,0 +1,1 @@
+import{_ as i}from"./slidev/VClicks-kWLYTJrV.js";import{_ as p}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";import{b as c,o as m,w as n,g as o,e as l,D as e,m as d,v as g,x as f,z as s}from"./modules/vue-D2_2azlL.js";import{I as _}from"./slidev/center-CLDVn-hs.js";import{u as Y,f as F}from"./slidev/context-CXJvKEzn.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";import"./lz-string-BCxBxReP.js";import"./modules/file-saver-B7oFTzqn.js";const v={__name:"introducing-metaxy.md__slidev_6",setup(Q){const{$clicksContext:r,$frontmatter:a}=Y();return r.setup(),(R,t)=>{const u=p,A=i;return m(),c(_,g(f(s(F)(s(a),5))),{default:n(()=>[t[1]||(t[1]=o("h1",null,"What Changes, What Doesn’t",-1)),t[2]||(t[2]=o("p",null,[o("strong",null,"Scenario:"),e(" You bump the "),o("code",null,"audio"),e(" code version from "),o("code",null,'"1"'),e(" to "),o("code",null,'"2"'),e(".")],-1)),l(u,d({"code-lz":"OYJwhgDgFgBAMgJQFA1TAagbQEToJYAmApgPYA8ARiAPQB8ltYArgXiQFwwBuATGdRVoxAYNSB4P8o1aAM3ABbIgGdOXAIzYAujAC0WoQGEcekCQgS6DZqw7c+AoWLPS5i5WvUo0e7bpgAxACI4vmAAxkT+RAAuRCGRbAB2jlKhLtwqMICg5BoeqOjeQgDKACpFOAUQRDFQRSRFRAAekY4M0Y3KtoIiotk5MAqRAJ4ANkQYMFJ4Q0PsAMQADFIqAOw8YAA0ISRDJCCzUgAcyVIha/3GANZEswBsAJzXvnp6p5EXRFoA7oSRUOw8EPVev1hqMvBMprMFstVhstjs9ocwMcXm89gBWW5EOYUFEkS6fb6/f6AtB9QYjPz+caTabzRYrdabba7GYHI4nM54q4zHg8EJotFEXH4r4EH5/AFA8mjYpFakQunQxlwllspEc15c9GY7HC96i8XEoA==="},{scale:.8}),null,16),l(A,null,{default:n(()=>[...t[0]||(t[0]=[o("ul",null,[o("li",null,[o("strong",null,"FaceDetection"),e(": depends on "),o("code",null,"Crop.frames"),e(" only. "),o("strong",null,"Untouched."),e(" No GPU cost.")]),o("li",null,[o("strong",null,"SpeechToText"),e(": depends on "),o("code",null,"Video.audio"),e(". "),o("strong",null,"Marked for reprocessing.")]),o("li",null,"You just saved half your GPU bill.")],-1)])]),_:1})]),_:1},16)}}};export{v as default};
