+++ 0.1.4/slides/2026-introducing-metaxy/dist/assets/slidev/center-CLDVn-hs.js
@@ -1,0 +1,1 @@
+import{f as t,o,g as s,j as c}from"../modules/vue-D2_2azlL.js";import{_ as r}from"../index-B2pMgQ-r.js";const a={},n={class:"slidev-layout center h-full grid place-content-center"},l={class:"my-auto"};function d(e,_){return o(),t("div",n,[s("div",l,[c(e.$slots,"default")])])}const m=r(a,[["render",d]]);export{m as I};
