+++ 0.1.4/slides/2026-introducing-metaxy/dist/assets/index-B2pMgQ-r.js
@@ -1,0 +1,90 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["./md-Cj42_pAF.js","./modules/unplugin-icons-BaqmoMqW.js","./modules/vue-D2_2azlL.js","./slidev/context-CXJvKEzn.js","./modules/shiki-bZSEl0vS.js","./modules/shiki-DVQfPmV_.css","./md-OhnM57yS.js","./slidev/VClick-CMerNFb9.js","./slidev/VClicks-kWLYTJrV.js","./slidev/default-DMaQI0-d.js","./md-BMrbsxsW.js","./slidev/center-CLDVn-hs.js","./md-j1mY04cH.js","./slidev/CodeBlockWrapper.vue_vue_type_script_setup_true_lang-FpMJoXbT.js","./slidev/TitleIcon.vue_vue_type_script_setup_true_lang-CyxFj2c-.js","./md-BS4GpBhh.js","./md-JqQiAHai.js","./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js","./lz-string-BCxBxReP.js","./modules/file-saver-B7oFTzqn.js","./md-g0s03i6h.js","./md-D7mj9DLL.css","./slidev/md-CRrTuXyh.js","./md-CwZeasl4.js","./md-Drxw-h1D.js","./md-DN5x4nm2.js","./slidev/entry-DGsRyvlP.js","./entry-15QH0Q6e.css","./slidev/overview-By4IQRoM.js","./slidev/NoteDisplay.vue_vue_type_style_index_0_lang-B-yBMlQt.js","./NoteDisplay-R4MfDMfa.css","./slidev/SlideWrapper-C87JTbq8.js","./SlideWrapper-DoEfutxU.css","./slidev/IconButton.vue_vue_type_script_setup_true_lang-DOdjJVYn.js","./slidev/useSlideInfo-DSFz2w09.js","./slidev/notes-BqfF6rY9.js","./slidev/TimerBar.vue_vue_type_script_setup_true_lang-BE3aPCgI.js","./slidev/Modal.vue_vue_type_script_setup_true_lang-Cjs_vcHA.js","./slidev/notes-edit-CacZzqU2.js","./slidev/presenter-2EkWCDdz.js","./slidev/shortcuts-Cyu2ygbd.js","./shortcuts--Adx2Pgt.css","./slidev/DrawingControls.vue_vue_type_style_index_0_lang-kjgh92NS.js","./DrawingControls-Cxk9a9ub.css","./presenter-C_fGJIsk.css","./slidev/play-C1490e_f.js","./play-D_7yfuao.css","./slidev/404-BCHbFmur.js","./404-n_Eqdu3W.css"])))=>i.map(i=>d[i]);
[…]

[…]
