+++ CONTRIBUTING.md
@@ -12,6 +12,38 @@ We are using [Devenv](https://devenv.sh/) to manage the development environment,
 
 To reinstall Python dependencies, run `uv sync --all-extras`.
 
+## Commit Messages
+
+We use [Conventional Commits](https://www.conventionalcommits.org/). PRs are squash-merged, so the **PR title becomes the commit message** in `main`. This means the PR title must follow conventional commit format — individual commit messages on the branch don't matter. A CI check enforces this.
+
+```
+<type>(<optional scope>): <description>
+```
+
+| Type | Purpose |
+|------|---------|
+| `feat` | New feature |
+| `fix` | Bug fix |
+| `docs` | Documentation only |
+| `refactor` | Code change that neither fixes a bug nor adds a feature |
+| `perf` | Performance improvement |
+| `test` | Adding or updating tests |
+| `chore` | Maintenance (deps, CI, releases) |
+
+Append `!` after the type/scope for breaking changes.
+
+Examples:
+
+```
+feat: add BigQuery metadata store support
+feat(cli): add --format flag to metaxy inspect
+fix: resolve race condition in metadata write
+refactor: simplify version calculation logic
+fix!: rename MetadataStore.push to MetadataStore.sync
+```
+
+Types `chore`, `ci`, `build`, `test`, and `style` are excluded from the changelog.
+
 ## Testing
 
 We use `pytest` for testing.
