### 2026-03-12 01d7aad
ci: add PR title check
