### 2026-03-12 a5ccea1
chore: adopt Conventional Commits (#1050)
## Summary

Start using [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/)
