### 2026-03-12 25c95ac
chore: add AI policy (#1051)
## Summary

Add policy for AI-generated contributions.
