+++ CONTRIBUTING.md
@@ -12,6 +12,43 @@ We are using [Devenv](https://devenv.sh/) to manage the development environment,
 
 To reinstall Python dependencies, run `uv sync --all-extras`.
 
+## Using AI-generated code
+
+We recognise that AI coding assistants are now a regular part of many
+developers' workflows and can improve productivity. Thoughtful use of these
+tools can be beneficial, but AI-generated PRs can sometimes lead to
+undesirable additional maintainer burden. PRs that appear to be fully
+generated by AI with little to no engagement from the author may be closed
+without further review.
+
+Human-generated mistakes tend to be easier to spot and reason about, and
+code review is intended to be a collaborative learning experience that
+benefits both submitter and reviewer. When a PR appears to have been
+generated without much engagement from the submitter, reviewers with access
+to AI tools could more efficiently generate the code directly, and since
+the submitter is not likely to learn from the review process, their time is
+more productively spent researching and reporting on the issue.
+
+We are not opposed to the use of AI tools in generating PRs, but recommend
+the following:
+
+- Only submit a PR if you are able to debug and own the changes yourself -
+  review all generated code to understand every detail. [Apache Datafusion has a useful explanation of **why fully AI-generated PRs without understanding are not helpful**](https://datafusion.apache.org/contributor-guide/index.html#why-fully-ai-generated-prs-without-understanding-are-not-helpful).
+- Match the style and conventions used in the rest of the codebase, including
+  PR titles and descriptions
+- Be upfront about AI usage and summarise what was AI-generated
+- If there are parts you don't fully understand, leave comments on your own PR
+  explaining what steps you took to verify correctness
+- Watch for AI's tendency to generate overly verbose comments, unnecessary
+  test cases, and incorrect fixes
+- Break down large PRs into smaller ones to make review easier
+
+PR authors are also responsible for disclosing any copyrighted materials in
+submitted contributions.
+
+> [!NOTE]
+> This policy has been taken from [`delta-io/delta-rs`](https://github.com/delta-io/delta-rs/) project. Thank you!
+
 ## Commit Messages
 
 We use [Conventional Commits](https://www.conventionalcommits.org/). PRs are squash-merged, so the **PR title becomes the commit message** in `main`. This means the PR title must follow conventional commit format — individual commit messages on the branch don't matter. A CI check enforces this.
