### 2026-03-12 ba65905
docs: fix ducklake example (#1042)
