### 2026-03-12 fad2e8a
chore: adopt `git-cliff` (#1049)
## Summary

Adopt conventional commits and start using [`git-cliff`](https://git-cliff.org/) to generate [CHANGELOG.md](http://CHANGELOG.md)
