+++ cliff.toml
@@ -1,0 +1,132 @@
+[git]
+conventional_commits = true
+filter_unconventional = true
+split_commits = false
+commit_preprocessors = [
+  { pattern = '^\((\w+)\)!:', replace = "($1):" },
+]
+commit_parsers = [
+  { message = "^feat", group = "### :sparkles: Added" },
+  { message = "^fix", group = "### :bug: Fixes" },
+  { message = "^docs", group = "### :book: Docs" },
+  { message = "^refactor", group = "### :wrench: Changed" },
+  { message = "^perf", group = "### :zap: Performance" },
+  { message = "^chore", skip = true },
+  { message = "^ci", skip = true },
+  { message = "^build", skip = true },
+  { message = "^test", skip = true },
+  { message = "^style", skip = true },
+]
+tag_pattern = "v[0-9].*"
+sort_commits = "newest"
+
+[changelog]
+header = """
+<!-- --8<-- [start:header] -->
+# Changelog
+
+All notable user-facing Metaxy changes are to be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), except for *experimental* features.
+
+While Metaxy's core functionality and versioning engine are **_stable_** and quite complete, the CLI and some of the more advanced APIs and integrations are considered **_experimental_**.
+
+!!! abstract
+
+    **_stable_** features are guaranteed to follow SemVer and won't receive breaking changes in between minor releases. Such changes will be announced with a deprecation warning.
+    A feature is considered *stable* if it's documented and doesn't have an `"Experimental"` badge.
+
+    **_experimental_** features may be changed or removed at any time without deprecation warnings. They may be documented and in this case must display an `"Experimental"` warning badge:
+
+    !!! warning "Experimental"
+        This functionality is experimental.
+
+<!-- --8<-- [end:header] -->
+<!-- --8<-- [start:releases] -->
+"""
+body = """
+{% for group, commits in commits | group_by(attribute="group") %}\
+{{ group | striptags | trim }}
+{% for commit in commits %}
+{%- if commit.body and commit.body is containing("## Changelog") -%}
+{%- set changelog_section = commit.body | split(pat="## Changelog") | last -%}
+{%- set lines = changelog_section | split(pat="\n") -%}
+{%- for line in lines -%}
+{%- set trimmed = line | trim -%}
+{%- if trimmed is starting_with("- ") or trimmed is starting_with("* ") %}
+{{ trimmed }}
+{%- endif -%}
+{%- endfor -%}
+{%- else %}
+- {{ commit.message | split(pat="\n") | first | trim }}
+{%- endif %}
+{%- endfor %}
+{% endfor %}\
+"""
+footer = """
+## 0.1.4
+
+### :bug: Fixes
+
+- Fixed a bug with `metaxy push` not re-pushing feature definitions on changes reverting them to a previous version.
+
+## 0.1.3
+
+### :sparkles: Added
+
+- `metaxy lock` now avoids loading feature definitions from metaxy.project entrypoints of Python packages that aren't dependencies (direct or transitive) of the current Python package when creating the lock file.
+
+### :wrench: Changed
+
+- Renamed `metaxy graph history` CLI command to `metaxy history`
+
+### :book: Docs
+
+- Added a section on inheritance with Metaxy features
+- Improved Dagster integration docs
+
+## 0.1.2
+
+### :sparkles: Added
+
[…]
