### 2026-03-12 26f70c4
chore: upgrade ty to 0.0.22 (#1053)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog entries (bullet points).
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
