+++ pyproject.toml
@@ -131,7 +131,7 @@ dev = [
   "deltalake>=1.4.2", # 1.4.2 fixed Ibis-generated datafusion dialect
   "mistune>=3.1.4",
   "dagster>=1.12.3",
-  "ty==0.0.21",
+  "ty==0.0.22",
   "datafusion>=51.0.0",
   "pytest-asyncio>=0.26.0",
   "pandas>=2.3.3",
