### 2026-03-13 909611d
chore: improve git-cliff setup (#1052)
