### 2026-03-13 441a69b
feat: explore python 3.14 support again
