### 2026-03-13 e029593
feat: code quality refinement
