### 2026-03-13 0b10580
fix: skip instead of fail docs
