+++ .github/workflows/_docs.yml
@@ -27,6 +27,7 @@ on:
 
 jobs:
   docs:
+    if: ${{ github.ref_type != 'branch' || github.ref_name == 'main' }}
     runs-on: depot-ubuntu-latest
     environment: ${{ inputs.environment }}
 
