### 2026-03-14 600d726
ci: upgrade gha actions (#1057)
## Summary

Various Node 20 out of service warnings are addressed by upgrading our GHA actions, now using Node24

## Changelog

upgraded GHA actions

## Test Plan

executed CI
