+++ .github/actions/publish/action.yml
@@ -10,7 +10,7 @@ runs:
   using: composite
   steps:
     - name: Install uv
-      uses: astral-sh/setup-uv@v5
+      uses: astral-sh/setup-uv@v7.5.0
 
     - name: Build package
       shell: bash
