+++ .github/actions/setup-devenv/action.yml
@@ -11,7 +11,7 @@ runs:
   using: composite
   steps:
     - name: Install Nix
-      uses: cachix/install-nix-action@v31
+      uses: cachix/install-nix-action@v31.10.1
       with:
         nix_path: nixpkgs=channel:nixpkgs-unstable
 
@@ -21,7 +21,7 @@ runs:
         name: devenv
 
     - name: Setup Magic Nix Cache
-      uses: DeterminateSystems/magic-nix-cache-action@main
+      uses: DeterminateSystems/magic-nix-cache-action@v13
 
     - name: Install devenv
       shell: bash
