+++ .github/workflows/slides.yaml
@@ -26,6 +26,9 @@ permissions:
   pages: write
   id-token: write
 
+env:
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: "true"
+
 jobs:
   build:
     name: Build slides and render PDF
@@ -36,9 +39,9 @@ jobs:
         file: ${{ fromJSON(github.event.inputs.slides_json || '["introducing-metaxy.md"]') }}
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
 
-      - uses: oven-sh/setup-bun@v2
+      - uses: oven-sh/setup-bun@v2.1.3
         with:
           bun-version: 1.3.1
 
@@ -58,7 +61,7 @@ jobs:
           echo "safe=$safe" >> "$GITHUB_OUTPUT"
 
       - name: Upload artifact
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7.0.0
         with:
           name: slides-${{ steps.safename.outputs.safe }}
           path: slides/2026-introducing-metaxy/*-export.pdf
@@ -69,7 +72,7 @@ jobs:
     needs: build
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/upload-artifact/merge@v4
+      - uses: actions/upload-artifact/merge@v7.0.0
         with:
           name: slide-pdfs
           pattern: slides-*
