+++ .github/workflows/release.yml
@@ -6,6 +6,9 @@ on:
       - 'v[0-9]+.[0-9]+.[0-9]+*'
   workflow_dispatch:
 
+env:
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: "true"
+
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
   cancel-in-progress: false
@@ -44,7 +47,7 @@ jobs:
       contents: read
       id-token: write
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6.0.2
       - uses: ./.github/actions/publish
         with:
           environment: testpypi
@@ -57,7 +60,7 @@ jobs:
       contents: read
       id-token: write
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6.0.2
       - uses: ./.github/actions/publish
         with:
           environment: pypi
@@ -86,17 +89,17 @@ jobs:
     permissions:
       contents: write
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6.0.2
         with:
           fetch-depth: 0
-      - uses: orhun/git-cliff-action@v4
+      - uses: orhun/git-cliff-action@v4.7.1
         id: cliff
         with:
           config: cliff.toml
           args: --latest --strip header
         env:
           OUTPUT: CHANGES.md
-      - uses: softprops/action-gh-release@v2
+      - uses: softprops/action-gh-release@v2.5.0
         with:
           body_path: CHANGES.md
           tag_name: ${{ github.ref_name }}
