+++ .github/workflows/_qa.yml
@@ -3,6 +3,9 @@ name: _QA (reusable)
 permissions:
   contents: read
 
+env:
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: "true"
+
 on:
   workflow_call:
     inputs:
@@ -41,8 +44,8 @@ jobs:
       claude-plugin: ${{ steps.filter.outputs.claude-plugin }}
     steps:
       - name: Checkout code
-        uses: actions/checkout@v4
-      - uses: dorny/paths-filter@v3
+        uses: actions/checkout@v6.0.2
+      - uses: dorny/paths-filter@v4.0.0
         id: filter
         with:
           filters: |
@@ -69,9 +72,9 @@ jobs:
     runs-on: depot-ubuntu-latest
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
       - name: Install uv
-        uses: astral-sh/setup-uv@v7
+        uses: astral-sh/setup-uv@v7.5.0
         with:
           enable-cache: true
       - name: Run Ruff
@@ -81,9 +84,9 @@ jobs:
     runs-on: depot-ubuntu-latest
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
       - name: Install uv
-        uses: astral-sh/setup-uv@v7
+        uses: astral-sh/setup-uv@v7.5.0
         with:
           enable-cache: true
       - uses: dprint/check@v2.3
@@ -104,7 +107,7 @@ jobs:
         python-version: ${{ fromJSON(inputs.python-versions) }}
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
       - name: Run ty
@@ -129,7 +132,7 @@ jobs:
         python-version: ${{ fromJSON(inputs.python-versions) }}
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
 
@@ -167,7 +170,7 @@ jobs:
 
       - name: Upload Coverage to Codecov
         if: ${{ !cancelled() }}
-        uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v5.5.2
         with:
           slug: anam-org/metaxy
           token: ${{ secrets.CODECOV_TOKEN }}
@@ -183,7 +186,7 @@ jobs:
 
       - name: Upload Test Results to Codecov
         if: ${{ !cancelled() }}
-        uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v5.5.2
         with:
           report_type: test_results
           slug: anam-org/metaxy
@@ -205,9 +208,9 @@ jobs:
       sessions: ${{ steps.sessions.outputs.sessions }}
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
       - name: Install uv
-        uses: astral-sh/setup-uv@v7
+        uses: astral-sh/setup-uv@v7.5.0
         with:
           enable-cache: true
       - name: Generate nox session matrix
@@ -226,7 +229,7 @@ jobs:
         session: ${{ fromJSON(needs.generate-nox-matrix.outputs.sessions) }}
     steps:
       - name: Checkout code
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
       - name: Run ${{ matrix.session }}
@@ -247,15 +250,15 @@ jobs:
[…]
