### 2026-03-15 9439d8b
fix: display help when mx is run without arguments
