### 2026-03-16 dba9a25
test: add regression test for CLI empty arguments
