### 2026-03-16 404bab1
test: simplify CLI regression tests and fix formatting
