### 2026-03-16 a328d4b
fix: display help when mx is run without arguments (#1059)
