### 2026-03-16 254db55
feat: config inheritance (#1060)
## Summary

Resolve #1061

<!-- Brief description of the changes -->

## Changelog

Configuration inheritance is now supported via a new `extends` config field.

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

Added tests.

<!-- How was this tested? -->
