### 2026-03-16 725fd59
Fix AttributeError: annotation on Python 3.10
