+++ pyproject.toml
@@ -33,7 +33,8 @@ dependencies = [
   "polars>=1.35.0",
   "polars-hash>=0.5.1",
   "pydantic>=2.11.9",
-  "pydantic-settings>=2.11.0",
+  "pydantic-settings>=2.11.0; python_version >= '3.11'",
+  "pydantic-settings>=2.11.0,<2.13.0; python_version < '3.11'",
   "python-dotenv>=1.0.0",
   "pyyaml>=6.0.0",
   "pytest-cases>=3.9.1",
