### 2026-03-16 28f7d93
fix(#952): implement SafeEnvSettingsSource for Python 3.10 interop
