+++ pyproject.toml
@@ -33,8 +33,7 @@ dependencies = [
   "polars>=1.35.0",
   "polars-hash>=0.5.1",
   "pydantic>=2.11.9",
-  "pydantic-settings>=2.11.0; python_version >= '3.11'",
-  "pydantic-settings>=2.11.0,<2.13.0; python_version < '3.11'",
+  "pydantic-settings>=2.11.0",
   "python-dotenv>=1.0.0",
   "pyyaml>=6.0.0",
   "pytest-cases>=3.9.1",
@@ -142,6 +141,7 @@ dev = [
   "deptry>=0.24.0",
   "nox>=2025.2.9",
   "nox-uv>=0.7.1",
+  "pydantic-settings>=2.13.0",
 ]
 
 examples = [
