### 2026-03-16 eca26a4
fix(ty): use getattr for safe attribute access in settings_customise_sources
