### 2026-03-16 3616c9e
test: add coverage for SafeEnvSettingsSource and issue #952
