### 2026-03-16 b82ee46
fix: resolve lint, format, and typecheck diagnostics in config and tests
