### 2026-03-16 014a833
feat: support Python 3.14 (#1055)
## Summary

resolves: #108, add support for Python 3.14
