+++ .github/workflows/main.yml
@@ -32,7 +32,7 @@ jobs:
       issues: write
       checks: write
     with:
-      python-versions: ${{ github.event.inputs.full-matrix == 'true' && '["3.10", "3.11", "3.12", "3.13"]' || '["3.10"]' }}
+      python-versions: ${{ github.event.inputs.full-matrix == 'true' && '["3.10", "3.11", "3.12", "3.13", "3.14"]' || '["3.10"]' }}
       run-windows: ${{ github.event.inputs.full-matrix == 'true' }}
       run-macos: ${{ github.event.inputs.full-matrix == 'true' }}
       dependency-matrix: ${{ github.event.inputs.full-matrix == 'true' }}
