### 2026-03-16 d89d1b4
chore: adopt deadcode (#1056)
## Summary

feat: eanble dead code checks in CI and pre-commit via `deadcode`
