### 2026-03-17 be3c1b0
feat: PostgreSQL metadata store (#743)
Add PostgreSQL-backed metadata store. PostgreSQL is used for I/O and when joining upstream features to prepare provenance metadata. The provenance computation itself is then run with Polars. This is because PostgreSQL doesn't support map/struct-like columns, so the versioning engine cannot run there.

Resolves #223.
