+++ tests/metadata_stores/test_postgresql.py
@@ -1,0 +1,1522 @@
+"""PostgreSQL-specific tests that don't apply to other stores.
+
+These tests focus on PostgreSQL-specific functionality like connection handling,
+struct ↔ JSONB serialization, and hash algorithms.
+
+Main provenance and integration tests run automatically via test_provenance_golden_reference.py
+when PostgreSQL is included in the `any_store` fixture.
+
+Requirements:
+    - pytest-postgresql installed (preferred for automatic test database setup)
+    - OR set POSTGRES_TEST_URL environment variable
+    - OR set PG_BIN environment variable to point to PostgreSQL bin directory
+"""
+
+from urllib.parse import parse_qsl, urlparse
+
+import ibis.expr.datatypes as dt
+import ibis.expr.schema as sch
+import narwhals as nw
+import polars as pl
+import polars.testing
+import pytest
+
+from metaxy._utils import collect_to_polars
+from metaxy.metadata_store.exceptions import HashAlgorithmNotSupportedError
+from metaxy.metadata_store.postgresql import PostgreSQLMetadataStore
+from metaxy.models.constants import METAXY_DATA_VERSION_BY_FIELD, METAXY_PROVENANCE_BY_FIELD
+from metaxy.models.feature import current_graph
+from metaxy.models.types import FeatureKey
+from metaxy.versioning.polars import PolarsVersioningEngine
+from metaxy.versioning.types import HashAlgorithm
+from tests.metadata_stores.conftest import _with_search_path
+
+
+@pytest.fixture
+def clean_postgres_db(postgresql_db: str, test_graph):
+    """Provide a clean PostgreSQL database for each test."""
+    with PostgreSQLMetadataStore(postgresql_db) as store:
+        for table_name in store.conn.list_tables():
+            if table_name.startswith("test_stores__"):
+                store.conn.drop_table(table_name, force=True)
+    yield postgresql_db
+
+
+def test_postgresql_native_implementation(postgresql_db: str, test_graph) -> None:
+    """Test that PostgreSQL store always uses Polars engine for versioning."""
+    import narwhals as nw
+
+    with PostgreSQLMetadataStore(postgresql_db) as store:
+        assert store.native_implementation() == nw.Implementation.POLARS
+
+
+def test_postgresql_connection_string_init(postgresql_db: str, test_graph) -> None:
+    """Test initialization with connection string."""
+    store = PostgreSQLMetadataStore(postgresql_db)
+    assert not store._is_open
+
+    with store.open():
+        assert store._is_open
+        assert store._conn is not None
+
+
+def test_postgresql_connection_params_init() -> None:
+    """Test initialization with connection params dict."""
+    params = {
+        "host": "localhost",
+        "port": 5432,
+        "database": "test_db",
+        "user": "test_user",
+        "password": "test_pass",
+    }
+
+    store = PostgreSQLMetadataStore(connection_params=params)
+    assert not store._is_open
+
+
+def test_postgresql_requires_connection_info() -> None:
+    """Test that store requires either connection_string or connection_params."""
+    with pytest.raises(ValueError, match="Must provide either connection_string or connection_params"):
+        PostgreSQLMetadataStore()
+
+
+def test_with_search_path_adds_options_when_absent() -> None:
+    """Search path option should be added while preserving existing query params."""
+    base_url = "postgresql://postgres@localhost:5432/test_db?sslmode=require&connect_timeout=10"
+
+    result = _with_search_path(base_url, "test_schema")
+    parsed = urlparse(result)
+    query_params = dict(parse_qsl(parsed.query, keep_blank_values=True))
+
[…]
