+++ src/metaxy/metadata_store/postgresql.py
@@ -1,0 +1,536 @@
+"""PostgreSQL metadata store with storage/compute separation.
+
+Uses PostgreSQL for storage and Polars for versioning compute.
+Metaxy system struct columns are round-tripped via JSON serialization.
+User struct columns are JSON-encoded on write and decode behavior on read depends on
+the effective SQL column type (JSON/JSONB).
+Always uses PolarsVersioningEngine since PostgreSQL lacks native struct support.
+
+Requirements:
+    - PostgreSQL 12+ (JSONB support)
+    - ibis-framework[postgres] package
+"""
+
+import json
+import warnings
+from collections.abc import Iterator, Sequence
+from contextlib import contextmanager
+from typing import TYPE_CHECKING, Any
+
+import narwhals as nw
+import polars as pl
+from narwhals.typing import Frame
+from pydantic import Field
+
+from metaxy._decorators import experimental, public
+from metaxy._utils import collect_to_polars
+from metaxy.metadata_store.exceptions import TableNotFoundError
+from metaxy.metadata_store.ibis import IbisMetadataStore, IbisMetadataStoreConfig
+from metaxy.models.constants import (
+    METAXY_DATA_VERSION_BY_FIELD,
+    METAXY_PROVENANCE_BY_FIELD,
+)
+from metaxy.models.plan import FeaturePlan
+from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.versioning.polars import PolarsVersioningEngine
+from metaxy.versioning.types import HashAlgorithm
+
+if TYPE_CHECKING:
+    import ibis
+    from ibis.expr.schema import Schema as IbisSchema
+
+    from metaxy.metadata_store.base import MetadataStore
+
+# Metaxy struct columns that need JSON-string ↔ Struct conversion
+_METAXY_STRUCT_COLUMNS = frozenset({METAXY_PROVENANCE_BY_FIELD, METAXY_DATA_VERSION_BY_FIELD})
+
+
+@public
+@experimental
+class PostgreSQLMetadataStoreConfig(IbisMetadataStoreConfig):
+    """Configuration for PostgreSQLMetadataStore.
+
+    Inherits connection_string, connection_params, table_prefix, auto_create_tables from IbisMetadataStoreConfig.
+    """
+
+    auto_cast_struct_for_jsonb: bool = Field(
+        default=True,
+        description="Auto-convert DataFrame Struct columns to JSON strings on write. Metaxy system columns are always converted.",
+    )
+
+
+@public
+@experimental
+class PostgreSQLMetadataStore(IbisMetadataStore):
+    """PostgreSQL metadata store with storage/compute separation.
+
+    Uses PostgreSQL for storage and Polars for versioning.
+    Filters push down to SQL WHERE clauses, then data materializes to Polars.
+
+    Example:
+        <!-- skip next -->
+        ```python
+        from metaxy.metadata_store.postgresql import PostgreSQLMetadataStore
+
+        store = PostgreSQLMetadataStore(connection_string="postgresql://user:pass@localhost:5432/metaxy")
+
+        with store:
+            increment = store.resolve_update(MyFeature)
+            store.write(MyFeature, increment.added)
+        ```
+    """
+
+    # Override to use Polars versioning engine (PostgreSQL has no native struct support)
+    versioning_engine_cls = PolarsVersioningEngine
+
+    def __init__(
+        self,
+        connection_string: str | None = None,
+        *,
+        connection_params: dict[str, Any] | None = None,
+        fallback_stores: list["MetadataStore"] | None = None,
+        auto_cast_struct_for_jsonb: bool = True,
+        **kwargs: Any,
+    ):
+        """Initialize PostgreSQL metadata store.
+
+        Args:
[…]
