+++ pyproject.toml
@@ -86,6 +86,10 @@ lancedb = [
   "lancedb>=0.15.0",
   "polars>=1.36.1",
 ]
+postgres = [
+  "ibis-framework[postgres]>=11.0.0",
+  "psycopg>=3.2.0",
+]
 mcp = [
   "fastmcp>=2.0.0",
 ]
@@ -119,9 +123,11 @@ dev = [
   "hypothesis>=6.142.4",
   "ruff>=0.14.4",
   "syrupy>=5.0.0",
-  "ibis-framework[duckdb,clickhouse]>=11.0.0",
+  "ibis-framework[duckdb,clickhouse,postgres]>=11.0.0",
   "lancedb>=0.15.0",
   "deadcode>=2.4.1",
+  "pytest-postgresql>=6.1.0",
+  "psycopg[binary]>=3.2.0", # Bundle libpq to avoid LD_LIBRARY_PATH issues
   "prek>=0.2.11",
   "pytest-env>=1.2.0",
   "pytest-cov>=7.0.0",
@@ -298,11 +304,14 @@ markers = [
   "duckdb: DuckDB store tests",
   "ducklake: DuckLake integration tests",
   "clickhouse: ClickHouse store tests",
-  "ibis: Ibis-based store tests (DuckDB, ClickHouse)",
+  "ibis: Ibis-based store tests (DuckDB, ClickHouse, PostgreSQL)",
   "native: Stores with native compute capabilities",
   "polars: Stores using Polars for compute",
   "delta: Delta Lake store tests",
   "motherduck: MotherDuck live integration tests (requires MOTHERDUCK_TOKEN)",
+  "lancedb: LanceDB store tests",
+  "postgresql: PostgreSQL store tests",
+  "integration: Integration tests (require external services)",
 ]
 
 [tool.deptry.package_module_name_map]
