+++ tests/test_config.py
@@ -1454,7 +1454,7 @@ class TestConfigInheritance:
         assert config.stores["shared"].config["root_path"] == "/child"
         assert config.stores["parent_only"].config["root_path"] == "/parent"
 
-    def test_entrypoints_replaced_when_set(self, tmp_path: Path) -> None:
+    def test_entrypoints_appended_when_set(self, tmp_path: Path) -> None:
         parent = tmp_path / "parent.toml"
         parent.write_text('entrypoints = ["parent.module"]\n')
 
@@ -1463,7 +1463,7 @@ class TestConfigInheritance:
 
         config = MetaxyConfig.load(child)
 
-        assert config.entrypoints == ["child.module"]
+        assert config.entrypoints == ["parent.module", "child.module"]
 
     def test_entrypoints_inherited_when_unset(self, tmp_path: Path) -> None:
         parent = tmp_path / "parent.toml"
@@ -1538,3 +1538,48 @@ class TestConfigInheritance:
         assert config.project == "inherited"
         assert config.store == "prod"
         assert config.entrypoints == ["mod.a"]
+
+    def test_ext_plugin_config_inherited(self, tmp_path: Path) -> None:
+        """Parent's ext plugin config (including extra fields) is inherited."""
+        parent = tmp_path / "parent.toml"
+        parent.write_text("[ext.sqlmodel]\nenable = true\ninject_primary_key = true\ninject_index = true\n")
+
+        child = tmp_path / "child.toml"
+        child.write_text('extends = "parent.toml"\nproject = "child"\n')
+
+        config = MetaxyConfig.load(child)
+
+        assert "sqlmodel" in config.ext
+        assert config.ext["sqlmodel"].enable is True
+        assert config.ext["sqlmodel"].model_extra
+        assert config.ext["sqlmodel"].model_extra["inject_primary_key"] is True
+        assert config.ext["sqlmodel"].model_extra["inject_index"] is True
+
+    def test_ext_plugin_config_shallow_merged(self, tmp_path: Path) -> None:
+        """Child ext entry replaces parent's entry for the same key (shallow merge)."""
+        parent = tmp_path / "parent.toml"
+        parent.write_text("[ext.sqlmodel]\nenable = true\ninject_primary_key = true\ninject_index = true\n")
+
+        child = tmp_path / "child.toml"
+        child.write_text('extends = "parent.toml"\n[ext.sqlmodel]\ninject_primary_key = false\n')
+
+        config = MetaxyConfig.load(child)
+
+        # Child's sqlmodel replaces parent's entirely — only child's fields are present
+        assert config.ext["sqlmodel"].enable is False  # default, not parent's True
+        assert config.ext["sqlmodel"].model_extra == {"inject_primary_key": False}
+
+    def test_ext_child_adds_new_plugin(self, tmp_path: Path) -> None:
+        """Child can add new plugins while preserving parent's plugins."""
+        parent = tmp_path / "parent.toml"
+        parent.write_text("[ext.myplugin]\nenable = false\ncustom_setting = true\n")
+
+        child = tmp_path / "child.toml"
+        child.write_text('extends = "parent.toml"\n[ext.otherplugin]\nenable = false\n')
+
+        config = MetaxyConfig.load(child)
+
+        assert "myplugin" in config.ext
+        assert config.ext["myplugin"].model_extra
+        assert config.ext["myplugin"].model_extra["custom_setting"] is True
+        assert "otherplugin" in config.ext
