+++ docs/reference/configuration.md
@@ -47,14 +47,14 @@ Metaxy supports templating environment variables in configuration files using th
 
 ### Configuration Inheritance
 
-A configuration file can inherit settings from another file using the [`extends`](#metaxy.config.MetaxyConfig.extends) field. This is especially useful in monorepos: a shared `metaxy.toml` can define stores, plugins, and other common settings.
+A configuration file can inherit settings from another file using the [`extend`](#metaxy.config.MetaxyConfig.extend) field. This is especially useful in monorepos: a shared `metaxy.toml` can define stores, plugins, and other common settings.
 
 ```toml title="metaxy.toml"
 project = "my_project"
-extends = "../base.toml"
+extend = "../base.toml"
 ```
 
-The `extends` path can be either relative to the file that contains it or an absolute path.
+The `extend` path can be either relative to the file that contains it or an absolute path.
 
 When a child config extends a parent, scalar fields are replaced, dictionary fields are shallow-merged, and array fields are extended.
 
