+++ docs/integrations/metadata-stores/databases/postgresql.md
@@ -5,70 +5,26 @@ description: "PostgreSQL as a metadata store backend."
 
 # Metaxy + PostgreSQL
 
-Metaxy implements [`PostgreSQLMetadataStore`][metaxy.metadata_store.postgresql.PostgreSQLMetadataStore]. It uses [PostgreSQL](https://www.postgresql.org/) for metadata storage with [Polars](https://pola.rs/) for versioning compute.
+!!! warning "Experimental"
+    This functionality is experimental.
 
-**Hashing**: All provenance hashing is computed in Polars using xxHash32 (default) or other algorithms from the [polars-hash](https://github.com/ion-elgreco/polars-hash) plugin. Only the pre-computed hash results are stored in PostgreSQL. No PostgreSQL extensions are required.
+Metadata managed by Metaxy can be stored in [`PostgreSQLMetadataStore`][metaxy.metadata_store.postgresql.PostgreSQLMetadataStore].
+It uses [PostgreSQL](https://www.postgresql.org/).
+This metadata store backend is limited in comparison to others, because PostgreSQL doesn't support map-like data types, and Metaxy's versioning engine can't run in the database.
+The local Polars versioning engine is used instead.
+This results in the following limitations for [`MetadataStore.resolve_update`][metaxy.MetadataStore.resolve_update]:
 
-!!! tip "Alembic Integration"
-
-    See [Alembic setup guide](../../plugins/sqlalchemy.md#alembic-integration) for instructions on using Alembic with Metaxy.
-
-## Installation
-
-```shell
-pip install 'metaxy[postgres]'
-```
+- **Increased I/O**: entire upstream metadata has to be fetched to memory
+- **Increased Memory footprint**: expect high memory usage, especially when having many upstream features
 
 ## Metaxy's Versioning Struct Columns
 
-Metaxy uses struct columns (`metaxy_provenance_by_field`, `metaxy_data_version_by_field`) to track field-level versioning. In Python world this corresponds to `dict[str, str]`. PostgreSQL doesn't have native struct types, so these columns should be stored as `JSONB` (recommended), though actual storage type depends on your table schema.
+PostgreSQL doesn't have native map-like or struct types, so it's recommended to store Metaxy's versioning columns as `JSONB`.
+As a convenience feature, `PostgreSQLMetadataStore` will automatically json-encodes `pl.Struct` columns when writing metadata and parse them to `pl.Struct` when reading.
+This behavior can be disabled with `auto_cast_struct_for_jsonb` configuration parameter. This setting only affects user-defined columns, while Metaxy's versioning columns are always encoded/parsed.
 
-### How PostgreSQL Handles Structs
+## API Reference
 
-PostgreSQL offers multiple approaches to represent structured data:
-
-| Type                                                                     | Description               | Use Case                                           |
-| ------------------------------------------------------------------------ | ------------------------- | -------------------------------------------------- |
-| [`JSONB`](https://www.postgresql.org/docs/current/datatype-json.html)    | Binary JSON with indexing | **Recommended for Metaxy** because of dynamic keys |
-| [`JSON`](https://www.postgresql.org/docs/current/datatype-json.html)     | Text-based JSON           | Less efficient than JSONB                          |
-| [Composite Types](https://www.postgresql.org/docs/current/rowtypes.html) | User-defined row types    | More performant than JSONB but keys are static     |
-
-!!! success "Recommended: `JSONB`"
-
[…]
