+++ CHANGELOG.md
@@ -21,15 +21,37 @@ While Metaxy's core functionality and versioning engine are **_stable_** and qui
 <!-- --8<-- [end:header] -->
 <!-- --8<-- [start:releases] -->
 
-## 1.4.0
+## v0.1.5 (18-03-2026)
 
-### :bug: Fixes
+### :sparkles: Features
+
+- PostgreSQL metadata store ([#743](https://github.com/anam-org/metaxy/pull/743) by [@geoHeil](https://github.com/geoHeil))
+- support Python 3.14 ([#1055](https://github.com/anam-org/metaxy/pull/1055) by [@geoHeil](https://github.com/geoHeil))
+- Configuration inheritance is now supported via a new `extend` config field. ([#1060](https://github.com/anam-org/metaxy/pull/1060) by [@danielgafni](https://github.com/danielgafni))
+
+### :bug: Bug Fixes
+
+- display help when mx is run without arguments ([#1059](https://github.com/anam-org/metaxy/pull/1059) by [@AnkitSharma-29](https://github.com/AnkitSharma-29))
+
+### :book: Docs
+
+- refine PostgreSQL docs ([#1066](https://github.com/anam-org/metaxy/pull/1066) by [@danielgafni](https://github.com/danielgafni))
+- fix ducklake example ([#1042](https://github.com/anam-org/metaxy/pull/1042) by [@geoHeil](https://github.com/geoHeil))
+
+### :heart: New Contributors
+
+- [@AnkitSharma-29](https://github.com/AnkitSharma-29) made their first contribution in [#1059](https://github.com/anam-org/metaxy/pull/1059)
+
+
+## 0.1.4 (12-03-2026)
+
+### :bug: Bug Fixes
 
 - Fixed a bug with `metaxy push` not re-pushing feature definitions on changes reverting them to a previous version.
 
-## 1.3.0
+## 0.1.3 (11-03-2026)
 
-### :sparkles: Added
+### :sparkles: Features
 
 - `metaxy lock` now avoids loading feature definitions from metaxy.project entrypoints of Python packages that aren't dependencies (direct or transitive) of the current Python package when creating the lock file.
 
@@ -42,35 +64,35 @@ While Metaxy's core functionality and versioning engine are **_stable_** and qui
 - Added a section on inheritance with Metaxy features
 - Improved Dagster integration docs
 
-## 1.2.0
+## 0.1.2 (06-03-2026)
 
-### :sparkles: Added
+### :sparkles: Features
 
 - A new `staleness_predicates` parameter has been added to [`resolve_update`][metaxy.MetadataStore.resolve_update]. It can be used to mark records as stale based on arbitrary Narwhals expressions, regardless of their Metaxy versions.
 - `MetadataStore.write` now fills `metaxy_data_version` and `metaxy_data_version_by_field` values in place of `Null`s on per-row basis instead of checking whether the whole column presence.
 - Added [Rebases](/guide/concepts/metadata-stores.md#rebasing-metadata-versions) functionality which allows backfilling metadata from historical feature versions. It's available via [CLI](/reference/cli.md#metaxy-metadata-rebase) or as [`MetadataStore.rebase`][metaxy.MetadataStore.rebase].
 
-## 1.1.0
+## 0.1.1 (27-02-2026)
 
-### :sparkles: Added
+### :sparkles: Features
 
 - A new top-level [`extra_features`](/reference/configuration.md#metaxy.config.MetaxyConfig.extra_features) configuration option has been added. It can be used to register additional external feature definitions from the metadata store on the feature graph. Learn more [here](/guide/concepts/definitions/external-features.md/#loading-extra-features).
[…]
