+++ .github/workflows/release.yml
@@ -66,7 +66,7 @@ jobs:
           environment: pypi
 
   check-stable-release:
-    needs: pypi
+    needs: [pypi, github-release]
     runs-on: depot-ubuntu-latest
     outputs:
       is-stable: ${{ steps.check.outputs.stable }}
@@ -84,26 +84,6 @@ jobs:
             echo "stable=false" >> "$GITHUB_OUTPUT"
           fi
 
-  github-release:
-    runs-on: depot-ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - uses: actions/checkout@v6.0.2
-        with:
-          fetch-depth: 0
-      - uses: orhun/git-cliff-action@v4.7.1
-        id: cliff
-        with:
-          config: cliff.toml
-          args: --latest --strip header
-        env:
-          OUTPUT: CHANGES.md
-      - uses: softprops/action-gh-release@v2.5.0
-        with:
-          body_path: CHANGES.md
-          tag_name: ${{ github.ref_name }}
-
   docs:
     needs: check-stable-release
     if: needs.check-stable-release.outputs.is-stable == 'true'
@@ -116,3 +96,17 @@ jobs:
       environment: github-pages
       docs-version: ${{ needs.check-stable-release.outputs.version }}
       docs-alias: 'stable'
+
+  github-release:
+    needs: pypi
+    runs-on: depot-ubuntu-latest
+    permissions:
+      contents: write
+    steps:
+      - uses: actions/checkout@v6.0.2
+      - name: Extract release notes from CHANGELOG.md
+        run: awk '/^## /{if(found) exit; found=1; next} found' CHANGELOG.md > CHANGES.md
+      - uses: softprops/action-gh-release@v2.5.0
+        with:
+          body_path: CHANGES.md
+          tag_name: ${{ github.ref_name }}
