### 2026-03-18 8583d71
Deployed 99ada7ed to 0.1.5 with MkDocs 1.6.1 and mike 2.1.3
