### 2026-03-18 6e696c3
refactor: move PostgreSQL metadata store to metaxy.ext (#1069)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
