+++ docs/integrations/metadata-stores/databases/postgresql.md
@@ -21,7 +21,7 @@ This results in the following limitations for [`MetadataStore.resolve_update`][m
 
 PostgreSQL doesn't have native map-like or struct types, so it's recommended to store Metaxy's versioning columns as `JSONB`.
 As a convenience feature, `PostgreSQLMetadataStore` will automatically json-encodes `pl.Struct` columns when writing metadata and parse them to `pl.Struct` when reading.
-This behavior can be disabled with `auto_cast_struct_for_jsonb` configuration parameter. This setting only affects user-defined columns, while Metaxy's versioning columns are always encoded/parsed.
+This behavior can be disabled with [`auto_cast_struct_for_jsonb`](#metaxy.ext.metadata_stores.postgresql.PostgreSQLMetadataStoreConfig.auto_cast_struct_for_jsonb) configuration parameter. This setting only affects user-defined columns, while Metaxy's versioning columns are always encoded/parsed.
 
 ## API Reference
 
