### 2026-03-18 c65bb76
feat: add example BigQuery E2E
Production-grade BigQuery metadata store example with SQLModel
features and dual-track Alembic migrations (system + feature tables).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
