+++ tests/ext/bigquery/test_native.py
@@ -1,4 +1,4 @@
-"""BigQuery-specific tests that don't apply to other stores."""
+"""BigQuery metadata store tests."""
 
 from unittest.mock import MagicMock, Mock, patch
 
@@ -7,8 +7,59 @@ import pytest
 from metaxy_testing.models import SampleFeature
 
 from metaxy.ext.bigquery import BigQueryMetadataStore
+from metaxy.metadata_store.base import MetadataStore
 from metaxy.models.feature import FeatureGraph
 from metaxy.versioning.types import HashAlgorithm
+from tests.metadata_stores.shared import (
+    CRUDTests,
+    DeletionTests,
+    DisplayTests,
+    FilterTests,
+    IbisMapTests,
+    MapDtypeTests,
+    ResolveUpdateTests,
+    VersioningTests,
+    WriteTests,
+)
+
+
+@pytest.mark.ibis
+@pytest.mark.native
+@pytest.mark.bigquery
+class TestBigQuery(
+    CRUDTests,
+    DeletionTests,
+    DisplayTests,
+    FilterTests,
+    IbisMapTests,
+    MapDtypeTests,
+    ResolveUpdateTests,
+    VersioningTests,
+    WriteTests,
+):
+    @pytest.fixture
+    def store(self, request: pytest.FixtureRequest) -> MetadataStore:
+        return BigQueryMetadataStore(
+            project_id=request.getfixturevalue("bigquery_project_id"),
+            dataset_id=request.getfixturevalue("bigquery_dataset"),
+            hash_algorithm=HashAlgorithm.MD5,
+            auto_create_tables=True,
+        )
+
+    @pytest.fixture
+    def named_store(self, request: pytest.FixtureRequest) -> MetadataStore:
+        return BigQueryMetadataStore(
+            project_id=request.getfixturevalue("bigquery_project_id"),
+            dataset_id=request.getfixturevalue("bigquery_dataset"),
+            hash_algorithm=HashAlgorithm.MD5,
+            auto_create_tables=True,
+            name="bigquery-test",
+        )
+
+
+# ---------------------------------------------------------------------------
+# BigQuery-specific unit tests (no live connection required)
+# ---------------------------------------------------------------------------
 
 
 @pytest.fixture
@@ -23,7 +74,6 @@ def mock_bigquery_connection():
 
 
 def test_bigquery_initialization_with_project_dataset():
-    """Test BigQuery store initialization with project and dataset."""
     store = BigQueryMetadataStore(
         project_id="test-project",
         dataset_id="test_dataset",
@@ -34,18 +84,15 @@ def test_bigquery_initialization_with_project_dataset():
 
 
 def test_bigquery_initialization_with_credentials_path():
-    """Test BigQuery store initialization with credentials path."""
     with patch("google.oauth2.service_account.Credentials.from_service_account_file") as mock_creds:
-        mock_creds_instance = Mock()
-        mock_creds.return_value = mock_creds_instance
+        mock_creds.return_value = Mock()
 
-        _ = BigQueryMetadataStore(
+        BigQueryMetadataStore(
             project_id="test-project",
             dataset_id="test_dataset",
             credentials_path="/path/to/creds.json",
         )
 
-        # Verify credentials were loaded using the recommended method
         mock_creds.assert_called_once_with(
             "/path/to/creds.json",
             scopes=["https://www.googleapis.com/auth/bigquery"],
@@ -53,9 +100,7 @@ def test_bigquery_initialization_with_credentials_path():
 
 
 def test_bigquery_initialization_with_invalid_credentials_path():
-    """Test BigQuery store initialization with invalid credentials path."""
[…]
