+++ examples/example-intro/src/example_intro/__entrypoint__.py
@@ -1,0 +1,6 @@
+"""Entrypoint for intro example.
+
+This module imports and registers the feature definitions.
+"""
+
+from example_intro import features  # noqa: F401
