+++ docs/examples/intro.md
@@ -1,0 +1,182 @@
+---
+title: "Tutorial"
+description: "Step-by-step tutorial building a video ML pipeline with incremental recomputation."
+---
+
+# Tutorial
+
+## Overview
+
+::: metaxy-example source-link
+    example: intro
+
+This tutorial walks through building a multimodal ML pipeline that processes video files. By the end you will understand Metaxy's core workflow: declaring features with dependencies, resolving incremental updates, and detecting when code changes require selective recomputation.
+
+The pipeline extracts faces and transcriptions from videos, then summarizes the transcriptions:
+
+::: metaxy-example graph
+    example: intro
+    scenario: "Initial pipeline run"
+    direction: LR
+
+## Project Setup
+
+Install the example's dependencies:
+
+```shell
+uv sync
+```
+
+A Metaxy project needs a `metaxy.toml` that tells Metaxy where to find feature definitions and which metadata store to use:
+
+<!-- dprint-ignore-start -->
+```toml title="metaxy.toml"
+--8<-- "example-intro/metaxy.toml"
+```
+<!-- dprint-ignore-end -->
+
+The `entrypoints` field points to a module that imports the feature definitions so Metaxy can discover them. The pipeline initializes Metaxy and obtains a store handle:
+
+<!-- dprint-ignore-start -->
+```python title="src/example_intro/pipeline.py"
+--8<-- "example-intro/src/example_intro/pipeline.py:init"
+```
+<!-- dprint-ignore-end -->
+
+## Defining Features
+
+### `"video"` -- root feature
+
+The root feature represents raw video files. It has no upstream dependencies — its provenance is provided externally via `metaxy_provenance_by_field` rather than derived from other features.
+
+<!-- dprint-ignore-start -->
+```python title="src/example_intro/features.py"
+--8<-- "example-intro/src/example_intro/features.py:video_feature"
+```
+<!-- dprint-ignore-end -->
+
+### `"video/faces"` -- face detection
+
+Depends on `"video"`. The `faces` field has a `code_version` that tracks the algorithm version.
+
+<!-- dprint-ignore-start -->
+```python title="src/example_intro/features.py" hl_lines="4"
+--8<-- "example-intro/src/example_intro/features.py:face_detection_feature"
+```
+<!-- dprint-ignore-end -->
+
+The `deps=[Video]` declaration tells Metaxy that when the parent's provenance changes, face detection must be recomputed. Changes to sibling features like `"video/transcription"` do not affect this feature.
+
+### `"video/transcription"` -- audio transcription
+
+Also depends on `"video"`, but is independent of `"video/faces"`.
+
+<!-- dprint-ignore-start -->
+```python title="src/example_intro/features.py" hl_lines="4"
+--8<-- "example-intro/src/example_intro/features.py:audio_transcription_feature"
+```
+<!-- dprint-ignore-end -->
+
+### `"video/summary"` -- summarization
+
+Depends on `"video/transcription"`. When the transcription algorithm changes, summaries are automatically recomputed. Changes to `"video/faces"` do not trigger recomputation here because there is no dependency path between them.
+
+<!-- dprint-ignore-start -->
+```python title="src/example_intro/features.py" hl_lines="4"
+--8<-- "example-intro/src/example_intro/features.py:summary_feature"
+```
+<!-- dprint-ignore-end -->
+
+## Running the Pipeline
+
+### Ingesting root samples
+
+Root features need explicit input samples. Each sample carries a `metaxy_provenance_by_field` column — a dictionary mapping field keys to their current data version. When a value changes, Metaxy knows the underlying data has been updated and downstream features need recomputation. The `"default"` key applies to all fields that do not have an explicit entry.
+
+<!-- dprint-ignore-start -->
[…]
