+++ examples/example-intro/README.md
@@ -1,0 +1,3 @@
+# Introduction
+
+View this example in [Metaxy docs](https://docs.metaxy.io/latest/examples/intro/).
