### 2026-03-18 e0dae1b
feat: add example Tutorial 101
Beginner-friendly video/audio ML pipeline tutorial with DuckDB.
Demonstrates feature dependencies, idempotent reruns, and selective
recomputation when code versions change.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
