+++ examples/example-intro/src/example_intro/pipeline.py
@@ -1,0 +1,143 @@
+"""Introductory pipeline demonstrating Metaxy's core workflow.
+
+This pipeline processes video files through a multi-step ML pipeline:
+  Video -> FaceDetection (mocked)
+  Video -> AudioTranscription (mocked) -> Summary (mocked)
+
+Each step uses resolve_update to determine which samples need processing,
+then writes the results back to the metadata store.
+"""
+
+import metaxy as mx
+import polars as pl
+
+from example_intro.features import AudioTranscription, FaceDetection, Summary, Video
+
+# --8<-- [start:init]
+# Initialize Metaxy: loads metaxy.toml and discovers features via entrypoints
+config = mx.init()
+store = config.get_store()
+# --8<-- [end:init]
+
+print("Intro Pipeline")
+print("=" * 60)
+
+# --8<-- [start:video_samples]
+# Step 1: Register video samples with provenance tracking.
+# metaxy_provenance_by_field tracks the data version of each field.
+# When provenance changes, downstream features are recomputed.
+video_samples = pl.DataFrame(
+    {
+        "video_id": ["v001", "v002", "v003"],
+        "path": [
+            "/data/videos/v001.mp4",
+            "/data/videos/v002.mp4",
+            "/data/videos/v003.mp4",
+        ],
+        "metaxy_provenance_by_field": [
+            {"default": "v001_hash_abc"},
+            {"default": "v002_hash_def"},
+            {"default": "v003_hash_ghi"},
+        ],
+    }
+)
+
+with store.open("w"):
+    video_increment = store.resolve_update(Video, samples=video_samples)
+    n_new_videos = len(video_increment.new)
+    if n_new_videos > 0:
+        store.write(Video, video_increment.new)
+        print(f"\n[Video] Wrote {n_new_videos} new samples")
+    else:
+        print("\n[Video] No new or changed samples")
+# --8<-- [end:video_samples]
+
+# --8<-- [start:face_detection]
+# Step 2: Run face detection on videos.
+# resolve_update without samples auto-resolves from upstream (Video).
+# It compares upstream provenance to find new/stale samples.
+with store.open("w"):
+    face_increment = store.resolve_update(FaceDetection)
+    face_to_process = pl.concat(
+        [face_increment.new.to_polars(), face_increment.stale.to_polars()]
+    )
+    if len(face_to_process) > 0:
+        # Mock face detection: add computed columns while preserving system columns
+        face_results = []
+        for row in face_to_process.iter_rows(named=True):
+            face_results.append(
+                {
+                    **row,
+                    "num_faces": 2,
+                    "face_embeddings_path": f"/data/faces/{row['video_id']}/embeddings.npy",
+                }
+            )
+        store.write(FaceDetection, pl.DataFrame(face_results))
+        print(f"[FaceDetection] Materialized {len(face_results)} samples")
+    else:
+        print("[FaceDetection] No new or changed samples")
+# --8<-- [end:face_detection]
+
+# --8<-- [start:audio_transcription]
+# Step 3: Transcribe audio from videos.
+# Same pattern: resolve_update finds what needs processing.
+with store.open("w"):
+    transcript_increment = store.resolve_update(AudioTranscription)
+    transcript_to_process = pl.concat(
+        [
+            transcript_increment.new.to_polars(),
+            transcript_increment.stale.to_polars(),
+        ]
+    )
+    if len(transcript_to_process) > 0:
+        # Mock transcription: add computed columns while preserving system columns
+        transcript_results = []
+        for row in transcript_to_process.iter_rows(named=True):
+            transcript_results.append(
+                {
+                    **row,
+                    "transcript": f"Transcript for {row['video_id']}.",
+                    "language": "en",
[…]
