### 2026-03-18 2664971
fix(sqlmodel): remove integration config (#1074)
## Summary

This functionality was always problematic, as it required the toml file to be evaluated and applied at import time, including usage outside of Metaxy.

This integration is now only available through the code interface.

## Changelog

the `SQLModel` integration configuration (`[ext.sqlmodel]` in `metaxy.toml`) has been removed. Please use the code interface instead.

## Test Plan

Updating tests.
