+++ src/metaxy/ext/sqlmodel/__init__.py
@@ -1,13 +1,13 @@
-from metaxy.ext.sqlmodel.config import SQLModelPluginConfig
 from metaxy.ext.sqlmodel.plugin import (
     BaseSQLModelFeature,
+    SQLModelFeatureConfig,
     SQLModelFeatureMeta,
     filter_feature_sqlmodel_metadata,
 )
 
 __all__ = [
+    "SQLModelFeatureConfig",
     "SQLModelFeatureMeta",
     "BaseSQLModelFeature",
-    "SQLModelPluginConfig",
     "filter_feature_sqlmodel_metadata",
 ]
