+++ src/metaxy/ext/sqlmodel/plugin.py
@@ -4,7 +4,7 @@ This module provides a combined metaclass that allows Metaxy Feature classes
 to also be SQLModel table classes, enabling seamless integration with SQLAlchemy/SQLModel ORMs.
 """
 
-from typing import TYPE_CHECKING, Any, ClassVar
+from typing import TYPE_CHECKING, Any, ClassVar, TypedDict
 
 from pydantic import AwareDatetime, BaseModel
 from sqlalchemy.types import JSON, DateTime
@@ -14,7 +14,6 @@ from sqlmodel.main import SQLModelMetaclass
 from metaxy import FeatureSpec
 from metaxy._decorators import public
 from metaxy.config import MetaxyConfig
-from metaxy.ext.sqlmodel.config import SQLModelPluginConfig
 from metaxy.models.constants import (
     ALL_SYSTEM_COLUMNS,
     METAXY_CREATED_AT,
@@ -48,6 +47,18 @@ class MetaxyTableInfo(BaseModel):
     feature_key: ValidatedFeatureKey
 
 
+@public
+class SQLModelFeatureConfig(TypedDict, total=False):
+    """SQLModel-specific feature configuration.
+
+    Pass via ``feature_config = FeatureConfig(ext={"sqlmodel": SQLModelFeatureConfig(...)})``
+    on your feature class or a shared base class.
+    """
+
+    inject_primary_key: bool
+    inject_index: bool
+
+
 class SQLModelFeatureMeta(MetaxyMeta, SQLModelMetaclass):
     def __new__(
         cls,
@@ -56,8 +67,6 @@ class SQLModelFeatureMeta(MetaxyMeta, SQLModelMetaclass):
         namespace: dict[str, Any],
         *,
         spec: FeatureSpecWithIDColumns | None = None,
-        inject_primary_key: bool | None = None,
-        inject_index: bool | None = None,
         **kwargs: Any,
     ) -> type[Any]:
         """Create a new SQLModel + Metaxy Feature class.
@@ -67,10 +76,6 @@ class SQLModelFeatureMeta(MetaxyMeta, SQLModelMetaclass):
             bases: Base classes
             namespace: Class namespace (attributes and methods)
             spec: Metaxy FeatureSpec (required for concrete features)
-            inject_primary_key: If True, automatically create composite primary key
-                including (metaxy_feature_version, *id_columns, metaxy_updated_at).
-            inject_index: If True, automatically create composite index
-                including (metaxy_feature_version, *id_columns, metaxy_updated_at).
             **kwargs: Additional keyword arguments (e.g., table=True for SQLModel)
 
         Returns:
@@ -82,12 +87,10 @@ class SQLModelFeatureMeta(MetaxyMeta, SQLModelMetaclass):
 
             namespace["model_config"] = ConfigDict(frozen=False)
 
-        # Check plugin config for defaults
-        sqlmodel_config = MetaxyConfig.get_plugin("sqlmodel", SQLModelPluginConfig)
-        if inject_primary_key is None:
-            inject_primary_key = sqlmodel_config.inject_primary_key
-        if inject_index is None:
-            inject_index = sqlmodel_config.inject_index
+        # Resolve SQLModel feature config from feature_config class attribute (inherited via MRO)
+        sqlmodel_config = cls._resolve_sqlmodel_config(namespace, bases)
+        inject_primary_key = sqlmodel_config.get("inject_primary_key", False)
+        inject_index = sqlmodel_config.get("inject_index", False)
 
         # If this is a concrete table (table=True) with a spec
         if kwargs.get("table") and spec is not None:
@@ -130,6 +133,20 @@ class SQLModelFeatureMeta(MetaxyMeta, SQLModelMetaclass):
[…]
