### 2026-03-18 1338a58
fix(sqlalchemy): remove integration config (#1075)
## Summary

This functionality was always problematic, as it required the toml file to be evaluated and applied at import time, including usage outside of Metaxy.

This integration is now only available through the code interface.

## Changelog

the `SQLAlchemy` integration configuration (`[ext.sqlalchemy]` in `metaxy.toml`) has been removed. Please use the code interface instead.

## Test Plan

Updating tests.
