+++ src/metaxy/ext/sqlalchemy/__init__.py
@@ -10,14 +10,12 @@ integration with migration tools:
 - `filter_feature_sqla_metadata`: Get URL and feature table metadata for a store
 """
 
-from metaxy.ext.sqlalchemy.config import SQLAlchemyConfig
 from metaxy.ext.sqlalchemy.plugin import (
     filter_feature_sqla_metadata,
     get_system_slqa_metadata,
 )
 
 __all__ = [
-    "SQLAlchemyConfig",
     "get_system_slqa_metadata",
     "filter_feature_sqla_metadata",
 ]
