+++ 0.1.6/slides/2026-introducing-metaxy/dist/assets/md-CwZeasl4.js
@@ -1,0 +1,1 @@
+import{_ as u}from"./slidev/VClicks-kWLYTJrV.js";import{b as i,o as c,w as n,g as t,e as g,D as l,v as p,x as d,z as r}from"./modules/vue-D2_2azlL.js";import{I as m}from"./slidev/default-DMaQI0-d.js";import{u as f,f as _}from"./slidev/context-CXJvKEzn.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";const b={__name:"introducing-metaxy.md__slidev_9",setup(k){const{$clicksContext:e,$frontmatter:s}=f();return e.setup(),(x,o)=>{const a=u;return c(),i(m,p(d(r(_)(r(s),8))),{default:n(()=>[o[1]||(o[1]=t("h1",null,"Works With Your Stack",-1)),g(a,null,{default:n(()=>[...o[0]||(o[0]=[t("p",null,[t("strong",null,"Storage: start small, scale up")],-1),t("ul",null,[t("li",null,[l("Prototype with "),t("strong",null,"DuckDB"),l(" on your laptop")]),t("li",null,[l("Deploy to "),t("strong",null,"ClickHouse"),l(", "),t("strong",null,"BigQuery"),l(", "),t("strong",null,"Delta Lake"),l(", "),t("strong",null,"LanceDB"),l(", "),t("strong",null,"DuckLake"),l(", or "),t("strong",null,"PostgreSQL")]),t("li",null,[l("Built on "),t("strong",null,"Ibis"),l(" + "),t("strong",null,"Narwhals"),l(" — adding new backends is straightforward")])],-1),t("p",null,[t("strong",null,"Orchestration: pluggable")],-1),t("ul",null,[t("li",null,[t("strong",null,"Dagster"),l(" integration with "),t("code",null,"@metaxify"),l(" decorator")]),t("li",null,[t("strong",null,"Ray"),l(" for distributed GPU workloads")]),t("li",null,"Or any orchestrator that can consume a diff")],-1)])]),_:1})]),_:1},16)}}};export{b as default};
