+++ 0.1.6/slides/2026-introducing-metaxy/dist/assets/slidev/default-DMaQI0-d.js
@@ -1,0 +1,1 @@
+import{f as o,o as t,j as s}from"../modules/vue-D2_2azlL.js";import{_ as r}from"../index-B2pMgQ-r.js";const c={},a={class:"slidev-layout default"};function n(e,l){return t(),o("div",a,[s(e.$slots,"default")])}const f=r(c,[["render",n]]);export{f as I};
