+++ 0.1.6/slides/2026-introducing-metaxy/dist/assets/slidev/TimerBar.vue_vue_type_script_setup_true_lang-BE3aPCgI.js
@@ -1,0 +1,1 @@
+import{d as w,B as a,f as u,F as B,a0 as $,z as x,o,k as m,n as y,e as M,w as N,i as f,a1 as D,a2 as T,N as z,_ as F}from"../modules/vue-D2_2azlL.js";import{n as V,c as b,B as E,s as I}from"../index-B2pMgQ-r.js";const L={class:"relative flex gap-px"},G=w({__name:"CurrentProgressBar",props:{clicksContext:{},current:{}},setup(r){const e=r,s=V(),t=a(()=>e.clicksContext??s.clicksContext.value),l=a(()=>e.current??s.currentSlideNo.value),{total:n}=s;return(d,g)=>(o(),u("div",L,[(o(!0),u(B,null,$(x(n)-1,i=>(o(),u("div",{key:i,class:y(["border-x border-b border-main h-4px transition-all",i<l.value?"bg-primary border-primary":""]),style:m({width:`${1/(x(n)-1)*100}%`})},[M(D,{name:"fade"},{default:N(()=>[i===l.value?(o(),u("div",{key:0,class:"h-full bg-primary op75 transition-all",style:m({width:`${t.value.total===0?0:t.value.current/(t.value.total+1)*100}%`})},null,4)):f("v-if",!0)]),_:2},1024)],6))),128))]))}});function P(){const r=a(()=>b.timer||"stopwatch"),e=a(()=>E(b.duration).seconds),s=T(100,{controls:!0}),t=z(I,"timer"),l=a(()=>t.value?.status),n=a(()=>(s.counter.value,t.value.status==="stopped"||!t.value.startedAt?0:Date.now()-t.value.startedAt)),d=a(()=>n.value/1e3),g=a(()=>d.value/e.value*100),i=a(()=>{if(r.value==="stopwatch"&&(t.value.status==="stopped"||!t.value.startedAt))return{h:"",m:"-",s:"--",ms:"-"};const c=r.value==="countdown"?e.value*1e3-n.value:n.value;let p=Math.floor(c/1e3/60/60).toString();p==="0"&&(p="");let v=Math.floor(c/1e3/60%60).toString();p&&(v=v.padStart(2,"0"));const C=Math.floor(c/1e3%60).toString().padStart(2,"0"),S=Math.floor(c%1e3/100).toString();return{h:p,m:v,s:C,ms:S}});function k(){s.pause(),t.value={status:"stopped",slides:{},startedAt:0,pausedAt:0}}function _(){t.value&&(t.value?.status==="stopped"?(t.value.status="running",t.value.startedAt=Date.now()):t.value.status==="paused"&&(t.value.status="running",t.value.startedAt=Date.now()-(t.value.pausedAt-t.value.startedAt)),s.resume())}function h(){t.value.status="paused",t.value.pausedAt=Date.now(),s.pause()}function A(){t.value.status==="running"?h():_()}return{state:t,status:l,timer:i,reset:k,toggle:A,resume:_,pause:h,passed:d,percentage:g,duration:e,mode:r}}const R={class:"border-b mt-px border-main relative flex h-4px"},H=w({__name:"TimerBar",setup(r){const e=F(P()),s=a(()=>e.status==="stopped"?"op50":e.status==="paused"?"bg-blue":e.percentage>80?"bg-yellow":e.percentage>100?"bg-red":"bg-green");return(t,l)=>(o(),u("div",R,[e.status!=="stopped"?(o(),u("div",{key:0,class:y(["h-4px",s.value]),style:m({width:`${e.percentage}%`})},null,6)):f("v-if",!0),f(" {{ timesplits }} ")]))}});export{G as _,H as a,P as u};

[…]
