### 2026-03-19 d9be2d4
[iceberg] Address PR review: sink_iceberg, native delete, docstrings
- Use pl.LazyFrame.sink_iceberg for streaming writes (Polars >= 1.39)
- Use PyIceberg native delete_filter instead of read-filter-overwrite
- Remove layout config (flat/nested) in favor of flat table names
- Add ibis-framework and polars>=1.39 to iceberg extras
- Add catalog.close() in _close for proper resource cleanup
- Add docstrings and section headers matching Delta store structure

Made-with: Cursor
