+++ src/metaxy/ext/metadata_stores/iceberg.py
@@ -2,10 +2,10 @@
 
 from __future__ import annotations
 
-from collections.abc import Iterator, Sequence
+from collections.abc import Callable, Iterator, Sequence
 from contextlib import contextmanager
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Literal, overload
+from typing import TYPE_CHECKING, Any, overload
 
 if TYPE_CHECKING:
     from pyiceberg.catalog import Catalog
@@ -13,6 +13,7 @@ if TYPE_CHECKING:
 import narwhals as nw
 import polars as pl
 from narwhals.typing import Frame
+from packaging.version import Version
 from pydantic import Field
 
 from metaxy._decorators import public
@@ -24,6 +25,18 @@ from metaxy.models.plan import FeaturePlan
 from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
 from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
+
+
+def _strip_casts() -> Callable[[Any], Any]:
+    """Unwrap ``CAST(x AS type)`` → ``x`` so PyIceberg's row-filter parser sees plain literals."""
+    from sqlglot import exp
+
+    def _transform(node: exp.Expression) -> exp.Expression:
+        if isinstance(node, exp.Cast):
+            return node.this
+        return node
+
+    return _transform
 
 
 @public
@@ -59,10 +72,6 @@ class IcebergMetadataStoreConfig(MetadataStoreConfig):
         default=None,
         description="Properties passed to pyiceberg.catalog.load_catalog.",
     )
-    layout: Literal["flat", "nested"] = Field(
-        default="nested",
-        description="Table naming layout ('nested' or 'flat').",
-    )
 
 
 @public
@@ -70,7 +79,11 @@ class IcebergMetadataStore(MetadataStore):
     """Apache Iceberg metadata store backed by [PyIceberg](https://py.iceberg.apache.org/).
 
     Stores feature metadata in Iceberg tables managed by a PyIceberg catalog.
-    Uses the Polars versioning engine for provenance calculations.
+    It uses the Polars versioning engine for provenance calculations.
+
+    !!! tip
+        If Polars 1.39 or greater is installed, lazy Polars frames are sinked via
+        `LazyFrame.sink_iceberg`, avoiding unnecessary materialization.
 
     Example:
 
@@ -78,8 +91,9 @@ class IcebergMetadataStore(MetadataStore):
         from metaxy.ext.metadata_stores.iceberg import IcebergMetadataStore
 
         store = IcebergMetadataStore(
-            warehouse="/path/to/warehouse",
-            namespace="metaxy",
+            warehouse="s3://my-bucket/warehouse",
+            namespace="ml_features",
+            catalog_properties={"type": "glue"},
         )
         ```
     """
@@ -95,15 +109,20 @@ class IcebergMetadataStore(MetadataStore):
         catalog_name: str = "metaxy",
         catalog_properties: dict[str, str] | None = None,
         fallback_stores: list[MetadataStore] | None = None,
-        layout: Literal["flat", "nested"] = "nested",
         **kwargs: Any,
     ) -> None:
-        if layout not in ("flat", "nested"):
-            raise ValueError(f"Invalid layout: {layout}. Must be 'flat' or 'nested'.")
+        """Initialize Apache Iceberg metadata store.
 
+        Args:
+            warehouse: Warehouse directory or URI for Iceberg data files.
+            namespace: Iceberg namespace for tables (Glue Database, SQL schema, etc.).
+            catalog_name: Local identifier for the PyIceberg catalog instance.
+            catalog_properties: Properties for [`pyiceberg.catalog.load_catalog`](https://py.iceberg.apache.org/api/#load-a-catalog).
[…]
