+++ pyproject.toml
@@ -84,6 +84,8 @@ dagster = [
 
 iceberg = [
   "pyiceberg[sql-sqlite]>=0.9.0",
+  "ibis-framework>=11.0.0",
+  "polars>=1.39.0",
 ]
 lancedb = [
   "lancedb>=0.15.0",
