### 2026-03-19 252fc7f
[iceberg] Expand _strip_casts docstring explaining PyIceberg parser limitation
Made-with: Cursor
