### 2026-03-19 ffec0d0
[iceberg] Add auto_create_namespace config option (default True)
Made-with: Cursor
