+++ docs/.glossary.md
@@ -46,7 +46,7 @@
 
 *[multi-environment setups]: Metaxy codebases which have more than one distinct Python environment. | /guide/concepts/projects/#multi-environment-setups
 
-*[metaxy.lock]: A file generated by Metaxy to track external feature definitions. | /guide/concepts/definitions/external-features#metaxy-lock-file
+*[metaxy.lock]: A file generated by Metaxy to track external feature definitions. | /guide/concepts/definitions/external-features/#metaxylock-file
 
 *[fully qualified field path]: A unique identifier for a field within a feature definition, consisting of the feature name and the field name. Example: `/raw/video:frames`.
 
