### 2026-03-19 aa290e6
docs: ray example with ops
