+++ examples/example-ray/src/example_ray/__init__.py
@@ -1,0 +1,1 @@
+"""Ray Data example package for Metaxy."""

[…]
