+++ examples/example-alembic/alembic/env.py
@@ -1,0 +1,183 @@
+"""Alembic env.py for Metaxy feature tables on DuckLake (DuckDB 1.5+).
+
+DuckLake is an OLAP lakehouse format. Like many OLAP systems, it does not
+support PRIMARY KEY, UNIQUE, or RETURNING (see
+https://ducklake.select/docs/stable/duckdb/advanced_features/constraints).
+
+This env.py adapts Alembic:
+
+- Constraints are stripped from the target metadata.
+- ``version_table_pk=False`` creates alembic_version without a primary key.
+- ``insert_returning`` is disabled on the dialect.
+- DuckLake partitioning and sorted-table DDL is emitted after each
+  ``create_table`` via an ``AlembicDuckDBImpl.create_table()`` override.
+- A monkey-patch on ``_columns_query`` replaces the ``pg_collation``
+  subquery with NULL so column reflection works on DuckDB 1.5
+  (upstream: https://github.com/Mause/duckdb_engine/issues/1391).
+
+The patches in this file target duckdb-engine 0.17.x, alembic 1.17.x, and
+DuckDB 1.5.x (pinned in pyproject.toml). Other versions may require adjustments.
+"""
+
+from functools import lru_cache
+
+import metaxy as mx
+from alembic.ddl.impl import DefaultImpl
+from duckdb_engine import Dialect as DuckDBDialect  # ty: ignore[unresolved-import]
+from metaxy.ext.metadata_stores.duckdb import DuckDBMetadataStore
+from metaxy.ext.sqlmodel import filter_feature_sqlmodel_metadata
+from sqlalchemy import JSON, PrimaryKeyConstraint, UniqueConstraint, null
+from sqlalchemy.types import UserDefinedType
+from sqlmodel import SQLModel
+
+
+class VARIANT(UserDefinedType):
+    """DuckDB 1.5 VARIANT type for semi-structured data."""
+
+    cache_ok = True
+
+    def get_col_spec(self, **kw):
+        return "VARIANT"
+
+
+import example_alembic.definitions  # noqa: F401  # ty: ignore[unresolved-import]
+from alembic import context
+
+# ---------------------------------------------------------------------------
+# Monkey-patch duckdb-engine: replace the pg_collation subquery with NULL.
+# DuckDB 1.5 removed pg_collation from pg_catalog.
+# ---------------------------------------------------------------------------
+_orig_columns_query = DuckDBDialect._columns_query.__wrapped__
+
+
+@lru_cache
+def _patched_columns_query(self, schema, has_filter_names, scope, kind):
+    query = _orig_columns_query(self, schema, has_filter_names, scope, kind)
+    new_columns = [
+        null().label("collation") if col.key == "collation" else col
+        for col in query.selected_columns
+    ]
+    return query.with_only_columns(*new_columns)
+
+
+DuckDBDialect._columns_query = _patched_columns_query
+
+
+# ---------------------------------------------------------------------------
+# DuckLake DDL implementation for Alembic
+# ---------------------------------------------------------------------------
+
+# Per-table DuckLake DDL: partitioning and sorting applied after create_table.
+_DUCKLAKE_TABLE_DDL: dict[str, list[str]] = {
+    "examples__alembic_demo": [
+        "ALTER TABLE examples__alembic_demo SET PARTITIONED BY (metaxy_feature_version)",
+        "ALTER TABLE examples__alembic_demo SET SORTED BY (metaxy_updated_at)",
+    ],
+}
+
+
+class AlembicDuckDBImpl(DefaultImpl):
+    """Alembic DDL for DuckDB/DuckLake.
+
+    Disables RETURNING (unsupported by DuckLake) and emits DuckLake-specific
[…]
