+++ examples/example-alembic/src/example_alembic/__init__.py
@@ -1,0 +1,1 @@
+"""Alembic migration example with DuckLake."""
