### 2026-03-20 15b6601
[clickhouse] convert Map to Struct on reads
