+++ src/metaxy/ext/metadata_stores/iceberg.py
@@ -4,11 +4,13 @@ from __future__ import annotations
 
 from collections.abc import Callable, Iterator, Sequence
 from contextlib import contextmanager
+from functools import cached_property
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, overload
+from typing import TYPE_CHECKING, Any, NewType, overload
 
 if TYPE_CHECKING:
     from pyiceberg.catalog import Catalog
+    from pyiceberg.table import Table
 
 import narwhals as nw
 import polars as pl
@@ -26,12 +28,14 @@ from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
 from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
 
+TableIdentifier = NewType("TableIdentifier", tuple[str, str])
+"""A ``(namespace, table_name)`` pair used by PyIceberg to locate a table within a catalog."""
+
 
 def _strip_casts() -> Callable[[Any], Any]:
-    """Unwrap ``CAST(x AS type)`` → ``x`` so PyIceberg's row-filter parser sees plain literals.
-
-    PyIceberg's pyparsing-based filter parser only accepts bare literals, not SQL casts.
-    Stripping CAST wrappers leaves clean ISO-8601 strings that PyIceberg can type-coerce.
+    """Unwrap ``CAST(x AS type)`` → ``x`` to match PyIceberg's
+    `row filter syntax <https://py.iceberg.apache.org/row-filter-syntax/>`_
+    which accepts bare literals but not SQL casts.
     """
     from sqlglot import exp
 
@@ -78,7 +82,7 @@ class IcebergMetadataStoreConfig(MetadataStoreConfig):
     )
     auto_create_namespace: bool = Field(
         default=True,
-        description="Automatically create the namespace on open if it does not exist.",
+        description="Automatically create the namespace on first write if it does not exist.",
     )
 
 
@@ -126,15 +130,15 @@ class IcebergMetadataStore(MetadataStore):
             warehouse: Warehouse directory or URI for Iceberg data files.
             namespace: Iceberg namespace for tables (Glue Database, SQL schema, etc.).
             catalog_name: Local identifier for the PyIceberg catalog instance.
-            catalog_properties: Properties for [`pyiceberg.catalog.load_catalog`](https://py.iceberg.apache.org/api/#load-a-catalog).
-            auto_create_namespace: Create the namespace on open if it does not exist.
+            catalog_properties: Properties for [`pyiceberg.catalog.load_catalog`][pyiceberg.catalog.load_catalog].
+            auto_create_namespace: Create the namespace on first write if it does not exist.
             fallback_stores: Ordered list of read-only fallback stores.
             **kwargs: Forwarded to [metaxy.metadata_store.base.MetadataStore][metaxy.metadata_store.base.MetadataStore].
         """
         self.namespace = namespace
         self.catalog_name = catalog_name
         self.auto_create_namespace = auto_create_namespace
-        self._catalog: Catalog | None = None
+        self._catalog_exists = False
 
         warehouse_str = str(warehouse)
         self._is_remote = not is_local_path(warehouse_str)
@@ -178,33 +182,37 @@ class IcebergMetadataStore(MetadataStore):
             yield engine
 
     def _open(self, mode: AccessMode) -> None:  # noqa: ARG002
-        from pyiceberg.catalog import load_catalog
-
         if not self._is_remote:
             Path(self._warehouse_uri).mkdir(parents=True, exist_ok=True)
-
-        self._catalog = load_catalog(
-            self.catalog_name,
-            **self._catalog_properties,
-        )
-        if self.auto_create_namespace:
-            self._catalog.create_namespace_if_not_exists(self.namespace)
+        _ = self.catalog
 
[…]
