### 2026-03-25 #10141 @geoHeil
I checked your database and found these people:

Top Recommendations

  1. Simon Provost — simonprovost

  - Expertise: AI/Data PhD, reviewed the closely related ReciPies JOSS paper
  (reproducible ML pipelines, feature engineering)
  - Why: Proven JOSS reviewer on the most topically similar accepted paper. Low
  COI risk.

  2. Francesco Bruzzesi — FBruzzesi

  - Expertise: Second-largest Narwhals contributor, dataframe interoperability
  - Why: Deep knowledge of the cross-dataframe abstraction layer Metaxy builds
  upon. Works at intella.tech (Italy), no institutional overlap. Low COI risk.

  3. Stefan Krawczyk — skrawcz

  - Expertise: Co-creator of Hamilton (compared in the paper), DAG-based feature
   computation and lineage
  - Why: Firsthand knowledge of the problem space and design trade-offs. Medium
  COI risk — Hamilton is discussed as a related tool. Flag to editor.

  4. Nick Crews — NickCrews

  - Expertise: Ibis contributor (229 commits), DuckDB expertise
  - Why: Can evaluate SQL-pushed version computation and backend-agnostic
  design. Works at ShipCreekGroup (Alaska), no institutional overlap. Low COI
  risk.

  5. Marios Papachristou — papachristoumarios

  - Expertise: Assistant Professor at Arizona State, previously reviewed the
  Feature-engine JOSS paper
  - Why: Academic perspective on feature engineering software with JOSS
  reviewing experience. Low COI risk.

  Alternates

  - Panagiotis Anagnostou (panagiotisanagnostou) — Also reviewed the ReciPies
  paper, complementary perspective
  - Thierry Jean (zilto) — Third-largest Hamilton contributor, practical DAG
  dataflow knowledge. Low-to-medium COI risk.
  - Robin van de Water (rvandewater) — Author of ReciPies, deeply understands
  this space. Low COI risk (Metaxy doesn't cite ReciPies).
