+++ src/metaxy/config/__init__.py
@@ -1,0 +1,10 @@
+from metaxy.config.models.metaxy_config import InvalidConfigError, MetaxyConfig
+from metaxy.config.models.plugin_config import PluginConfig
+from metaxy.config.models.store_config import StoreConfig
+
+__all__ = [
+    "InvalidConfigError",
+    "MetaxyConfig",
+    "StoreConfig",
+    "PluginConfig",
+]
