+++ src/metaxy/config/models/metaxy_config.py
@@ -1,14 +1,10 @@
-"""Configuration system for Metaxy using pydantic-settings."""
-# pyright: reportImportCycles=false
-
 import os
 import warnings
 from collections.abc import Iterator
 from contextlib import contextmanager
 from contextvars import ContextVar
-from functools import cached_property
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Literal, TypeAlias, TypeVar, overload
+from typing import TYPE_CHECKING, Any, Literal, TypeVar, overload
 
 import tomli_w
 from pydantic import Field as PydanticField
@@ -21,7 +17,10 @@ from pydantic_settings import (
 from typing_extensions import Self
 
 from metaxy._decorators import public
-from metaxy._toml_source import MetaxyTomlSource, discover_config_with_parents
+from metaxy.config.metaxy_source import MetaxyTomlSource, discover_config_with_parents
+from metaxy.config.models.plugin_config import PluginConfig
+from metaxy.config.models.store_config import StoreConfig
+from metaxy.config.utils import _collect_dict_keys, _remove_none_values
 from metaxy.models.feature_selection import FeatureSelection
 
 if TYPE_CHECKING:
@@ -29,18 +28,22 @@ if TYPE_CHECKING:
         MetadataStore,
     )
 
-T = TypeVar("T")
+
+BUILTIN_PLUGINS = {}
 
 
-def _collect_dict_keys(d: dict[str, Any], prefix: str = "") -> list[str]:
-    """Recursively collect all keys from a nested dict as dot-separated paths."""
-    keys = []
-    for key, value in d.items():
-        full_key = f"{prefix}.{key}" if prefix else key
-        keys.append(full_key)
-        if isinstance(value, dict):
-            keys.extend(_collect_dict_keys(value, full_key))
-    return keys
+PluginConfigT = TypeVar("PluginConfigT", bound=PluginConfig)
+StoreTypeT = TypeVar("StoreTypeT", bound="MetadataStore")
+
+# Global config visible to all threads, set via MetaxyConfig.set() / .load().
+# ContextVar overlay for per-context overrides via MetaxyConfig.use().
+# get() checks the ContextVar first, falling back to the global.
+_global_config: "MetaxyConfig | None" = None
+_config_override: ContextVar["MetaxyConfig | None"] = ContextVar("_config_override", default=None)
+
+# Used by load() to pass the config file path into settings_customise_sources
+# without monkey-patching the classmethod.
+_toml_file_override: ContextVar[Path | None] = ContextVar("_toml_file_override")
 
 
 @public
@@ -82,116 +85,6 @@ class InvalidConfigError(Exception):
             An InvalidConfigError with context from the config.
         """
         return cls(message, config_file=config._config_file)
-
-
-MetadataStoreT: TypeAlias = type["MetadataStore"]
-
-
-@public
-class StoreConfig(BaseSettings):
-    """Configuration options for metadata stores."""
-
-    model_config = SettingsConfigDict(
-        extra="forbid",  # Only type and config fields allowed
-        frozen=True,
-    )
-
-    # Store the import path as string (internal field)
-    # Uses alias="type" so TOML and constructor use "type"
-    # Annotated as str | type to allow passing class objects directly
-    type: str = PydanticField(
-        description='Full import path to metadata store class (e.g., `"metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore"`)',
-    )
-
-    config: dict[str, Any] = PydanticField(
-        default_factory=dict,
-        description="Store-specific configuration parameters (constructor kwargs). Includes `fallback_stores`, database connection parameters, etc.",
-    )
-
-    @field_validator("type", mode="before")
-    @classmethod
[…]
