### 2026-03-25 ffa78c7
fix: do not validate incomplete MetaxyConfig when `extend` is set (#1087)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
