+++ tests/test_toml_source.py
@@ -1,0 +1,214 @@
+"""Tests for MetaxyTomlSource."""
+
+from pathlib import Path
+
+import pytest
+from pydantic_settings import BaseSettings
+
+from metaxy._toml_source import MetaxyTomlSource, _discover_config_in_dir
+
+
+class _DummySettings(BaseSettings):
+    pass
+
+
+STORE_TYPE = "metaxy.ext.metadata_stores.delta.DeltaMetadataStore"
+
+
+# --- Basic loading ---
+
+
+def test_load_metaxy_toml(tmp_path: Path) -> None:
+    config_file = tmp_path / "metaxy.toml"
+    config_file.write_text('project = "test"\nstore = "dev"\n')
+
+    source = MetaxyTomlSource(_DummySettings, config_file=config_file)
+    data = source()
+
+    assert data["project"] == "test"
+    assert data["store"] == "dev"
+
+
+def test_load_pyproject_toml(tmp_path: Path) -> None:
+    config_file = tmp_path / "pyproject.toml"
+    config_file.write_text('[project]\nname = "foo"\n\n[tool.metaxy]\nproject = "from-pyproject"\n')
+
+    source = MetaxyTomlSource(_DummySettings, config_file=config_file)
+    data = source()
+
+    assert data["project"] == "from-pyproject"
+    assert "tool" not in data
+
+
+def test_env_var_expansion(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+    monkeypatch.setenv("TEST_DB_HOST", "prod.db.example.com")
+    config_file = tmp_path / "metaxy.toml"
+    config_file.write_text('[stores.prod.config]\nhost = "${TEST_DB_HOST}"\n')
+
+    source = MetaxyTomlSource(_DummySettings, config_file=config_file)
+    data = source()
+
+    assert data["stores"]["prod"]["config"]["host"] == "prod.db.example.com"
+
+
+def test_no_config_file_returns_empty() -> None:
+    source = MetaxyTomlSource(_DummySettings, config_file=None)
+    assert source() == {}
+
+
+# --- Extend chain resolution ---
+
+
+def test_extend_child_overrides_parent(tmp_path: Path) -> None:
+    parent = tmp_path / "parent.toml"
+    parent.write_text('project = "parent"\nstore = "parent_store"\n')
+
+    child = tmp_path / "child.toml"
+    child.write_text('extend = "parent.toml"\nproject = "child"\n')
+
+    source = MetaxyTomlSource(_DummySettings, config_file=child)
+    data = source()
+
+    assert data["project"] == "child"
+    assert data["store"] == "parent_store"
+    assert "extend" not in data
+
+
+def test_extend_stores_merged(tmp_path: Path) -> None:
+    parent = tmp_path / "parent.toml"
+    parent.write_text(
+        f'[stores.prod]\ntype = "{STORE_TYPE}"\n[stores.prod.config]\nroot_path = "/prod"\n'
+        f'[stores.parent_only]\ntype = "{STORE_TYPE}"\n'
+    )
+    child = tmp_path / "child.toml"
+    child.write_text(
+        f'extend = "parent.toml"\n[stores.prod]\ntype = "{STORE_TYPE}"\n[stores.prod.config]\nroot_path = "/override"\n'
+    )
+
+    source = MetaxyTomlSource(_DummySettings, config_file=child)
+    data = source()
+
+    assert data["stores"]["prod"]["config"]["root_path"] == "/override"
+    assert "parent_only" in data["stores"]
+
+
+def test_extend_entrypoints_appended(tmp_path: Path) -> None:
+    parent = tmp_path / "parent.toml"
+    parent.write_text('entrypoints = ["parent.mod"]\n')
[…]
