+++ src/metaxy/config.py
@@ -2,16 +2,14 @@
 # pyright: reportImportCycles=false
 
 import os
-import re
 import warnings
-from collections.abc import Iterator, Mapping, Sequence
+from collections.abc import Iterator
 from contextlib import contextmanager
 from contextvars import ContextVar
 from functools import cached_property
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, TypeAlias, TypeVar, overload
 
-import tomli
 import tomli_w
 from pydantic import Field as PydanticField
 from pydantic import PrivateAttr, field_validator, model_validator
@@ -23,6 +21,7 @@ from pydantic_settings import (
 from typing_extensions import Self
 
 from metaxy._decorators import public
+from metaxy._toml_source import MetaxyTomlSource, discover_config_with_parents
 from metaxy.models.feature_selection import FeatureSelection
 
 if TYPE_CHECKING:
@@ -31,9 +30,6 @@ if TYPE_CHECKING:
     )
 
 T = TypeVar("T")
-
-# Pattern for ${VAR} or ${VAR:-default} syntax
-_ENV_VAR_PATTERN = re.compile(r"\$\{([^}:]+)(?::-([^}]*))?\}")
 
 
 def _collect_dict_keys(d: dict[str, Any], prefix: str = "") -> list[str]:
@@ -86,99 +82,6 @@ class InvalidConfigError(Exception):
             An InvalidConfigError with context from the config.
         """
         return cls(message, config_file=config._config_file)
-
-
-def _expand_env_vars(value: Any) -> Any:
-    """Recursively expand environment variables in config values.
-
-    Supports:
-    - ${VAR} - substitutes with environment variable value, empty string if not set
-    - ${VAR:-default} - substitutes with environment variable value, or default if not set
-
-    Args:
-        value: The value to expand (can be string, dict, list, or other)
-
-    Returns:
-        The value with environment variables expanded
-    """
-    if isinstance(value, str):
-
-        def replace_match(match: re.Match[str]) -> str:
-            var_name = match.group(1)
-            default = match.group(2)  # None if no default specified
-            env_value = os.environ.get(var_name)
-            if env_value is not None:
-                return env_value
-            elif default is not None:
-                return default
-            else:
-                return ""
-
-        return _ENV_VAR_PATTERN.sub(replace_match, value)
-    elif isinstance(value, Mapping):
-        return {k: _expand_env_vars(v) for k, v in value.items()}
-    elif isinstance(value, Sequence):
-        return [_expand_env_vars(item) for item in value]
-    else:
-        return value
-
-
-class TomlConfigSettingsSource(PydanticBaseSettingsSource):
-    """Custom settings source for TOML configuration files.
-
-    Auto-discovers configuration in this order:
-    1. Explicit file path if provided
-    2. metaxy.toml in current directory (preferred)
-    3. pyproject.toml [tool.metaxy] section (fallback)
-    4. No config (returns empty dict)
-    """
-
-    def __init__(self, settings_cls: type[BaseSettings], toml_file: Path | None = None):
-        super().__init__(settings_cls)
-        self.toml_file = toml_file or self._discover_config_file()
-        self.toml_data = self._load_toml()
-
-    def _discover_config_file(self) -> Path | None:
-        """Auto-discover config file."""
-        # Prefer metaxy.toml
-        if Path("metaxy.toml").exists():
-            return Path("metaxy.toml")
-
-        # Fallback to pyproject.toml
-        if Path("pyproject.toml").exists():
-            return Path("pyproject.toml")
-
-        return None
-
-    def _load_toml(self) -> dict[str, Any]:
-        """Load TOML file and extract metaxy config.
-
[…]

[…]
