### 2026-03-25 fc43094
refactor: redesign MetadataStore to be composed of storage and compute
