### 2026-03-25 4d46a7b
refactor: redesign MetadataStore to be composed of storage and compute
