+++ src/metaxy/ext/clickhouse/metadata_store.py
@@ -1,75 +1,18 @@
-"""This module implements [`IbisMetadataStore`][metaxy.ext.ibis.metadata_store.IbisMetadataStore] for ClickHouse.
+"""ClickHouse metadata store — thin wrapper that composes ClickHouseEngine + IbisStorageConfig."""
 
-It takes care of some ClickHouse-specific logic such as `nw.Struct` type conversion against ClickHouse types such as `Map(K,V)`."""
+from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, cast
+from typing import Any, cast
 
-import narwhals as nw
-from narwhals.typing import FrameT
 from pydantic import Field
 
-if TYPE_CHECKING:
-    import ibis
-    from ibis.expr.schema import Schema as IbisSchema
-
-    from metaxy.metadata_store.base import MetadataStore
-
 from metaxy._decorators import public
-from metaxy.ext.ibis.metadata_store import (
-    Frame,
-    IbisMetadataStore,
-    IbisMetadataStoreConfig,
+from metaxy.ext.ibis.engine import (
+    IbisStorageConfig,
+    IbisStoreBackcompat,
 )
-from metaxy.ext.ibis.versioning import IbisVersioningEngine
-from metaxy.models.types import FeatureKey
-from metaxy.versioning.types import HashAlgorithm
-
-
-class ClickHouseVersioningEngine(IbisVersioningEngine):
-    """Versioning engine for ClickHouse backend.
-
-    Overrides concat_strings_over_groups to use ClickHouse-compatible
-    syntax with collect() (groupArray) + arrayStringConcat.
-    """
-
-    def concat_strings_over_groups(
-        self,
-        df: FrameT,
-        source_column: str,
-        target_column: str,
-        group_by_columns: list[str],
-        order_by_columns: list[str],
-        separator: str = "|",
-    ) -> FrameT:
-        """Concatenate string values within groups using ClickHouse window functions.
-
-        Uses collect() (groupArray) + arrayStringConcat instead of group_concat().over()
-        which generates invalid SQL for ClickHouse.
-        """
-        import ibis
-        import ibis.expr.datatypes as dt
-        import ibis.expr.types
-
-        assert df.implementation == nw.Implementation.IBIS, "Only Ibis DataFrames are accepted"
-        ibis_table: ibis.expr.types.Table = cast(ibis.expr.types.Table, df.to_native())
-
-        @ibis.udf.scalar.builtin
-        def arrayStringConcat(arr: dt.Array[dt.String], sep: str) -> str:
-            """ClickHouse arrayStringConcat() function."""
-            ...
-
-        effective_order_by = order_by_columns if order_by_columns else group_by_columns
-        window = ibis.window(
-            group_by=group_by_columns,
-            order_by=[ibis_table[col] for col in effective_order_by],
-        )
-
-        arr_expr = ibis_table[source_column].cast("string").collect().over(window)
-        concat_expr = arrayStringConcat(arr_expr, separator)
-
-        ibis_table = ibis_table.mutate(**{target_column: concat_expr})
-
-        return cast(FrameT, nw.from_native(ibis_table))
+from metaxy.ext.ibis.metadata_store import IbisMetadataStoreConfig
+from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
 
 
 @public
@@ -94,555 +37,117 @@ class ClickHouseMetadataStoreConfig(IbisMetadataStoreConfig):
 
 
 @public
-class ClickHouseMetadataStore(IbisMetadataStore):
-    """
-    [ClickHouse](https://clickhouse.com/) metadata store using [Ibis](https://ibis-project.org/) backend.
[…]
