### 2026-03-25 0923c9c
refactor: redesign MetadataStore to be composed of storage and compute
