+++ src/metaxy/ext/duckdb/metadata_store.py
@@ -1,38 +1,28 @@
-"""DuckDB metadata store - thin wrapper around IbisMetadataStore."""
+"""DuckDB metadata store -- thin wrapper that composes DuckDBEngine + IbisStorageConfig."""
 
-from collections.abc import Iterable, Sequence
-from contextlib import contextmanager
+from __future__ import annotations
+
+from collections.abc import Sequence
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
-from urllib.parse import urlsplit
+from typing import Any, cast
 
-from pydantic import BaseModel, Field
-from typing_extensions import Self
-
-if TYPE_CHECKING:
-    from metaxy.metadata_store.base import MetadataStore
+from pydantic import Field
 
 from metaxy._decorators import public
-from metaxy.ext.duckdb import (
+from metaxy.ext.duckdb.engine import ExtensionSpec, _normalise_extensions
+from metaxy.ext.duckdb.handlers.ducklake import (
+    DuckDBDuckLakeHandler,
     DuckDBPyConnection,
     DuckLakeAttachmentManager,
     DuckLakeConfig,
-    MotherDuckCatalogConfig,
 )
-from metaxy.ext.ibis.metadata_store import IbisMetadataStore, IbisMetadataStoreConfig
-from metaxy.metadata_store.types import AccessMode
-from metaxy.versioning.types import HashAlgorithm
-
-
-@public
-class ExtensionSpec(BaseModel):
-    """DuckDB extension specification accepted by DuckDBMetadataStore."""
-
-    name: str
-    repository: str = "core"
-    """Extension repository: `"core"` for official extensions, `"community"` for community extensions."""
-    init_sql: Sequence[str] = ()
-    """SQL statements to execute immediately after loading the extension."""
+from metaxy.ext.ibis.engine import (
+    DuckLakeStorageConfig,
+    IbisStorageConfig,
+    IbisStoreBackcompat,
+)
+from metaxy.ext.ibis.metadata_store import IbisMetadataStoreConfig
+from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
 
 
 @public
@@ -67,25 +57,9 @@ class DuckDBMetadataStoreConfig(IbisMetadataStoreConfig):
     )
 
 
-def _normalise_extensions(
-    extensions: Iterable[str | ExtensionSpec],
-) -> list[ExtensionSpec]:
-    """Coerce extension inputs into ExtensionSpec instances."""
-    normalised: list[ExtensionSpec] = []
-    for ext in extensions:
-        if isinstance(ext, str):
-            normalised.append(ExtensionSpec(name=ext))
-        elif isinstance(ext, ExtensionSpec):
-            normalised.append(ext)
-        else:
-            raise TypeError(f"DuckDB extensions must be strings or ExtensionSpec instances, got {type(ext).__name__}.")
-    return normalised
-
-
 @public
-class DuckDBMetadataStore(IbisMetadataStore):
-    """
-    [DuckDB](https://duckdb.org/) metadata store using [Ibis](https://ibis-project.org/) backend.
+class DuckDBMetadataStore(IbisStoreBackcompat):
+    """[DuckDB](https://duckdb.org/) metadata store using [Ibis](https://ibis-project.org/) backend.
 
     Example: Local File
         ```py
@@ -94,271 +68,160 @@ class DuckDBMetadataStore(IbisMetadataStore):
 
     Example: With extensions
         ```py
-        # With extensions
         store = DuckDBMetadataStore("md:my_database", extensions=["spatial"])
         ```
     """
 
     def __init__(
         self,
-        database: str | Path,
+        database: str | Path = "",  # noqa: ARG002
         *,
[…]
