### 2026-03-25 b80fff4
refactor: redesign MetadataStore to be composed of storage and compute
