+++ src/metaxy/metadata_store/base.py
@@ -1,8 +1,7 @@
-"""Abstract base class for metadata storage backends."""
+"""Base class for metadata storage backends."""
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
 from collections.abc import Iterator, Mapping, Sequence
 from contextlib import AbstractContextManager, contextmanager
 from datetime import datetime, timezone
@@ -17,12 +16,14 @@ from typing_extensions import Self
 
 from metaxy._decorators import public
 from metaxy.config import MetaxyConfig
+from metaxy.metadata_store.compute_engine import ComputeEngine
 from metaxy.metadata_store.exceptions import (
     FeatureNotFoundError,
     StoreNotOpenError,
     SystemDataNotFoundError,
     VersioningEngineMismatchError,
 )
+from metaxy.metadata_store.storage_config import StorageConfig
 from metaxy.metadata_store.system.keys import METAXY_SYSTEM_KEY_PREFIX
 from metaxy.metadata_store.types import AccessMode
 from metaxy.metadata_store.utils import (
@@ -77,6 +78,13 @@ def _is_map_column(df: Frame, col_name: str) -> bool:
 
 # TypeVar for config types - used for typing from_config method
 MetadataStoreConfigT = TypeVar("MetadataStoreConfigT", bound="MetadataStoreConfig")
+
+
+def _unpickle_store(cls: type[MetadataStore], state: dict[str, Any]) -> MetadataStore:
+    """Reconstruct a MetadataStore from pickled state, bypassing __new__/__init__."""
+    obj = object.__new__(cls)
+    obj.__dict__.update(state)
+    return obj
 
 
 @public
@@ -161,20 +169,16 @@ def _cast_present_system_columns(
 
 
 @public
-class MetadataStore(ABC):
-    """
-    Abstract base class for metadata storage backends.
-    """
+class MetadataStore:
+    """Base class for metadata storage backends."""
 
     # Subclasses can override this to disable auto_create_tables warning
     # Set to False for stores where table creation is not applicable
     _should_warn_auto_create_tables: bool = True
 
-    # Subclasses must define the versioning engine class to use
-    versioning_engine_cls: type[VersioningEngine]
-
     def __init__(
         self,
+        engine: ComputeEngine,
         *,
         name: str | None = None,
         hash_algorithm: HashAlgorithm | None = None,
@@ -182,6 +186,7 @@ class MetadataStore(ABC):
         fallback_stores: Sequence[MetadataStore | str] | None = None,
         auto_create_tables: bool | None = None,
         materialization_id: str | None = None,
+        storage: Sequence[StorageConfig] | None = None,
     ):
         """
         Initialize the metadata store.
@@ -213,6 +218,10 @@ class MetadataStore(ABC):
             materialization_id: Optional external orchestration ID.
                 If provided, all metadata writes will include this ID in the `metaxy_materialization_id` column.
                 Can be overridden per [`MetadataStore.write`][metaxy.MetadataStore.write] call.
+            engine: Optional compute engine for delegation.
+                When set, I/O methods delegate to the engine.
+            storage: Optional sequence of storage configurations.
+                Used by the engine to locate feature data.
 
         Raises:
             ValueError: If fallback stores use different hash algorithms or truncation lengths
@@ -227,6 +236,9 @@ class MetadataStore(ABC):
         self._exit_stack: list[AbstractContextManager[Any]] = []
         self._transaction_timestamp: datetime | None = None
         self._soft_delete_in_progress: bool = False
+        self._engine: ComputeEngine = engine
+        self._storage: list[StorageConfig] = list(storage) if storage else []
+        self.versioning_engine_cls = engine.versioning_engine_cls
 
[…]
