+++ tests/migrations/test_migration_integration.py
@@ -1,1030 +1,0 @@
-"""Integration tests for new migration system with multi-snapshot scenarios.
-
-These tests use TempFeatureModule to create realistic graph evolution scenarios
-and test the full migration workflow: detect → execute → verify.
-"""
-
-from datetime import datetime, timezone
-
-import polars as pl
-import pytest
-from metaxy_testing import TempFeatureModule, add_metaxy_provenance_column
-from metaxy_testing.models import SampleFeatureSpec
-from syrupy.assertion import SnapshotAssertion
-
-from metaxy import (
-    FeatureDep,
-    FeatureKey,
-    FieldDep,
-    FieldKey,
-    FieldSpec,
-)
-from metaxy._utils import collect_to_polars
-from metaxy.config import MetaxyConfig
-from metaxy.ext.metadata_stores.delta import DeltaMetadataStore
-from metaxy.metadata_store.system import SystemTableStorage
-from metaxy.migrations import MigrationExecutor, detect_diff_migration
-from metaxy.models.feature import FeatureGraph
-
-
-@pytest.fixture(autouse=True)
-def setup_default_config():
-    """Set up default MetaxyConfig for all tests so features use project='default'."""
-    config = MetaxyConfig(project="default", stores={})
-    MetaxyConfig.set(config)
-    yield
-    MetaxyConfig.reset()
-
-
-def migrate_store_to_graph(
-    source_store: DeltaMetadataStore,
-    target_graph: FeatureGraph,
-) -> DeltaMetadataStore:
-    """Create new store with target graph context but source store's data.
-
-    This includes system tables (snapshots, migrations, events) so that
-    migration detection can find the previous snapshot.
-
-    For DeltaMetadataStore, we just create a new store pointing to the same
-    root_path since data is persisted on disk.
-    """
-    # DeltaMetadataStore persists data to disk, so we can just create a new
-    # instance pointing to the same location
-    return DeltaMetadataStore(root_path=source_store._root_uri)
-
-
-@pytest.fixture
-def simple_graph_v1():
-    """Simple graph with one feature."""
-    temp_module = TempFeatureModule("test_integration_simple_v1")
-
-    spec_v1 = SampleFeatureSpec(
-        key=FeatureKey(["test_integration", "simple"]),
-        fields=[
-            FieldSpec(key=FieldKey(["default"]), code_version="1"),
-        ],
-    )
-
-    temp_module.write_features({"Simple": spec_v1})
-    yield temp_module.graph
-    temp_module.cleanup()
-
-
-@pytest.fixture
-def simple_graph_v2():
-    """Simple graph with code_version changed."""
-    temp_module = TempFeatureModule("test_integration_simple_v2")
-
-    spec_v2 = SampleFeatureSpec(
-        key=FeatureKey(["test_integration", "simple"]),
-        fields=[
-            FieldSpec(key=FieldKey(["default"]), code_version="2"),  # Changed!
-        ],
-    )
-
-    temp_module.write_features({"Simple": spec_v2})
-    yield temp_module.graph
-    temp_module.cleanup()
-
-
-@pytest.fixture
-def upstream_downstream_v1():
-    """Graph with upstream and downstream features."""
-    temp_module = TempFeatureModule("test_integration_chain_v1")
-
-    upstream_spec = SampleFeatureSpec(
-        key=FeatureKey(["test_integration", "upstream"]),
-        fields=[
-            FieldSpec(key=FieldKey(["default"]), code_version="1"),
-        ],
-    )
-
-    downstream_spec = SampleFeatureSpec(
-        key=FeatureKey(["test_integration", "downstream"]),
-        deps=[FeatureDep(feature=FeatureKey(["test_integration", "upstream"]))],
-        fields=[
[…]
