+++ src/metaxy/migrations/models.py
@@ -1,516 +1,0 @@
-"""Type-safe migration models with Python class paths.
-
-Migration system features:
-- Python class paths for polymorphic deserialization via discriminated unions
-- Struct-based storage for graph data
-- Event-based status tracking
-"""
-
-from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Annotated, Any, Literal
-
-import pydantic
-from pydantic import AliasChoices, TypeAdapter
-from pydantic import Field as PydanticField
-from pydantic.types import AwareDatetime
-
-if TYPE_CHECKING:
-    from metaxy.graph.diff.diff_models import GraphDiff
-    from metaxy.metadata_store.base import MetadataStore
-
-
-class OperationConfig(pydantic.BaseModel):
-    """Configuration for a migration operation.
-
-    The structure directly matches the YAML - no nested 'config' field.
-    All operation-specific fields are defined directly on the operation class.
-
-    Required fields:
-    - type: Full Python class path to operation class (e.g., "metaxy.migrations.ops.DataVersionReconciliation")
-
-    Optional fields:
-    - features: List of feature keys this operation applies to
-      - Required for FullGraphMigration
-      - Optional for DiffMigration (features determined from graph diff)
-    - All other fields are operation-specific and defined by the operation class
-
-    Example (FullGraphMigration):
-        {
-            "type": "anam_data_utils.migrations.PostgreSQLBackfill",
-            "features": ["raw_video", "scene"],
-            "postgresql_url": "postgresql://...",  # Direct field, no nesting
-            "batch_size": 1000
-        }
-
-    Example (DiffMigration):
-        {
-            "type": "metaxy.migrations.ops.DataVersionReconciliation",
-        }
-
-    Note:
-        The 'type' field is stored as a string and only imported when the operation
-        needs to be instantiated via the Migration.operations property. This allows
-        reading migration configurations even if the operation classes have been
-        renamed or don't exist yet.
-    """
-
-    model_config = pydantic.ConfigDict(extra="allow")
-
-    type: str  # Python class path as string - imported lazily when needed
-    features: list[str] = pydantic.Field(default_factory=list)
-
-
-class Migration(pydantic.BaseModel, ABC):
-    """Abstract base class for all migrations.
-
-    Subclasses must define:
-    - migration_type: Literal field with class path for discriminated union deserialization
-    - execute(): Migration logic
-    - get_affected_features(): Return list of affected feature keys
-
-    The migration_type field is used as a discriminator for automatic polymorphic deserialization.
-
-    All migrations form a chain via parent IDs (like git commits):
-    - parent: ID of parent migration ("initial" for first migration)
-    """
-
-    model_config = pydantic.ConfigDict(extra="forbid")
-
-    # Use AliasChoices to accept both "id" (from generated YAML) and "migration_id" (from tests/manual YAML)
-    migration_id: str = PydanticField(validation_alias=AliasChoices("id", "migration_id"), serialization_alias="id")
-    parent: str  # Parent migration ID or "initial"
-    created_at: AwareDatetime
-
-    @abstractmethod
-    def execute(
-        self,
-        store: "MetadataStore",
-        project: str,
-        *,
-        dry_run: bool = False,
-    ) -> "MigrationResult":
-        """Execute the migration.
-
-        Args:
-            store: Metadata store to operate on
-            project: Project name for event tracking
-            dry_run: If True, only validate without executing
-
-        Returns:
-            MigrationResult with execution details
-
-        Raises:
-            Exception: If migration fails
-        """
-        pass
-
-    @abstractmethod
-    def get_affected_features(self, store: "MetadataStore", project: str | None) -> list[str]:
-        """Get list of affected feature keys in topological order.
-
-        Args:
[…]
