### 2026-03-26 9378994
feat(duckdb): support Lance format from DuckDB via DuckDBLanceHandler
