+++ src/metaxy/ext/duckdb/handlers/lance.py
@@ -1,0 +1,155 @@
+"""DuckDB + Lance storage handler.
+
+Reads and writes Lance tables through DuckDB's `lance` community extension.
+Uses `__lance_scan()` for reads and `COPY TO (FORMAT LANCE)` for writes.
+
+```python
+from metaxy.metadata_store.base import MetadataStore
+from metaxy.metadata_store.storage_config import LanceStorageConfig
+from metaxy.ext.metadata_stores.duckdb import DuckDBEngine, ExtensionSpec
+
+handler = DuckDBLanceHandler()
+engine = DuckDBEngine(
+    database=":memory:",
+    handler=handler,
+    extensions=[*handler.required_extensions()],
+)
+store = MetadataStore(
+    engine=engine,
+    storage=[LanceStorageConfig(format="lance", location="/data/lance")],
+)
+```
+"""
+
+from __future__ import annotations
+
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, Any
+
+import narwhals as nw
+from narwhals.typing import Frame
+
+from metaxy.ext.metadata_stores.duckdb import ExtensionSpec
+from metaxy.metadata_store.storage_config import LanceStorageConfig, StorageConfig
+from metaxy.metadata_store.storage_handler import StorageHandler
+from metaxy.models.types import FeatureKey
+
+if TYPE_CHECKING:
+    from ibis.backends.sql import SQLBackend
+
+
+class DuckDBLanceHandler(StorageHandler["SQLBackend"]):
+    """Reads and writes Lance tables through DuckDB's `lance` extension.
+
+    Uses `__lance_scan()` for reads and `COPY TO (FORMAT LANCE)` for writes,
+    keeping all I/O inside DuckDB.
+    """
+
+    def can_read(self, storage_config: StorageConfig, key: FeatureKey) -> bool:  # noqa: ARG002
+        return isinstance(storage_config, LanceStorageConfig)
+
+    def can_write(self, storage_config: StorageConfig, key: FeatureKey) -> bool:  # noqa: ARG002
+        return isinstance(storage_config, LanceStorageConfig)
+
+    def required_extensions(self) -> list[ExtensionSpec]:
+        """DuckDB extensions this handler needs."""
+        return [ExtensionSpec(name="lance", repository="community")]
+
+    # -- helpers --------------------------------------------------------------
+
+    def _table_uri(self, storage_config: LanceStorageConfig, key: FeatureKey) -> str:
+        location = storage_config.location.rstrip("/")
+        return f"{location}/{key.table_name}.lance"
+
+    # -- CRUD -----------------------------------------------------------------
+
+    def read(
+        self,
+        conn: SQLBackend,
+        storage_config: StorageConfig,
+        key: FeatureKey,
+        *,
+        filters: Sequence[nw.Expr] | None = None,
+        columns: Sequence[str] | None = None,
+    ) -> nw.LazyFrame[Any] | None:
+        assert isinstance(storage_config, LanceStorageConfig)
+        uri = self._table_uri(storage_config, key)
+
+        raw = conn.con  # ty: ignore[unresolved-attribute]
+        try:
+            raw.execute(f"SELECT 1 FROM __lance_scan('{uri}') LIMIT 0")
+        except Exception:
+            return None
+
+        view_name = f"_lance_{key.table_name}"
+        raw.execute(f"CREATE OR REPLACE TEMP VIEW \"{view_name}\" AS SELECT * FROM __lance_scan('{uri}')")
+
+        ibis_table = conn.table(view_name)
+        nw_frame: nw.LazyFrame[Any] = nw.from_native(ibis_table, eager_only=False)  # type: ignore[assignment]
+
+        if filters:
[…]
