### 2026-03-26 b1056d5
feat(duckdb): support Lance format from DuckDB via DuckDBLanceHandler
