### 2026-03-26 9f47e99
feat(iceberg): add Apache Iceberg metadata store (#1033)
## Motivation

Closes #1032

## Summary
- Add IcebergMetadataStore backend using PyIceberg with Polars
versioning engine
- Support flat and nested table naming layouts, schema evolution, custom
namespaces
- Add tests, documentation, and pytest marker for Iceberg

---------

Co-authored-by: Ashutosh Agarwal <ashutosh.agarwal@Ashutoshs-MacBook-Pro.local>
