+++ src/metaxy/ext/metadata_stores/iceberg.py
@@ -1,0 +1,355 @@
+"""Apache Iceberg metadata store implemented with PyIceberg."""
+
+from __future__ import annotations
+
+from collections.abc import Callable, Iterator, Sequence
+from contextlib import contextmanager
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, NewType, overload
+
+if TYPE_CHECKING:
+    from pyiceberg.catalog import Catalog
+    from pyiceberg.table import Table
+
+import narwhals as nw
+import polars as pl
+from narwhals.typing import Frame
+from packaging.version import Version
+from pydantic import Field
+
+from metaxy._decorators import public
+from metaxy._utils import collect_to_polars
+from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
+from metaxy.metadata_store.types import AccessMode
+from metaxy.metadata_store.utils import is_local_path
+from metaxy.models.plan import FeaturePlan
+from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.versioning.polars import PolarsVersioningEngine
+from metaxy.versioning.types import HashAlgorithm
+
+TableIdentifier = NewType("TableIdentifier", tuple[str, str])
+"""A ``(namespace, table_name)`` pair used by PyIceberg to locate a table within a catalog."""
+
+
+def _strip_casts() -> Callable[[Any], Any]:
+    """Unwrap ``CAST(x AS type)`` → ``x`` to match PyIceberg's
+    `row filter syntax <https://py.iceberg.apache.org/row-filter-syntax/>`_
+    which accepts bare literals but not SQL casts.
+    """
+    from sqlglot import exp
+
+    def _transform(node: exp.Expression) -> exp.Expression:
+        if isinstance(node, exp.Cast):
+            return node.this
+        return node
+
+    return _transform
+
+
+@public
+class IcebergMetadataStoreConfig(MetadataStoreConfig):
+    """Configuration for IcebergMetadataStore.
+
+    Example:
+        ```toml title="metaxy.toml"
+        [stores.dev]
+        type = "metaxy.ext.metadata_stores.iceberg.IcebergMetadataStore"
+
+        [stores.dev.config]
+        warehouse = "/path/to/warehouse"
+        namespace = "metaxy"
+
+        [stores.dev.config.catalog_properties]
+        type = "sql"
+        ```
+    """
+
+    warehouse: str | Path = Field(
+        description="Warehouse directory or URI where Iceberg tables are stored.",
+    )
+    namespace: str = Field(
+        default="metaxy",
+        description="Iceberg namespace for feature tables.",
+    )
+    catalog_name: str = Field(
+        default="metaxy",
+        description="Name of the Iceberg catalog.",
+    )
+    catalog_properties: dict[str, str] | None = Field(
+        default=None,
+        description="Properties passed to pyiceberg.catalog.load_catalog.",
+    )
+    auto_create_namespace: bool = Field(
+        default=True,
+        description="Automatically create the namespace on first write if it does not exist.",
+    )
+
+
+@public
+class IcebergMetadataStore(MetadataStore):
+    """Apache Iceberg metadata store backed by [PyIceberg](https://py.iceberg.apache.org/).
+
+    Stores feature metadata in Iceberg tables managed by a PyIceberg catalog.
+    It uses the Polars versioning engine for provenance calculations.
+
+    !!! tip
+        If Polars 1.39 or greater is installed, lazy Polars frames are sinked via
+        `LazyFrame.sink_iceberg`, avoiding unnecessary materialization.
+
+    Example:
+
+        ```py
+        from metaxy.ext.metadata_stores.iceberg import IcebergMetadataStore
+
+        store = IcebergMetadataStore(
+            warehouse="s3://my-bucket/warehouse",
+            namespace="ml_features",
[…]
