### 2026-03-26 #10141 @abhishektiwari
👋👋 @simonprovost @FBruzzesi @skrawcz, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)?  The submission under consideration is "Metaxy: Record-Level Feature Metadata Management for GPU-Accelerated ML Pipeline".

Thanks
Abhi
