### 2026-03-26 60082af
test: split shared test suite for MetadataStore implementations to more granular modules (#1091)
## Summary

Resolve #644
