+++ tests/metadata_stores/shared/write.py
@@ -1,0 +1,594 @@
+"""Write test pack for metadata stores."""
+
+import random
+import time
+import warnings
+from datetime import datetime, timedelta, timezone
+
+import narwhals as nw
+import polars as pl
+import pytest
+from metaxy_testing.models import SampleFeatureSpec
+
+from metaxy import BaseFeature, FeatureDep, FeatureGraph, FeatureKey, FeatureSpec, FieldDep, FieldKey, FieldSpec
+from metaxy._utils import collect_to_polars
+from metaxy.metadata_store import MetadataStore
+from metaxy.metadata_store.warnings import MetaxyColumnMissingWarning
+from metaxy.models.constants import (
+    METAXY_CREATED_AT,
+    METAXY_DATA_VERSION,
+    METAXY_DATA_VERSION_BY_FIELD,
+    METAXY_DELETED_AT,
+    METAXY_FEATURE_VERSION,
+    METAXY_MATERIALIZATION_ID,
+    METAXY_PROJECT_VERSION,
+    METAXY_PROVENANCE,
+    METAXY_PROVENANCE_BY_FIELD,
+    METAXY_UPDATED_AT,
+)
+
+
+class WriteTests:
+    """Tests for write overrides, materialization_id, and proper dataframe writes."""
+
+    # --- Fixtures for proper dataframe write tests ---
+
+    @pytest.fixture
+    def features(self, graph: FeatureGraph):
+        """Define test features: a root feature and a downstream feature."""
+
+        class RootFeature(
+            BaseFeature,
+            spec=SampleFeatureSpec(
+                key=FeatureKey(["root", "feature"]),
+                fields=[
+                    FieldSpec(key=FieldKey(["field_a"]), code_version="1"),
+                    FieldSpec(key=FieldKey(["field_b"]), code_version="1"),
+                ],
+            ),
+        ):
+            """A root feature with no dependencies."""
+
+            pass
+
+        class DownstreamFeature(
+            BaseFeature,
+            spec=SampleFeatureSpec(
+                key=FeatureKey(["downstream", "feature"]),
+                deps=[
+                    FeatureDep(feature=FeatureKey(["root", "feature"])),
+                ],
+                fields=[
+                    FieldSpec(
+                        key=FieldKey(["output"]),
+                        code_version="1",
+                        deps=[
+                            FieldDep(
+                                feature=FeatureKey(["root", "feature"]),
+                                fields=[FieldKey(["field_a"]), FieldKey(["field_b"])],
+                            )
+                        ],
+                    ),
+                ],
+            ),
+        ):
+            """A downstream feature that depends on the root feature."""
+
+            pass
+
+        return {
+            "RootFeature": RootFeature,
+            "DownstreamFeature": DownstreamFeature,
+        }
+
+    @pytest.fixture
+    def RootFeature(self, features: dict[str, type[BaseFeature]]):
+        return features["RootFeature"]
+
+    @pytest.fixture
+    def DownstreamFeature(self, features: dict[str, type[BaseFeature]]):
+        return features["DownstreamFeature"]
+
+    # --- Proper dataframe write tests ---
+
+    def test_root_feature_with_provenance_by_field_only(self, store: MetadataStore, RootFeature):
+        """Root features should only require metaxy_provenance_by_field.
+
+        When writing a root feature (no dependencies) with only
+        metaxy_provenance_by_field, no warnings should be emitted since
+        the provenance can be computed from provenance_by_field.
+        """
+        df = pl.DataFrame(
+            {
+                "sample_uid": [1, 2, 3],
+                METAXY_PROVENANCE_BY_FIELD: [
+                    {"field_a": "hash_a1", "field_b": "hash_b1"},
[…]
