+++ tests/metadata_stores/shared/resolve_update.py
@@ -17,6 +17,7 @@ from __future__ import annotations
 
 import warnings
 from collections.abc import Mapping
+from pathlib import Path
 from typing import Any
 
 import narwhals as nw
@@ -50,7 +51,7 @@ from metaxy.models.constants import (
 from metaxy.models.field import FieldSpec
 from metaxy.models.plan import FeaturePlan
 from metaxy.models.types import FieldKey
-from metaxy.versioning.types import Increment, LazyIncrement
+from metaxy.versioning.types import HashAlgorithm, Increment, LazyIncrement
 
 # Type alias for feature plan output
 FeaturePlanOutput = tuple[FeatureGraph, Mapping[FeatureKey, type[BaseFeature]], FeaturePlan]
@@ -363,7 +364,7 @@ class OptionalDependencyCases:
 def _compute_golden_increment_for_optional_deps(
     child_plan: FeaturePlan,
     upstream_data: Mapping[str, pl.DataFrame],
-    hash_algorithm,
+    hash_algorithm: HashAlgorithm,
 ) -> Increment:
     """Compute golden increment using PolarsVersioningEngine for optional dependency tests."""
     from metaxy.versioning.polars import PolarsVersioningEngine
@@ -406,7 +407,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         root_feature: type[BaseFeature],
-    ):
+    ) -> None:
         """Test that resolve_update raises ValueError for root features without samples."""
         with store:
             with pytest.raises(ValueError, match="root feature"):
@@ -417,7 +418,7 @@ class ResolveUpdateTests:
         store: MetadataStore,
         root_feature: type[BaseFeature],
         graph: FeatureGraph,
-    ):
+    ) -> None:
         """Test resolve_update for root features with provided samples."""
         # Generate sample data using the parametric strategy
         feature_spec = root_feature.spec()
@@ -466,7 +467,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         feature_plan_config: FeaturePlanOutput,
-    ):
+    ) -> None:
         """Test resolve_update for downstream features with upstream dependencies."""
         graph, upstream_features, child_plan = feature_plan_config
 
@@ -538,7 +539,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         feature_plan_config: FeaturePlanOutput,
-    ):
+    ) -> None:
         """Test that resolve_update correctly detects added/changed/removed samples."""
         graph, upstream_features, child_plan = feature_plan_config
 
@@ -616,7 +617,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         graph: FeatureGraph,
-    ):
+    ) -> None:
         """Test resolve_update with lazy=True returns lazy frames with correct implementation."""
 
         # Create a feature graph with multiple parents (realistic scenario)
@@ -734,7 +735,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         feature_plan_config: FeaturePlanOutput,
-    ):
+    ) -> None:
         """Test that calling resolve_update multiple times is idempotent."""
         graph, upstream_features, child_plan = feature_plan_config
 
@@ -799,7 +800,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         graph: FeatureGraph,
-    ):
+    ) -> None:
         """Test that resolve_update accepts feature classes as filter keys."""
 
         class UpstreamFeature(
@@ -852,7 +853,7 @@ class ResolveUpdateTests:
         self,
         store: MetadataStore,
         graph: FeatureGraph,
-    ):
+    ) -> None:
[…]
