### 2026-03-26 d1a4618
ci: generate integration matrix with nox
