+++ tests/ext/sqlalchemy/test_native.py
@@ -6,38 +6,26 @@ This module tests the SQLAlchemy integration functionality including:
 3. MetadataStore sqlalchemy_url retrieval
 """
 
-import pytest
-from sqlmodel import Field
+from unittest.mock import MagicMock
 
-from metaxy import FeatureKey, FeatureSpec, FieldKey, FieldSpec
+import pytest
+from sqlalchemy import Column, MetaData, String, Table
+
+from metaxy import BaseFeature, FeatureKey, FeatureSpec, FieldKey, FieldSpec
 from metaxy.config import MetaxyConfig
 from metaxy.ext.sqlalchemy import (
     filter_feature_sqla_metadata,
     get_system_slqa_metadata,
 )
-from metaxy.ext.sqlmodel import BaseSQLModelFeature
+from metaxy.ext.sqlalchemy.plugin import _get_store_sqlalchemy_url
 from metaxy.metadata_store.ibis import IbisMetadataStore
 
-# Test fixtures and helper classes
-
-
-@pytest.fixture(scope="function", autouse=False)
-def _clear_sqlmodel_metadata():
-    """Clear SQLModel metadata before tests that define Features."""
-    from sqlmodel import SQLModel
-
-    # Clear all tables from the metadata before the test
-    SQLModel.metadata.clear()
-
-    yield
-
-    # Clear again after the test
-    SQLModel.metadata.clear()
+# Test fixtures
 
 
 @pytest.fixture
-def config_project_a(_clear_sqlmodel_metadata):
-    """Config for project A with both project_a and project_b features defined."""
+def config_with_features():
+    """Config with two features from different projects, using plain SQLAlchemy MetaData."""
     config = MetaxyConfig(
         project="project_a",
         store="test_store",
@@ -49,125 +37,47 @@ def config_project_a(_clear_sqlmodel_metadata):
         },
     )
     with config.use():
-        # Define ProjectAFeature within the config context
-        # Set __metaxy_project__ as class attribute to ensure it's captured at registration time
-        class ProjectAFeature(
-            BaseSQLModelFeature,
-            table=True,
+
+        class FeatureA(
+            BaseFeature,
             spec=FeatureSpec(
                 key=FeatureKey(["project_a", "feature"]),
                 id_columns=["sample_uid"],
                 fields=[FieldSpec(key=FieldKey(["value"]), code_version="1")],
             ),
         ):
-            """Feature for project A."""
-
             __metaxy_project__ = "project_a"
-            sample_uid: str = Field(primary_key=True)
+            sample_uid: str
             value: str
 
-        # Also define ProjectBFeature (for explicit_project test)
-        # Set __metaxy_project__ as class attribute to ensure it's captured at registration time
-        class ProjectBFeature(
-            BaseSQLModelFeature,
-            table=True,
+        class FeatureB(
+            BaseFeature,
             spec=FeatureSpec(
                 key=FeatureKey(["project_b", "feature"]),
                 id_columns=["sample_uid"],
                 fields=[FieldSpec(key=FieldKey(["value"]), code_version="1")],
             ),
         ):
-            """Feature for project B."""
-
             __metaxy_project__ = "project_b"
-            sample_uid: str = Field(primary_key=True)
+            sample_uid: str
             value: str
 
         yield config
 
 
 @pytest.fixture
-def config_project_b(_clear_sqlmodel_metadata):
-    """Config for project B."""
-    config = MetaxyConfig(
-        project="project_b",
-        store="test_store",
-        stores={  # ty: ignore[invalid-argument-type]
-            "test_store": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
-                "config": {"database": ":memory:"},
-            }
-        },
-    )
[…]
