+++ .github/workflows/_qa.yml
@@ -247,7 +247,7 @@ jobs:
 
   test-integrations:
     needs: [filter, generate-nox-matrix]
-    if: ${{ inputs.run-all-checks || needs.filter.outputs.test == 'true' }}
+    if: ${{ (inputs.run-all-checks || needs.filter.outputs.test == 'true') && needs.generate-nox-matrix.outputs.integrations != '[]' }}
     permissions:
       contents: read
     runs-on: depot-ubuntu-latest
@@ -261,46 +261,29 @@ jobs:
         python-version: ${{ fromJSON(inputs.python-versions) }}
         integration: ${{ fromJSON(needs.generate-nox-matrix.outputs.integrations) }}
     steps:
-      - name: Check if integration should run
-        id: should-run
-        env:
-          INTEGRATION_CHANGED: ${{ needs.filter.outputs[format('{0}', matrix.integration)] }}
-          CORE_CHANGED: ${{ needs.filter.outputs.core }}
-          RUN_ALL: ${{ inputs.run-all-checks }}
-        run: |
-          if [[ "$RUN_ALL" == "true" || "$CORE_CHANGED" == "true" || "$INTEGRATION_CHANGED" == "true" ]]; then
-            echo "run=true" >> "$GITHUB_OUTPUT"
-          else
-            echo "run=false" >> "$GITHUB_OUTPUT"
-          fi
-
       - name: Checkout code
-        if: steps.should-run.outputs.run == 'true'
         uses: actions/checkout@v6.0.2
       - name: Setup devenv
-        if: steps.should-run.outputs.run == 'true'
         uses: ./.github/actions/setup-devenv
         with:
-          extras: "duckdb ${{ matrix.integration }}"
+          install-deps: "minimal"
 
       - name: Run Tests with Coverage
-        if: steps.should-run.outputs.run == 'true'
         env:
           COVERAGE_PROCESS_START: ${{ github.workspace }}/pyproject.toml
           COVERAGE_FILE: ${{ github.workspace }}/.coverage
-        run: |
-          SITE_PACKAGES=$(python -c "import sysconfig; print(sysconfig.get_path('purelib'))")
-          cp .github/coverage_subprocess.pth "$SITE_PACKAGES/" || echo "Warning: Could not install coverage subprocess support"
-          uv run pytest tests/ext/${{ matrix.integration }}/ --cov=src/metaxy --cov-branch --durations=10 --junit-xml=./junit.xml -o junit_family=legacy
+        run: >
+          uv run nox -s "integration(integration='${{ matrix.integration }}')"
+          -- --cov=src/metaxy --cov-branch --junit-xml=./junit.xml -o junit_family=legacy
 
       - name: Combine Coverage Data
-        if: ${{ always() && steps.should-run.outputs.run == 'true' }}
+        if: always()
         run: |
           uv run coverage combine --keep || true
           uv run coverage xml -o coverage.xml
 
       - name: Upload Coverage to Codecov
-        if: ${{ !cancelled() && steps.should-run.outputs.run == 'true' }}
+        if: ${{ !cancelled() }}
         uses: codecov/codecov-action@v5.5.2
         with:
           slug: anam-org/metaxy
@@ -308,7 +291,7 @@ jobs:
           flags: ${{ matrix.integration }}-Python${{ matrix.python-version }}
 
       - name: Upload Test Results to Codecov
-        if: ${{ !cancelled() && steps.should-run.outputs.run == 'true' }}
+        if: ${{ !cancelled() }}
         uses: codecov/codecov-action@v5.5.2
         with:
[…]
