### 2026-03-26 6c14a31
ci: use per-integration devenv profiles
