+++ .github/workflows/_qa.yml
@@ -210,6 +210,7 @@ jobs:
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
         with:
+          profile: ci-duckdb
           extras: "duckdb delta"
 
       - name: Run Tests with Coverage
@@ -266,6 +267,7 @@ jobs:
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
         with:
+          profile: ci-${{ matrix.integration }}
           install-deps: "minimal"
 
       - name: Run Tests with Coverage
@@ -348,6 +350,7 @@ jobs:
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
         with:
+          profile: ci-duckdb
           install-deps: "minimal"
       - name: Run ${{ matrix.session }}
         run: uv run nox -s "${{ matrix.session }}"
@@ -450,6 +453,8 @@ jobs:
         uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
+        with:
+          profile: ci-docs
       - name: Setup Bun
         uses: oven-sh/setup-bun@v2.1.3
         with:
