### 2026-03-26 aa17a36
refactor: move metadata stores to `ext/<integration>/`
