+++ src/metaxy/ext/iceberg/legacy_store.py
@@ -1,0 +1,301 @@
+"""Apache Iceberg metadata store implemented with PyIceberg."""
+
+from __future__ import annotations
+
+from collections.abc import Callable, Iterator, Sequence
+from contextlib import contextmanager
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, overload
+
+if TYPE_CHECKING:
+    from pyiceberg.catalog import Catalog
+    from pyiceberg.table import Table
+
+import narwhals as nw
+import polars as pl
+from narwhals.typing import Frame
+from packaging.version import Version
+
+from metaxy._decorators import public
+from metaxy._utils import collect_to_polars
+from metaxy.ext.iceberg.config import IcebergMetadataStoreConfig, TableIdentifier
+from metaxy.metadata_store.base import MetadataStore
+from metaxy.metadata_store.types import AccessMode
+from metaxy.metadata_store.utils import is_local_path
+from metaxy.models.plan import FeaturePlan
+from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.versioning.polars import PolarsVersioningEngine
+from metaxy.versioning.types import HashAlgorithm
+
+
+def _strip_casts() -> Callable[[Any], Any]:
+    """Unwrap ``CAST(x AS type)`` → ``x`` to match PyIceberg's row filter syntax."""
+    from sqlglot import exp
+
+    def _transform(node: exp.Expression) -> exp.Expression:
+        if isinstance(node, exp.Cast):
+            return node.this
+        return node
+
+    return _transform
+
+
+@public
+class IcebergMetadataStore(MetadataStore):
+    """Apache Iceberg metadata store backed by [PyIceberg](https://py.iceberg.apache.org/).
+
+    Stores feature metadata in Iceberg tables managed by a PyIceberg catalog.
+    It uses the Polars versioning engine for provenance calculations.
+
+    !!! tip
+        If Polars 1.39 or greater is installed, lazy Polars frames are sinked via
+        `LazyFrame.sink_iceberg`, avoiding unnecessary materialization.
+
+    Example:
+
+        ```py
+        from metaxy.ext.metadata_stores.iceberg import IcebergMetadataStore
+
+        store = IcebergMetadataStore(
+            warehouse="s3://my-bucket/warehouse",
+            namespace="ml_features",
+            catalog_properties={"type": "glue"},
+        )
+        ```
+    """
+
+    _should_warn_auto_create_tables = False
+    versioning_engine_cls = PolarsVersioningEngine
+
+    namespace: str
+    catalog_name: str
+    auto_create_namespace: bool
+    _catalog: Catalog | None
+    _is_remote: bool
+    _warehouse_uri: str
+    _catalog_properties: dict[str, str]
+
+    def __init__(
+        self,
+        warehouse: str | Path | None = None,
+        *,
+        namespace: str = "metaxy",  # noqa: ARG002
+        catalog_name: str = "metaxy",  # noqa: ARG002
+        catalog_properties: dict[str, str] | None = None,  # noqa: ARG002
+        auto_create_namespace: bool = True,  # noqa: ARG002
+        fallback_stores: list[MetadataStore] | None = None,  # noqa: ARG002
+        **kwargs: Any,  # noqa: ARG002
+    ) -> None:
+        pass  # __new__ already initialized via MetadataStore.__init__
+
+    def __new__(
+        cls,
+        warehouse: str | Path | None = None,
+        *,
+        namespace: str = "metaxy",
+        catalog_name: str = "metaxy",
+        catalog_properties: dict[str, str] | None = None,
+        auto_create_namespace: bool = True,
+        fallback_stores: list[MetadataStore] | None = None,
+        **kwargs: Any,
+    ) -> IcebergMetadataStore:
+        if warehouse is None:
[…]
