+++ noxfile.py
@@ -1,3 +1,7 @@
+import os
+import shutil
+from pathlib import Path
+
 import nox
 import nox_uv
 
@@ -21,6 +25,22 @@ INTEGRATIONS = [
 
 IBIS_VERSIONS = ["11", "12"]
 
+COVERAGE_SUBPROCESS_PTH = Path(".github/coverage_subprocess.pth")
+
+
+def _install_coverage_subprocess(session: nox.Session) -> None:
+    """Install coverage subprocess support in the nox venv when running under CI coverage."""
+    if not os.environ.get("COVERAGE_PROCESS_START") or not COVERAGE_SUBPROCESS_PTH.exists():
+        return
+    site_packages = session.run(
+        "python",
+        "-c",
+        "import sysconfig; print(sysconfig.get_path('purelib'))",
+        silent=True,
+    )
+    assert site_packages is not None
+    shutil.copy(COVERAGE_SUBPROCESS_PTH, site_packages.strip())
+
 
 @nox_uv.session(
     python="3.10",
@@ -32,8 +52,12 @@ IBIS_VERSIONS = ["11", "12"]
 @nox.parametrize("integration", INTEGRATIONS)
 def integration(session: nox.Session, integration: str) -> None:
     """Run tests for a single integration."""
-    session.install(f".[{integration}]")
+    session.install("-e", f".[{integration}]")
+    _install_coverage_subprocess(session)
     session.run("pytest", f"tests/ext/{integration}/", "--durations=10", *session.posargs)
+    if any(arg.startswith("--cov") for arg in session.posargs):
+        session.run("coverage", "combine", "--keep", success_codes=[0, 1])
+        session.run("coverage", "xml", "-o", "coverage.xml")
 
 
 @nox_uv.session(
