### 2026-03-27 #10141 @FBruzzesi
Hey everyone, first and foremost thanks @abhishektiwari for tagging me and thanks @geoHeil for the mention!
At the current time I am also with no bandwidth to review the paper properly. I might be interested in a less hectic period - probably around the start of summer.

Best of luck,
Francesco

---

<details><summary>Details</summary>

1. I have never performed a paper review, I would need some initial time to figure out everything that needs to be done and how
2. I just started a new job, and currently relocating across countries and looking for new apartment.
