### 2026-03-27 d928939
feat: implement metaxy metadata read command (#706)
