### 2026-03-27 9f47149
style: fix linting and formatting issues
