+++ tests/cli/test_cli_metadata_read.py
@@ -1,13 +1,15 @@
 """Tests for the metadata read CLI command."""
 
 import json
-import pytest
-import polars as pl
 
+import polars as pl
+import pytest
 from metaxy_testing import TempMetaxyProject
+
 
 def _define_features():
     from metaxy_testing.models import SampleFeatureSpec
+
     from metaxy import BaseFeature, FeatureKey, FieldKey, FieldSpec
 
     class FilesRoot(
@@ -24,19 +26,24 @@ def test_metadata_read_basic(metaxy_project: TempMetaxyProject, capsys: pytest.C
     """Test basic metadata read without options."""
     with metaxy_project.with_features(_define_features):
         from metaxy.models.types import FeatureKey
+
         graph = metaxy_project.graph
         store = metaxy_project.stores["dev"]
 
-        upstream_data = pl.DataFrame({
-            "sample_uid": [1, 2, 3],
-            "value": ["val_1", "val_2", "val_3"],
-            "category": ["A", "B", "A"],
-            "metaxy_provenance_by_field": [{"default": f"hash{i}"} for i in range(1, 4)],
-        })
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1, 2, 3],
+                "value": ["val_1", "val_2", "val_3"],
+                "category": ["A", "B", "A"],
+                "metaxy_provenance_by_field": [{"default": f"hash{i}"} for i in range(1, 4)],
+            }
+        )
         with graph.use(), store.open("w"):
             store.write(FeatureKey(["files_root"]), upstream_data)
-        
-        result = metaxy_project.run_cli(["metadata", "read", "files_root", "-f", "markdown"], capsys=capsys, check=False)
+
+        result = metaxy_project.run_cli(
+            ["metadata", "read", "files_root", "-f", "markdown"], capsys=capsys, check=False
+        )
         print("STDOUT:", result.stdout)
         print("STDERR:", result.stderr)
         assert result.returncode == 0
@@ -49,17 +56,20 @@ def test_metadata_read_json(metaxy_project: TempMetaxyProject, capsys: pytest.Ca
     """Test metadata read with JSON output."""
     with metaxy_project.with_features(_define_features):
         from metaxy.models.types import FeatureKey
+
         graph = metaxy_project.graph
         store = metaxy_project.stores["dev"]
-        upstream_data = pl.DataFrame({
-            "sample_uid": [1, 2, 3],
-            "value": ["val_1", "val_2", "val_3"],
-            "category": ["A", "B", "A"],
-            "metaxy_provenance_by_field": [{"default": f"hash{i}"} for i in range(1, 4)],
-        })
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1, 2, 3],
+                "value": ["val_1", "val_2", "val_3"],
+                "category": ["A", "B", "A"],
+                "metaxy_provenance_by_field": [{"default": f"hash{i}"} for i in range(1, 4)],
+            }
+        )
         with graph.use(), store.open("w"):
             store.write(FeatureKey(["files_root"]), upstream_data)
-        
+
         result = metaxy_project.run_cli(["metadata", "read", "files_root", "-f", "json"], capsys=capsys)
         assert result.returncode == 0
         data = json.loads(result.stdout)
@@ -71,26 +81,28 @@ def test_metadata_read_select_and_filter(metaxy_project: TempMetaxyProject, caps
     """Test metadata read with --select and --filter."""
     with metaxy_project.with_features(_define_features):
         from metaxy.models.types import FeatureKey
+
         graph = metaxy_project.graph
         store = metaxy_project.stores["dev"]
-        upstream_data = pl.DataFrame({
[…]
