### 2026-03-27 af9c00b
test: expand metadata read tests for better coverage
