+++ tests/cli/test_cli_metadata_read.py
@@ -21,6 +21,33 @@ def _define_features():
     ):
         pass
 
+    class FilesParquet(
+        BaseFeature,
+        spec=SampleFeatureSpec(
+            key=FeatureKey(["files_parquet"]),
+            fields=[FieldSpec(key=FieldKey(["default"]), code_version="1")],
+        ),
+    ):
+        pass
+
+    class FilesMarkdown(
+        BaseFeature,
+        spec=SampleFeatureSpec(
+            key=FeatureKey(["files_markdown"]),
+            fields=[FieldSpec(key=FieldKey(["default"]), code_version="1")],
+        ),
+    ):
+        pass
+
+    class FilesJSON(
+        BaseFeature,
+        spec=SampleFeatureSpec(
+            key=FeatureKey(["files_json"]),
+            fields=[FieldSpec(key=FieldKey(["default"]), code_version="1")],
+        ),
+    ):
+        pass
+
 
 def test_metadata_read_basic(metaxy_project: TempMetaxyProject, capsys: pytest.CaptureFixture[str]):
     """Test basic metadata read without options."""
@@ -197,3 +224,180 @@ def test_metadata_read_output_file(metaxy_project: TempMetaxyProject, tmp_path, 
         content = output_file.read_text()
         assert "sample_uid,value,category" in content
         assert "1,val_1,A" in content
+
+
+def test_metadata_read_parquet_file(metaxy_project: TempMetaxyProject, tmp_path, capsys: pytest.CaptureFixture[str]):
+    """Test output to a parquet file."""
+    with metaxy_project.with_features(_define_features):
+        from metaxy.models.types import FeatureKey
+
+        graph = metaxy_project.graph
+        store = metaxy_project.stores["dev"]
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1],
+                "value": ["val_1"],
+                "category": ["A"],
+                "metaxy_provenance_by_field": [{"default": "hash"}],
+            }
+        )
+        with graph.use(), store.open("w"):
+            store.write(FeatureKey(["files_parquet"]), upstream_data)
+
+        output_file = tmp_path / "output.parquet"
+        result = metaxy_project.run_cli(
+            ["metadata", "read", "files_parquet", "-f", "parquet", "-o", str(output_file)], capsys=capsys
+        )
+
+        assert result.returncode == 0
+        assert output_file.exists()
+        # Verify it's a valid parquet
+        df = pl.read_parquet(output_file)
+        assert len(df) == 1
+        assert "sample_uid" in df.columns
+
+
+def test_metadata_read_markdown_file(metaxy_project: TempMetaxyProject, tmp_path, capsys: pytest.CaptureFixture[str]):
+    """Test markdown output to a file."""
+    with metaxy_project.with_features(_define_features):
+        from metaxy.models.types import FeatureKey
+
+        graph = metaxy_project.graph
+        store = metaxy_project.stores["dev"]
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1],
+                "value": ["val_1"],
+                "category": ["A"],
+                "metaxy_provenance_by_field": [{"default": "hash"}],
+            }
+        )
+        with graph.use(), store.open("w"):
+            store.write(FeatureKey(["files_markdown"]), upstream_data)
+
+        output_file = tmp_path / "output.md"
+        result = metaxy_project.run_cli(
+            ["metadata", "read", "files_markdown", "-f", "markdown", "-o", str(output_file)], capsys=capsys
+        )
+
+        assert result.returncode == 0
+        assert output_file.exists()
+        content = output_file.read_text()
+        assert "| sample_uid |" in content
+
+
+def test_metadata_read_json_file(metaxy_project: TempMetaxyProject, tmp_path, capsys: pytest.CaptureFixture[str]):
[…]
