### 2026-03-27 19834b7
address PR review: dynamic table name, collect_to_polars, IbisMetadataStore opt, output refactor, assert_frame_equal
