### 2026-03-27 03cfba2
test(cli): fix coverage mock setup to pass rigorous metaxy schema tests
