+++ tests/cli/test_cli_metadata_read.py
@@ -406,3 +406,184 @@ def test_metadata_read_explicit_store(metaxy_project: TempMetaxyProject, capsys:
         assert result.returncode == 0
         data = json.loads(result.stdout)
         assert data[0]["sample_uid"] == 1
+
+
+def test_metadata_read_invalid_format(metaxy_project: TempMetaxyProject, capsys: pytest.CaptureFixture[str]):
+    """Test reading with an unsupported format."""
+    with metaxy_project.with_features(_define_features):
+        from metaxy.models.types import FeatureKey
+
+        graph = metaxy_project.graph
+        store = metaxy_project.stores["dev"]
+        # Must write data first, otherwise it fails on "feature not found"
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1],
+                "value": ["val_1"],
+                "category": ["A"],
+                "metaxy_provenance_by_field": [{"default": "hash"}],
+            }
+        )
+        with graph.use(), store.open("w"):
+            store.write(FeatureKey(["files_root"]), upstream_data)
+
+        result = metaxy_project.run_cli(["metadata", "read", "files_root", "-f", "invalid"], capsys=capsys, check=False)
+        assert result.returncode == 1
+        assert "Unsupported format: invalid" in (result.stderr + result.stdout)
+
+
+def test_metadata_read_csv_stdout(metaxy_project: TempMetaxyProject, capsys: pytest.CaptureFixture[str]):
+    """Test reading and outputting CSV to stdout."""
+    with metaxy_project.with_features(_define_features):
+        from metaxy.models.types import FeatureKey
+
+        graph = metaxy_project.graph
+        store = metaxy_project.stores["dev"]
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1],
+                "default": ["val_1"],
+                "metaxy_provenance_by_field": [{"default": "hash"}],
+            }
+        )
+        with graph.use(), store.open("w"):
+            store.write(FeatureKey(["files_root"]), upstream_data)
+
+        # Use --select to keep output clean and predictable
+        result = metaxy_project.run_cli(
+            ["metadata", "read", "files_root", "-f", "csv", "--select", "sample_uid", "--select", "default"],
+            capsys=capsys,
+        )
+        assert result.returncode == 0
+        # Check that CSV header and data are in stdout (handle Windows line endings)
+        clean_out = result.stdout.replace("\r", "")
+        assert "sample_uid,default" in clean_out
+        assert "1,val_1" in clean_out
+
+
+def test_metadata_read_markdown_stdout(metaxy_project: TempMetaxyProject, capsys: pytest.CaptureFixture[str]):
+    """Test reading and outputting Markdown to stdout."""
+    with metaxy_project.with_features(_define_features):
+        from metaxy.models.types import FeatureKey
+
+        graph = metaxy_project.graph
+        store = metaxy_project.stores["dev"]
+        upstream_data = pl.DataFrame(
+            {
+                "sample_uid": [1],
+                "default": ["val_1"],
+                "metaxy_provenance_by_field": [{"default": "hash"}],
+            }
+        )
+        with graph.use(), store.open("w"):
+            store.write(FeatureKey(["files_root"]), upstream_data)
+
+        # Use --select to limit columns and avoid system columns cluttering the check
+        result = metaxy_project.run_cli(
+            ["metadata", "read", "files_root", "-f", "markdown", "--select", "sample_uid", "--select", "default"],
+            capsys=capsys,
+        )
+        assert result.returncode == 0
+        assert "sample_uid" in result.stdout
+        assert "val_1" in result.stdout
[…]
