### 2026-03-27 6089e56
test(cli): bypass strict pyre typing for MagicMock to resolve CI typecheck
