### 2026-03-27 ad802ce
style(cli): auto-format metadata.py with ruff to fix CI qa/format check
