### 2026-03-27 66ee494
feat: `Map` type support via `polars-map` (experimental) (#1104)
## Summary

<!-- Brief description of the changes -->

Just recently I discovered the problems we had with `Map` columns were because of Polars. All the other tools supported by Metaxy support the `Map` type, while [Polars doesn't](https://github.com/pola-rs/polars/issues/8385), and it converts it to `List(Struct(key, value))` on reads.

This PR is attempting to solve this via the [`polars-map`](https://pypi.org/project/polars-map/) plugin.

The new functionality is guarded by `enable_map_datatype` global Metaxy config parameter.
[…]
