+++ tests/versioning/test_arrow_map.py
@@ -1,0 +1,170 @@
+"""Unit tests for Arrow Map conversion utilities."""
+
+from __future__ import annotations
+
+import polars as pl
+import pyarrow as pa
+import pytest
+from polars_map import Map
+
+from metaxy.versioning._arrow_map import (
+    convert_extension_maps_to_native,
+    convert_maps_to_polars_map,
+    convert_structs_to_maps,
+)
+
+MAP_STR_STR = Map(pl.String(), pl.String())
+
+
+# ── Fixtures ──────────────────────────────────────────────────────────
+
+
+@pytest.fixture
+def struct_df() -> pl.DataFrame:
+    """Polars DataFrame with a Struct column."""
+    return pl.DataFrame(
+        {
+            "sample_uid": ["s1", "s2"],
+            "metaxy_provenance_by_field": [
+                {"alpha": "a1", "beta": "b1"},
+                {"alpha": "a2", "beta": "b2"},
+            ],
+        }
+    )
+
+
+@pytest.fixture
+def multi_column_df() -> pl.DataFrame:
+    """Polars DataFrame with metaxy struct columns and non-metaxy columns."""
+    return pl.DataFrame(
+        {
+            "sample_uid": ["s1"],
+            "metaxy_data_version_by_field": [{"x": "1", "y": "2"}],
+            "other_struct": [{"x": "3", "y": "4"}],
+            "score": [0.9],
+        }
+    )
+
+
+@pytest.fixture
+def list_kv_df() -> pl.DataFrame:
+    """Polars DataFrame with a List(Struct({key, value})) column (as Arrow Map appears in Polars)."""
+    return pl.DataFrame(
+        {
+            "sample_uid": ["s1", "s2"],
+            "metaxy_provenance_by_field": [
+                [{"key": "alpha", "value": "a1"}, {"key": "beta", "value": "b1"}],
+                [{"key": "alpha", "value": "a2"}, {"key": "beta", "value": "b2"}],
+            ],
+        }
+    )
+
+
+# ── convert_structs_to_maps ──────────────────────────────────────────
+
+
+class TestConvertStructsToMaps:
+    """Tests for the write-path Struct to polars_map.Map conversion."""
+
+    def test_struct_to_map_basic(self, struct_df: pl.DataFrame) -> None:
+        """Struct column becomes polars_map.Map, values accessible via .map.get()."""
+        result = convert_structs_to_maps(struct_df, columns=["metaxy_provenance_by_field"])
+
+        assert result.schema["metaxy_provenance_by_field"] == MAP_STR_STR
+        assert result["metaxy_provenance_by_field"].map.get("alpha").to_list() == ["a1", "a2"]  # ty: ignore[unresolved-attribute]
+        assert result["metaxy_provenance_by_field"].map.get("beta").to_list() == ["b1", "b2"]  # ty: ignore[unresolved-attribute]
+
+    def test_roundtrip_to_native_arrow_map(self, struct_df: pl.DataFrame) -> None:
+        """Struct → polars_map.Map → Arrow extension → native Arrow MapArray."""
+        converted = convert_structs_to_maps(struct_df, columns=["metaxy_provenance_by_field"])
+        arrow_table = converted.to_arrow()
+        native = convert_extension_maps_to_native(arrow_table)
+
+        map_field = native.schema.field("metaxy_provenance_by_field")
+        assert pa.types.is_map(map_field.type)
+
+        row0 = native.column("metaxy_provenance_by_field")[0].as_py()
+        assert row0 == [("alpha", "a1"), ("beta", "b1")]
+
+    def test_only_specified_columns_converted(self, multi_column_df: pl.DataFrame) -> None:
+        """Only specified columns are converted, others remain unchanged."""
[…]
