+++ src/metaxy/_utils.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import Any, TypeAlias, cast, overload
 
 import narwhals as nw
@@ -7,25 +8,75 @@ from narwhals.typing import DataFrameT, Frame, FrameT, LazyFrameT
 PolarsCompatibleFrame: TypeAlias = Frame | pl.DataFrame | pl.LazyFrame
 
 
+def _is_polars_map_dtype(dtype: pl.DataType) -> bool:
+    """Check if a Polars dtype is a polars_map.Map extension type without importing polars_map."""
+    return isinstance(dtype, pl.datatypes.classes.BaseExtension) and getattr(dtype, "_name", None) == "polars_map.map"
+
+
+def find_map_columns(df: Frame) -> list[str]:
+    """Return column names that have a Map type in the underlying frame.
+
+    Returns empty list when enable_map_datatype is not set.
+    Inspects the native backend (Polars, Arrow, Ibis) for Map-typed columns.
+    """
+    from metaxy.config import MetaxyConfig
+
+    if not MetaxyConfig.get().enable_map_datatype:
+        return []
+
+    if df.implementation == nw.Implementation.POLARS:
+        native = df.to_native()
+        schema = native.collect_schema() if isinstance(native, pl.LazyFrame) else native.schema
+        return [name for name, dtype in schema.items() if _is_polars_map_dtype(dtype)]
+
+    if df.implementation == nw.Implementation.PYARROW:
+        import pyarrow as pa
+
+        table = cast(pa.Table, df.to_native())
+        return [field.name for field in table.schema if pa.types.is_map(field.type)]
+
+    if df.implementation == nw.Implementation.IBIS:
+        import ibis
+        import ibis.expr.datatypes as dt
+
+        table = cast(ibis.Table, df.to_native())
+        return [name for name, dtype in table.schema().items() if isinstance(dtype, dt.Map)]
+
+    warnings.warn(
+        f"cannot identify Map columns for unsupported narwhals implementation {df.implementation}",
+        stacklevel=2,
+    )
+    return []
+
+
+def has_polars_map_columns(df: pl.DataFrame | pl.LazyFrame) -> bool:
+    """Check if a Polars DataFrame/LazyFrame has any polars_map.Map columns."""
+    return bool(find_map_columns(cast(Frame, nw.from_native(df))))
+
+
 def collect_to_polars(frame: PolarsCompatibleFrame) -> pl.DataFrame:
     """Helper to convert a Narwhals frame into an eager Polars DataFrame.
 
     Avoids unnecessary re-materialization when the frame is already a Polars-backed
-    eager DataFrame.
+    eager DataFrame. Preserves polars_map.Map columns when enable_map_datatype is set.
     """
-    if isinstance(frame, nw.DataFrame):
-        if frame.implementation == nw.Implementation.POLARS:
-            return cast(pl.DataFrame, frame.to_native())
-        return frame.to_polars()
-
-    if isinstance(frame, nw.LazyFrame):
-        return cast(pl.DataFrame, lazy_frame_to_polars(frame).collect())
-
     if isinstance(frame, pl.DataFrame):
         return frame
-
     if isinstance(frame, pl.LazyFrame):
         return cast(pl.DataFrame, frame.collect())
+
+    if isinstance(frame, (nw.DataFrame, nw.LazyFrame)):
+        if frame.implementation == nw.Implementation.POLARS:
+            native = frame.to_native()
+            return cast(pl.DataFrame, native.collect() if isinstance(native, pl.LazyFrame) else native)
+
+        map_cols = find_map_columns(frame)
+        result = frame.collect().to_polars() if isinstance(frame, nw.LazyFrame) else frame.to_polars()
+        if map_cols:
[…]

[…]
