### 2026-03-27 ce4d225
refactor(cli): consolidate output streams for metadata read resolving review feedback
