### 2026-03-27 09a6dd4
feat(iceberg): `Map` type support (experimental) (#1108)
See https://app.graphite.com/github/pr/anam-org/metaxy/1104
