+++ tests/ext/iceberg/test_native.py
@@ -19,6 +19,7 @@ from tests.metadata_stores.shared import (
     DeletionTests,
     DisplayTests,
     FilterTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
@@ -44,6 +45,7 @@ class TestIceberg(
     DeletionTests,
     DisplayTests,
     FilterTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
