+++ tests/ext/delta/test_native.py
@@ -17,6 +17,7 @@ from tests.metadata_stores.shared import (
     DeletionTests,
     DisplayTests,
     FilterTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
@@ -46,6 +47,7 @@ class TestDelta(
     DeletionTests,
     DisplayTests,
     FilterTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
