+++ src/metaxy/ext/metadata_stores/delta.py
@@ -17,6 +17,7 @@ from pydantic import Field
 
 from metaxy._decorators import public
 from metaxy._utils import collect_to_polars
+from metaxy.config import MetaxyConfig
 from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
 from metaxy.metadata_store.types import AccessMode
 from metaxy.metadata_store.utils import is_local_path
@@ -24,6 +25,13 @@ from metaxy.models.plan import FeaturePlan
 from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
 from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
+
+
+def _map_columns_from_delta_schema(schema: deltalake.Schema) -> list[str]:
+    """Return column names that have a native Arrow Map type in the Delta schema."""
+    from deltalake.schema import MapType
+
+    return [field.name for field in schema.fields if isinstance(field.type, MapType)]
 
 
 @public
@@ -268,6 +276,10 @@ class DeltaMetadataStore(MetadataStore):
         mode = write_opts.pop("mode", "append")
         storage_options = write_opts.pop("storage_options", None)
 
+        if MetaxyConfig.get().enable_map_datatype:
+            self._write_with_map_columns(df, table_uri, mode, storage_options, write_opts)
+            return
+
         # Check if we can use sink_delta (Polars >= 1.37, native Polars LazyFrame)
         can_sink = (
             df.implementation == nw.Implementation.POLARS
@@ -294,6 +306,41 @@ class DeltaMetadataStore(MetadataStore):
                 storage_options=storage_options,
                 delta_write_options=write_opts or None,
             )
+
+    def _write_with_map_columns(
+        self,
+        df: Frame,
+        table_uri: str,
+        mode: Literal["error", "append", "ignore", "overwrite"],
+        storage_options: dict[str, str] | None,
+        write_opts: dict[str, Any],
+    ) -> None:
+        """Collect to Arrow and convert Struct *_by_field columns to native MapArray before writing."""
+        from metaxy.models.constants import METAXY_DATA_VERSION_BY_FIELD, METAXY_PROVENANCE_BY_FIELD
+        from metaxy.versioning._arrow_map import convert_extension_maps_to_native, convert_structs_to_maps
+
+        df_native = self._cast_enum_to_string(collect_to_polars(df))
+        df_native = convert_structs_to_maps(
+            df_native, columns=[METAXY_PROVENANCE_BY_FIELD, METAXY_DATA_VERSION_BY_FIELD]
+        )
+        arrow_table = convert_extension_maps_to_native(df_native.to_arrow())
+        deltalake.write_deltalake(
+            table_uri,
+            arrow_table,
+            mode=mode,
+            storage_options=storage_options,
+            **(write_opts or {}),
+        )
+
+    def _read_map_columns(self, lf: pl.LazyFrame, feature_key: FeatureKey) -> pl.LazyFrame:
+        """Convert native Delta Map columns back to Polars Map dtype on read."""
+        delta_table = self._open_delta_table(feature_key, without_files=True)
+        map_columns = _map_columns_from_delta_schema(delta_table.schema())
+        if map_columns:
+            from metaxy.versioning._arrow_map import convert_maps_to_polars_map
+
+            return convert_maps_to_polars_map(lf, columns=map_columns)
+        return lf
 
     def _drop_feature(self, feature_key: FeatureKey) -> None:
         """Drop Delta table for the specified feature using soft delete.
@@ -377,6 +424,9 @@ class DeltaMetadataStore(MetadataStore):
             table_uri,
             storage_options=self.storage_options or None,
[…]
