### 2026-03-27 0748322
feat(duckdb): `Map` type support (experimental) (#1105)
See https://app.graphite.com/github/pr/anam-org/metaxy/1104
