### 2026-03-27 870aa05
feat(clickhouse): `Map` type support (experimental) (#1107)
See https://app.graphite.com/github/pr/anam-org/metaxy/1104
