+++ tests/ext/clickhouse/test_native.py
@@ -15,6 +15,8 @@ from tests.metadata_stores.shared import (
     DeletionTests,
     DisplayTests,
     FilterTests,
+    IbisMapTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
@@ -41,6 +43,8 @@ class TestClickHouse(
     DeletionTests,
     DisplayTests,
     FilterTests,
+    IbisMapTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
