### 2026-03-27 bc47f45
feat(lancedb): support `Map` type (experimental)
