+++ tests/ext/lancedb/test_native.py
@@ -20,6 +20,7 @@ from tests.metadata_stores.shared import (
     DeletionTests,
     DisplayTests,
     FilterTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
@@ -33,6 +34,7 @@ class TestLanceDB(
     DeletionTests,
     DisplayTests,
     FilterTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
