+++ tests/ext/duckdb/test_ducklake.py
@@ -12,6 +12,8 @@ from tests.metadata_stores.shared import (
     DeletionTests,
     DisplayTests,
     FilterTests,
+    IbisMapTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
@@ -27,6 +29,8 @@ class TestDuckDBDuckLake(
     DeletionTests,
     DisplayTests,
     FilterTests,
+    IbisMapTests,
+    MapDtypeTests,
     ResolveUpdateTests,
     VersioningTests,
     WriteTests,
