### 2026-03-27 865ee26
docs: document `Map` datatype usage and `enable_map_datatype` config option (#1115)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
