+++ docs/integrations/metadata-stores/databases/clickhouse.md
@@ -10,6 +10,10 @@ description: "Learn how to use ClickHouse as a Metaxy metadata store."
 
 1. extremely fast
 
+!!! tip "Recommended: enable [`Map` datatype](../../../guide/concepts/metadata-stores.md#map-datatype)"
+
+    ClickHouse natively supports the `Map` type. Enabling [`enable_map_datatype`](../../../reference/configuration.md#metaxy.config.MetaxyConfig.enable_map_datatype) preserves `Map` columns across read and write operations.
+
 ## Installation
 
 ```shell
@@ -38,9 +42,9 @@ ClickHouse offers multiple approaches to represent Metaxy's structured versionin
 
     - **Good performance** for key-value lookups
 
-!!! warning "Special Map columns handling"
+??? info "Legacy: Struct-to-Map conversion without `enable_map_datatype`"
 
-    Metaxy transforms its system columns (`metaxy_provenance_by_field`, `metaxy_data_version_by_field`):
+    When [`enable_map_datatype`](../../../reference/configuration.md#metaxy.config.MetaxyConfig.enable_map_datatype) is **not** enabled, Metaxy transforms its system columns (`metaxy_provenance_by_field`, `metaxy_data_version_by_field`):
 
     - **Reading**: System Map columns are converted into [Ibis Structs](https://ibis-project.org/reference/datatypes#ibis.expr.datatypes.core.Struct) (e.g., `Struct[{"field_a": str, "field_b": str}]`)
 
