+++ docs/.glossary.md
@@ -51,3 +51,5 @@
 *[fully qualified field path]: A unique identifier for a field within a feature definition, consisting of the feature name and the field name. Example: `/raw/video:frames`.
 
 *[syntactic sugar]: Simplified syntax for common Metaxy constructs. | /guide/concepts/syntactic-sugar
+
+*[field-level versioning columns]: columns used to track field-level versioning information: `metaxy_provenance_by_field` and `metaxy_data_version_by_field`. | /reference/system-columns/
