### 2026-03-28 47eb56f
docs: add scope subgroups to changelog
