+++ CHANGELOG.md
@@ -21,22 +21,115 @@ While Metaxy's core functionality and versioning engine are **_stable_** and qui
 <!-- --8<-- [end:header] -->
 <!-- --8<-- [start:releases] -->
 
+## v0.1.7 (28-03-2026)
+
+This release adds experimental support for the [`Map` datatype](https://arrow.apache.org/docs/python/generated/pyarrow.map_.html#pyarrow.map_) and a new [Apache Iceberg](https://iceberg.apache.org/) metadata store.
+
+### :sparkles: Features
+
+- A new experimental `enable_map_datatype` global setting has been added to Metaxy's configuration. When enabled, it uses [`polars-map`](https://pypi.org/project/polars-map/) (must be installed) to preserve the Arrow `Map` type on Polars frames across Metaxy operations. ([#1104](https://github.com/anam-org/metaxy/pull/1104) by [@danielgafni](https://github.com/danielgafni))
+
+#### Claude
+
+- update Claude plugin with `Map` support info ([#1117](https://github.com/anam-org/metaxy/pull/1117) by [@danielgafni](https://github.com/danielgafni))
+- update Claude plugin with External Features docs ([#1086](https://github.com/anam-org/metaxy/pull/1086) by [@danielgafni](https://github.com/danielgafni))
+
+#### Clickhouse
+
+- `Map` type support (experimental) ([#1107](https://github.com/anam-org/metaxy/pull/1107) by [@danielgafni](https://github.com/danielgafni))
+
+#### Duckdb
+
+- `Map` type support (experimental) ([#1105](https://github.com/anam-org/metaxy/pull/1105) by [@danielgafni](https://github.com/danielgafni))
+
+#### Deltalake
+
+- `Map` type support (experimental) ([#1108](https://github.com/anam-org/metaxy/pull/1108) by [@danielgafni](https://github.com/danielgafni))
+
+#### Iceberg
+
+- `Map` type support (experimental) ([#1108](https://github.com/anam-org/metaxy/pull/1108) by [@danielgafni](https://github.com/danielgafni))
+- add Apache Iceberg metadata store ([#1033](https://github.com/anam-org/metaxy/pull/1033) by [@ashutosh1807](https://github.com/ashutosh1807))
+
+
+### :bug: Bug Fixes
+
+- do not validate incomplete MetaxyConfig when `extend` is set ([#1087](https://github.com/anam-org/metaxy/pull/1087) by [@danielgafni](https://github.com/danielgafni))
+
+#### Dagster
+
+- log runtime metadata for `Map` type correctly ([#1120](https://github.com/anam-org/metaxy/pull/1120) by [@danielgafni](https://github.com/danielgafni))
+
+#### Iceberg
+
+- use appropriate warehouse path on Windows ([#1094](https://github.com/anam-org/metaxy/pull/1094) by [@danielgafni](https://github.com/danielgafni))
+
+
+### :book: Docs
+
+- add scope subgroups to changelog ([#1123](https://github.com/anam-org/metaxy/pull/1123) by [@danielgafni](https://github.com/danielgafni))
+- update `enable_map_datatype` docstring ([#1121](https://github.com/anam-org/metaxy/pull/1121) by [@danielgafni](https://github.com/danielgafni))
+- document `Map` datatype usage and `enable_map_datatype` config option ([#1115](https://github.com/anam-org/metaxy/pull/1115) by [@danielgafni](https://github.com/danielgafni))
+- fix broken metaxy.lock glossary link ([#1078](https://github.com/anam-org/metaxy/pull/1078) by [@danielgafni](https://github.com/danielgafni))
+
+#### Iceberg
+
[…]
