+++ 0.1.7/slides/2026-introducing-metaxy/dist/assets/md-Drxw-h1D.js
@@ -1,0 +1,1 @@
+import{_ as i}from"./slidev/VClicks-kWLYTJrV.js";import{b as u,o as p,w as o,g as e,e as c,D as n,v as d,x as m,z as a}from"./modules/vue-D2_2azlL.js";import{I as g}from"./slidev/default-DMaQI0-d.js";import{u as f,f as _}from"./slidev/context-CXJvKEzn.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";const B={__name:"introducing-metaxy.md__slidev_10",setup(x){const{$clicksContext:r,$frontmatter:l}=f();return r.setup(),(h,t)=>{const s=i;return p(),u(g,d(m(a(_)(a(l),9))),{default:o(()=>[t[1]||(t[1]=e("h1",null,"Running in Production",-1)),c(s,null,{default:o(()=>[...t[0]||(t[0]=[e("p",null,[e("a",{href:"https://anam.ai/blog/metaxy",target:"_blank"},"Anam"),n(": "),e("strong",null,"millions of samples"),n(" in production since Dec 2025, powering "),e("strong",null,"Cara 3"),n(" training data.")],-1),e("p",null,"The pipeline processes raw video and audio through multiple stages:",-1),e("ul",null,[e("li",null,"Face detection and cropping"),e("li",null,"Audio extraction and transcription"),e("li",null,"Embedding generation")],-1),e("p",null,[e("strong",null,"Before Metaxy:"),n(" A single code change triggered full reprocessing across all stages.")],-1),e("p",null,[e("strong",null,"After Metaxy:"),n(" Only the records and fields that actually depend on the change are reprocessed.")],-1)])]),_:1})]),_:1},16)}}};export{B as default};
