+++ 0.1.7/slides/2026-introducing-metaxy/dist/assets/chunk-WDWNAAEG-C-FBSnpL.js
@@ -1,0 +1,1 @@
+import{a,f as w,l as x}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";var d=a((e,t,i,r)=>{e.attr("class",i);const{width:o,height:h,x:n,y:c}=u(e,t);w(e,h,o,r);const s=l(n,c,o,h,t);e.attr("viewBox",s),x.debug(`viewBox configured: ${s} with padding: ${t}`)},"setupViewPortForSVG"),u=a((e,t)=>{const i=e.node()?.getBBox()||{width:0,height:0,x:0,y:0};return{width:i.width+t*2,height:i.height+t*2,x:i.x,y:i.y}},"calculateDimensionsWithPadding"),l=a((e,t,i,r,o)=>`${e-o} ${t-o} ${i} ${r}`,"createViewBox");export{d as s};
