### 2026-03-28 ceb4d80
fix: address PR #1109 review feedback from danielgafni
