+++ src/metaxy/cli/metadata.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import json
-from typing import TYPE_CHECKING, Annotated, Any
+from typing import TYPE_CHECKING, Annotated, Any, Literal
 
 import cyclopts
 import narwhals as nw
@@ -115,11 +115,6 @@ def status(
 
         # Handle empty result for --all-features
         if selector.all_features and not selector:
-            _output_no_features_warning(format, project_version)
-            return
-
-        # If no valid features remain
-        if not selector:
             _output_no_features_warning(format, project_version)
             return
 
@@ -754,14 +749,21 @@ def read(
             help="Columns to select. Can be repeated.",
         ),
     ] = None,
+    format: Annotated[
+        Literal["csv", "json", "parquet", "markdown"],
+        cyclopts.Parameter(
+            name=["--format", "-f"],
+            help="Output format: markdown, json, csv, parquet.",
+        ),
+    ] = "markdown",
     query: Annotated[
         str | None,
         cyclopts.Parameter(
             name=["--query"],
             help=(
                 "Arbitrary SQL query to execute against the metadata. "
-                "The table name matches the feature's table_name "
-                "(e.g. 'my__feature' for feature key 'my/feature')."
+                "The table is available as both its specific table name (e.g., 'files_root') "
+                "and under the generic alias 'metadata'."
             ),
         ),
     ] = None,
@@ -772,13 +774,6 @@ def read(
             help="Output file path (default: print to stdout).",
         ),
     ] = None,
-    format: Annotated[
-        str,
-        cyclopts.Parameter(
-            name=["--format", "-f"],
-            help="Output format: markdown, json, csv, parquet.",
-        ),
-    ] = "markdown",
 ) -> None:
     """Read feature metadata and optionally execute arbitrary SQL against it.
 
@@ -801,14 +796,16 @@ def read(
 
     from metaxy._utils import collect_to_polars
     from metaxy.cli.context import AppContext
-    from metaxy.cli.utils import CLIError, CLIErrorCode, exit_with_error
+    from metaxy.cli.utils import CLIError, CLIErrorCode, exit_with_error, load_graph_for_command
     from metaxy.metadata_store.ibis import IbisMetadataStore
 
     filters = filters or []
     context = AppContext.get()
     metadata_store = context.get_store(store)
 
-    selector.resolve("plain")
+    # Load graph before resolution. We use "plain" format for read command's resolution phase.
+    graph = load_graph_for_command(context, None, metadata_store, "plain")
+    selector.resolve("plain", graph=graph)
 
     if not selector:
         exit_with_error(
@@ -829,79 +826,60 @@ def read(
                 if select:
                     df = df.select(*select)
 
-                # For IbisMetadataStore, try to execute SQL in the DB before collecting
+                # For IbisMetadataStore, execute SQL in the DB by extending the lazy Ibis frame.
+                # We compile the already-filtered df expression into a CTE so the user's SQL
+                # query runs inside the database (great perf: no collect needed).
+                # Both table_name and 'metadata' are provided as aliases.
                 if query and isinstance(metadata_store, IbisMetadataStore):
-                    # Create a temporary view with the dynamic table name and run the query
-                    # This allows the backend (DuckDB, Postgres, etc.) to handle the SQL
-                    try:
-                        conn = metadata_store.conn
-                        # Many backends support conn.sql(query) directly
-                        # But we need to make sure the 'table_name' is registered or reachable
-                        # Most Ibis backends allow querying an existing expression by name if we alias it
-                        sql_query = (
-                            str(query).replace("metadata", table_name) if "metadata" in str(query) else str(query)
-                        )
[…]
