### 2026-03-28 2e577d2
style: remove verbose AI-ish comments
