### 2026-03-28 2fe643f
fix: resolve CLI regressions and linter warnings from metadata Refactor
