### 2026-03-30 d202fa9
refactor: rename and move _collect_to_arrow to `metaxy._utils.collect_to_arrow`
