### 2026-03-30 2315403
feat: decide whether to use fallback stores based on record count
