### 2026-03-30 d5e3fee
refactor: rename and move _collect_to_arrow to `metaxy._utils.collect_to_arrow` (#1130)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
