+++ pyproject.toml
@@ -156,6 +156,7 @@ dev = [
   "mkdocs-metaxy",
   "pytest-postgresql>=8.0.0",
   "polars-map>=0.1.1",
+  "tach>=0.34.0",
 ]
 
 examples = [
