### 2026-03-31 d0eac27
feat(cli): more detailed and user-friendly output for `metaxy lock` (#1135)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
