+++ src/metaxy/utils/lock_file.py
@@ -111,13 +111,63 @@ def load_lock_file(config: MetaxyConfig) -> list[FeatureDefinition]:
     return definitions
 
 
+class FeatureChange(BaseModel):
+    """A single feature that was added or removed."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    project: str
+    key: str
+    version: str
+
+
+class FeatureUpdate(BaseModel):
+    """A single feature whose definition changed."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    project: str
+    key: str
+    old_version: str
+    new_version: str
+    metadata_only: bool
+
+
+class LockFileResult(BaseModel):
+    """Result of a lock file generation.
+
+    Attributes:
+        added: Features present in the new lock file but not the old one.
+        updated: Features whose definition changed between old and new lock files.
+        removed: Features present in the old lock file but not the new one.
+        unchanged_count: Number of features that are identical between old and new lock files.
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    added: list[FeatureChange]
+    updated: list[FeatureUpdate]
+    removed: list[FeatureChange]
+    unchanged_count: int = 0
+
+    @property
+    def count(self) -> int:
+        """Total number of features in the new lock file."""
+        return len(self.added) + len(self.updated) + self.unchanged_count
+
+    @property
+    def changed(self) -> bool:
+        """Whether any features were added, updated, or removed."""
+        return bool(self.added or self.updated or self.removed)
+
+
 def generate_lock_file(
     store: MetadataStore,
     output_path: Path,
     *,
     exclude_project: str | None = None,
     selection: FeatureSelection | None = None,
-) -> int:
+) -> LockFileResult:
     """Generate a lock file with feature definitions from a metadata store.
 
     Automatically discovers external dependencies by analyzing the current FeatureGraph.
@@ -132,7 +182,7 @@ def generate_lock_file(
             beyond those discovered from the graph's dependencies.
 
     Returns:
-        Number of features locked.
+        Result with the number of features and whether the lock file changed.
 
     Raises:
         FeatureNotFoundError: If any required features are not found in the store.
@@ -158,8 +208,7 @@ def generate_lock_file(
     external_keys_needed = referenced_dep_keys - local_keys
 
     if not external_keys_needed and not selection:
-        _write_lock_file(output_path, LockFile(features={}))
-        return 0
+        return _write_lock_file(output_path, LockFile(features={}))
 
     # Use nullcontext if store is already open, otherwise open it
     cm = nullcontext(store) if store._is_open else store
@@ -173,8 +222,7 @@ def generate_lock_file(
         _apply_selection(selection, all_store_definitions, local_keys, external_keys_needed)
 
     if not external_keys_needed:
-        _write_lock_file(output_path, LockFile(features={}))
-        return 0
+        return _write_lock_file(output_path, LockFile(features={}))
 
     # Resolve transitive dependencies in memory
     definitions, missing = _resolve_transitive_deps(external_keys_needed, all_store_definitions, local_keys, graph)
@@ -189,9 +237,7 @@ def generate_lock_file(
 
     # Build and write lock file (validates versions match)
     lock_file = _build_lock_file(definitions, db_versions, graph)
-    _write_lock_file(output_path, lock_file)
-
-    return len(definitions)
+    return _write_lock_file(output_path, lock_file)
 
 
 def _load_all_features_from_store(
@@ -344,8 +390,69 @@ def _build_lock_file(
     return LockFile(features=features)
 
 
-def _write_lock_file(output_path: Path, lock_file: LockFile) -> None:
-    """Write lock file with header comment."""
-    header = "# Generated by `metaxy lock`.\n\n"
[…]

[…]
