### 2026-03-31 b685d7e
fix: improve `Map` datatype support in `MetadataStore.resolve_update`
