+++ pyproject.toml
@@ -155,7 +155,7 @@ dev = [
   "mermaid-py>=0.8.0",
   "mkdocs-metaxy",
   "pytest-postgresql>=8.0.0",
-  "polars-map>=0.1.1",
+  "polars-map>=0.2.0",
 ]
 
 examples = [
