### 2026-03-31 c3a7de3
feat: [core] Add FeatureSpec.unique field
