+++ src/metaxy/models/feature_spec.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import hashlib
 import json
-from collections.abc import Mapping, Sequence
+from collections.abc import Mapping, Sequence, Set
 from functools import cached_property
-from typing import TYPE_CHECKING, Annotated, Any, TypeAlias, overload
+from typing import TYPE_CHECKING, Annotated, Any, Literal, TypeAlias, overload
 
 import narwhals as nw
 import pydantic
@@ -171,13 +171,33 @@ class FeatureDep(pydantic.BaseModel):
 IDColumns: TypeAlias = Sequence[str]  # non-bound, should be used for feature specs with arbitrary id columns
 
 CoercibleToFeatureDep: TypeAlias = FeatureDep | type["BaseFeature"] | str | Sequence[str] | FeatureKey
+UniqueKeep: TypeAlias = Literal["any", "latest"]
+
+
+def _validate_column_sequence(value: Any, field_name: str) -> tuple[str, ...]:
+    """Validate column names and preserve the user-provided order."""
+    if isinstance(value, str):
+        raise ValueError(f"{field_name} must be a sequence of column names, not a bare string")
+    if isinstance(value, Mapping):
+        raise ValueError(f"{field_name} must be a sequence of column names, not a mapping")
+    if isinstance(value, Set):
+        raise ValueError(f"{field_name} must be an ordered sequence of column names, not a set")
+    try:
+        return tuple(dict.fromkeys(value))
+    except TypeError as exc:
+        raise ValueError(f"{field_name} must be a sequence of column names") from exc
 
 
 def _validate_id_columns(value: Any) -> tuple[str, ...]:
-    """Coerce id_columns to tuple."""
-    if isinstance(value, tuple):
-        return value
-    return tuple(value)
+    return _validate_column_sequence(value, "id_columns")
+
+
+def _validate_unique_subset(value: Any) -> tuple[str, ...]:
+    return _validate_column_sequence(value, "unique.subset")
+
+
+def _validate_unique_order_by(value: Any) -> tuple[str, ...]:
+    return _validate_column_sequence(value, "unique.order_by")
 
 
 def _validate_deps(value: Any) -> list[FeatureDep]:
@@ -207,10 +227,67 @@ def _validate_deps(value: Any) -> list[FeatureDep]:
 
 
 @public
+class Unique(FrozenBaseModel):
+    """Read-time uniqueness settings for a feature."""
+
+    subset: Annotated[tuple[str, ...], BeforeValidator(_validate_unique_subset)] = pydantic.Field(
+        ...,
+        min_length=1,
+        description=(
+            "Columns used to determine uniqueness. Records with identical values in these columns "
+            "are considered duplicates. Repeated column names are ignored after their first occurrence."
+        ),
+    )
+    keep: UniqueKeep = pydantic.Field(
+        default="any",
+        description=(
+            "Strategy for choosing which row to keep among duplicates. "
+            '"any" picks the row with the lexicographically greatest feature ID. '
+            '"latest" keeps the row with the lexicographically greatest order_by values.'
+        ),
+    )
+    order_by: Annotated[tuple[str, ...], BeforeValidator(_validate_unique_order_by)] | None = pydantic.Field(
+        default=None,
+        min_length=1,
+        description=(
+            'Columns defining the logical order of duplicates when keep="latest". '
+            "The lexicographically greatest values win, with feature ID columns as final logical tie-breakers. "
+            "NULL sorts before every non-NULL value. Physical metadata breaks remaining ties; rejecting "
+            "conflicting ties is the writer's responsibility. Repeated column names are ignored after "
+            "their first occurrence."
+        ),
+    )
+
+    @pydantic.model_validator(mode="after")
[…]
