+++ src/metaxy/ext/ray/datasource.py
@@ -89,12 +89,17 @@ class MetaxyDatasource(Datasource):
         staleness_predicates: Passed to [`resolve_update`][metaxy.MetadataStore.resolve_update]
             in incremental mode.
         filters: Sequence of Narwhals filter expressions to apply.
-        columns: Subset of columns to include. Metaxy's system columns are always included.
+        columns: Exact subset of columns to return. Columns needed to resolve the read are loaded
+            internally and omitted unless requested.
         allow_fallback: If `True`, check fallback stores on main store miss.
-        with_feature_history: If `True`, only return rows with current feature version.
+        with_feature_history: If `True`, include rows from historical feature versions.
+            If `False`, restrict results to the current feature version.
         feature_version: Explicit feature version to filter by (mutually exclusive with `with_feature_history=False`).
-        with_sample_history: Whether to deduplicate samples within `id_columns` groups ordered by `metaxy_created_at`.
+        with_sample_history: Whether to include historical materializations instead of selecting
+            the current row per `id_columns` group using
+            `coalesce(metaxy_deleted_at, metaxy_updated_at)`.
         include_soft_deleted: If `True`, include soft-deleted rows in the result.
+        apply_unique: If `True`, apply the feature's `FeatureSpec.unique` settings.
     """
 
     def __init__(
@@ -112,6 +117,7 @@ class MetaxyDatasource(Datasource):
         with_feature_history: bool = False,
         with_sample_history: bool = False,
         include_soft_deleted: bool = False,
+        apply_unique: bool = True,
     ):
         self.config = mx.init(config)
         self.store = store
@@ -124,6 +130,7 @@ class MetaxyDatasource(Datasource):
         self.with_feature_history = with_feature_history
         self.with_sample_history = with_sample_history
         self.include_soft_deleted = include_soft_deleted
+        self.apply_unique = apply_unique
 
         self._feature_key = mx.coerce_to_feature_key(feature)
 
@@ -141,6 +148,7 @@ class MetaxyDatasource(Datasource):
                 with_feature_history=self.with_feature_history,
                 with_sample_history=self.with_sample_history,
                 include_soft_deleted=self.include_soft_deleted,
+                apply_unique=self.apply_unique,
             )
 
     def _read_incremental_lazy(self) -> nw.LazyFrame[Any]:
