### 2026-03-31 d353501
feat: [core] Add deduplicate_by field to FeatureSpec
