+++ tests/metadata_stores/core/test_deduplicate_by.py
@@ -1,0 +1,432 @@
+"""Tests for FeatureSpec.unique runtime deduplication in the read path."""
+
+from __future__ import annotations
+
+from collections.abc import Callable, Iterator
+from contextlib import contextmanager
+from datetime import datetime, timedelta, timezone
+from pathlib import Path
+from types import SimpleNamespace
+from typing import Any
+from unittest.mock import patch
+
+import narwhals as nw
+import polars as pl
+import pytest
+from metaxy import FeatureGraph, FieldKey, FieldSpec, Unique
+from metaxy.ext.polars.handlers.delta import DeltaMetadataStore
+from metaxy.metadata_store.base import MetadataStore
+from metaxy.models.constants import (
+    ALL_SYSTEM_COLUMNS,
+    METAXY_CREATED_AT,
+    METAXY_DELETED_AT,
+    METAXY_MATERIALIZATION_ID,
+    METAXY_UPDATED_AT,
+)
+from metaxy.utils import collect_to_polars
+from metaxy_testing.models import SampleFeature, SampleFeatureSpec
+
+
+@contextmanager
+def fixed_transaction_time(store: MetadataStore, timestamp: datetime) -> Iterator[None]:
+    """Set a deterministic transaction timestamp for writes within this context."""
+    store._transaction_timestamp = timestamp  # noqa: SLF001
+    try:
+        yield
+    finally:
+        store._transaction_timestamp = None  # noqa: SLF001
+
+
+# ========== Fixtures ==========
+
+
+@pytest.fixture
+def feature_factory(graph: FeatureGraph) -> Callable[..., type[SampleFeature]]:
+    _ = graph
+
+    def build_feature(
+        *,
+        key: str,
+        unique: Unique | None = None,
+    ) -> type[SampleFeature]:
+        class TestFeature(
+            SampleFeature,
+            spec=SampleFeatureSpec(
+                key=key,
+                fields=[FieldSpec(key=FieldKey(["default"]), code_version="1")],
+                unique=unique,
+            ),
+        ):
+            content_hash: str | None = None
+
+        return TestFeature
+
+    return build_feature
+
+
+@pytest.fixture
+def dedup_feature(feature_factory: Callable[..., type[SampleFeature]]) -> type[SampleFeature]:
+    return feature_factory(key="test/deduplicated", unique=Unique(subset=("content_hash",)))
+
+
+@pytest.fixture
+def no_dedup_feature(feature_factory: Callable[..., type[SampleFeature]]) -> type[SampleFeature]:
+    return feature_factory(key="test/no_dedup")
+
+
+@pytest.fixture
+def delta_store(tmp_path: Path) -> DeltaMetadataStore:
+    return DeltaMetadataStore(root_path=tmp_path / "delta")
+
+
+def _make_data(rows: list[dict[str, Any]]) -> pl.DataFrame:
+    return pl.DataFrame(
+        [{**row, "metaxy_provenance_by_field": {"default": f"prov_{row['sample_uid']}"}} for row in rows]
+    )
+
+
+@pytest.fixture
+def data_factory() -> Callable[[list[dict[str, Any]]], pl.DataFrame]:
+    return _make_data
+
+
+@pytest.fixture
+def three_rows_data(data_factory: Callable[[list[dict[str, Any]]], pl.DataFrame]) -> pl.DataFrame:
+    return data_factory(
+        [
+            {"sample_uid": "s1", "content_hash": "aaa"},
+            {"sample_uid": "s2", "content_hash": "aaa"},
+            {"sample_uid": "s3", "content_hash": "bbb"},
+        ]
+    )
+
+
+@pytest.fixture
+def two_duplicate_rows_data(data_factory: Callable[[list[dict[str, Any]]], pl.DataFrame]) -> pl.DataFrame:
+    return data_factory(
[…]
