### 2026-03-31 42c4f10
feat: [core] Add deduplicate_by field to FeatureSpec
