+++ tests/metaxy_testing/src/metaxy_testing/models.py
@@ -21,6 +21,7 @@ if TYPE_CHECKING:
 
     from metaxy.models.feature_spec import (
         CoercibleToFeatureDep,
+        Deduplication,
         FeatureDep,
         IDColumns,
     )
@@ -29,29 +30,32 @@ if TYPE_CHECKING:
 
 # Type aliases
 DefaultFeatureCols: TypeAlias = tuple[Literal["sample_uid"],]
-TestingUIDCols: TypeAlias = list[str]
+TestingUIDCols: TypeAlias = tuple[str, ...]
 
 
 def _validate_sample_feature_spec_id_columns(
     value: Any,
-) -> list[str]:
-    """Coerce id_columns to list for SampleFeatureSpec."""
+) -> tuple[str, ...]:
+    """Coerce id_columns to tuple for SampleFeatureSpec."""
     if value is None:
-        return ["sample_uid"]
-    if isinstance(value, list):
+        return ("sample_uid",)
+    if isinstance(value, tuple):
         return value
-    return list(value)
+    if isinstance(value, str):
+        return (value,)
+    return tuple(value)
 
 
 class SampleFeatureSpec(FeatureSpec):
     """A testing implementation of FeatureSpec that has a `sample_uid` ID column. Has to be moved to tests."""
 
     id_columns: Annotated[
-        pydantic.SkipValidation[list[str]],
+        tuple[str, ...],
         BeforeValidator(_validate_sample_feature_spec_id_columns),
     ] = pydantic.Field(
-        default_factory=lambda: ["sample_uid"],
-        description="List of columns that uniquely identify a row. They will be used by Metaxy in joins.",
+        default_factory=lambda: ("sample_uid",),
+        min_length=1,
+        description="Columns that uniquely identify a row. They will be used by Metaxy in joins.",
     )
 
     if TYPE_CHECKING:
@@ -65,6 +69,7 @@ class SampleFeatureSpec(FeatureSpec):
             deps: list[FeatureDep] | None = None,
             fields: Sequence[str | FieldSpec] | None = None,
             metadata: Mapping[str, JsonValue] | None = None,
+            deduplication: Deduplication | Mapping[str, Any] | None = None,
             **kwargs: Any,
         ) -> None: ...
 
@@ -77,6 +82,7 @@ class SampleFeatureSpec(FeatureSpec):
             deps: list[CoercibleToFeatureDep] | None = None,
             fields: Sequence[str | FieldSpec] | None = None,
             metadata: Mapping[str, JsonValue] | None = None,
+            deduplication: Deduplication | Mapping[str, Any] | None = None,
             **kwargs: Any,
         ) -> None: ...
 
@@ -89,6 +95,7 @@ class SampleFeatureSpec(FeatureSpec):
             deps: list[FeatureDep] | list[CoercibleToFeatureDep] | None = None,
             fields: Sequence[str | FieldSpec] | None = None,
             metadata: Mapping[str, JsonValue] | None = None,
+            deduplication: Deduplication | Mapping[str, Any] | None = None,
             **kwargs: Any,
         ) -> None: ...
 
