+++ src/metaxy/metadata_store/base.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from collections.abc import Iterator, Mapping, Sequence
+from collections.abc import Iterable, Iterator, Mapping, Sequence
 from contextlib import AbstractContextManager, contextmanager
 from datetime import datetime, timezone
 from types import TracebackType
@@ -781,8 +781,11 @@ class MetadataStore(ABC):
             with_feature_history: If True, include rows from all historical feature versions.
                 By default (False), only returns rows with the currently registered feature version.
             with_sample_history: If True, include all historical materializations per sample.
-                By default (False), deduplicates samples within `id_columns` groups ordered by `metaxy_created_at`.
+                By default (False), deduplicates samples within `id_columns` groups ordered by
+                `coalesce(metaxy_deleted_at, metaxy_updated_at)`, keeping the row with the latest timestamp.
+                This disables `FeatureSpec.deduplication`, since historical rows are being requested.
             include_soft_deleted: If `True`, include soft-deleted rows in the result. Previous historical materializations of the same feature version will be effectively removed from the output otherwise.
+                This also disables `FeatureSpec.deduplication`, so deleted rows are not collapsed into active ones.
             with_store_info: If `True`, return a tuple of (LazyFrame, MetadataStore) where
                 the MetadataStore is the store that actually contained the feature (which
                 may be a fallback store if allow_fallback=True).
@@ -796,9 +799,10 @@ class MetadataStore(ABC):
             ValueError: If both feature_version and with_feature_history=False are provided
 
         !!! info
-            When this method is called with default arguments, it will return the latest (by `metaxy_created_at`)
-            metadata for the current feature version excluding soft-deleted rows. Therefore, it's perfectly suitable
-            for most use cases.
+            When this method is called with default arguments, it returns the current metadata for the
+            active feature version, excluding soft-deleted rows. The current row for each `id_columns`
+            group is selected by ordering on `coalesce(metaxy_deleted_at, metaxy_updated_at)` and keeping
+            the latest timestamp. Therefore, it's perfectly suitable for most use cases.
 
         !!! warning
             The order of rows is not guaranteed.
@@ -813,6 +817,7 @@ class MetadataStore(ABC):
 
         feature_key = self._resolve_feature_key(feature)
         is_system_table = self._is_system_table(feature_key)
+        feature_plan: FeaturePlan | None = None
 
         # Sync external features if auto-sync is enabled (default)
         # This call is a no-op most of the time and is very lightweight when it's not
@@ -849,10 +854,23 @@ class MetadataStore(ABC):
         if user_filters:
             read_columns = None
         elif columns and not is_system_table:
-            # Add only system columns that aren't already in the user's columns list
-            columns_set = set(columns)
-            missing_system_cols = [c for c in ALL_SYSTEM_COLUMNS if c not in columns_set]
-            read_columns = [*columns, *missing_system_cols]
+            # Only load dedup columns when deduplication will actually be applied:
+            # requires both with_sample_history=False and filter_deleted=True
+            additional_columns: Iterable[str] | None = None
+            if not with_sample_history:
+                graph = current_graph()
+                definition = graph.feature_definitions_by_key.get(feature_key)
+                if definition is not None:
+                    feature_plan = graph.get_feature_plan(feature_key)
+
+                # Current-view logic groups by id_columns; deduplication needs its own columns too
[…]

[…]
