+++ src/metaxy/metadata_store/ibis.py
@@ -6,10 +6,12 @@ Supports any SQL database that Ibis supports:
 - And 20+ other backends
 """
 
+from __future__ import annotations
+
 from abc import ABC, abstractmethod
 from collections.abc import Iterator, Sequence
 from contextlib import contextmanager
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, overload
 
 import narwhals as nw
 from narwhals.typing import Frame
@@ -27,9 +29,9 @@ from metaxy.metadata_store.exceptions import (
     StoreNotOpenError,
     TableNotFoundError,
 )
-from metaxy.metadata_store.types import AccessMode
+from metaxy.metadata_store.types import AccessMode, TableIdentifier
 from metaxy.models.plan import FeaturePlan
-from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.models.types import CoercibleToFeatureKey
 from metaxy.versioning.ibis import IbisVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
 
@@ -159,7 +161,6 @@ class IbisMetadataStore(MetadataStore, ABC):
             store = IbisMetadataStore(backend="clickhouse", connection_params={"host": "localhost", "port": 9000})
             ```
         """
-        from ibis.backends.sql import SQLBackend
 
         self.connection_string = connection_string
         self.backend = backend
@@ -172,29 +173,34 @@ class IbisMetadataStore(MetadataStore, ABC):
             versioning_engine=versioning_engine,
         )
 
-    def _has_feature_impl(self, feature: CoercibleToFeatureKey) -> bool:
-        feature_key = self._resolve_feature_key(feature)
-        table_name = self.get_table_name(feature_key)
-        return table_name in self.conn.list_tables()
+    def _has_feature_impl(self, table_id: TableIdentifier) -> bool:
+        return self.get_table_name(table_id) in self.conn.list_tables()
+
+    @overload
+    def get_table_name(self, table_id: TableIdentifier) -> str: ...
+
+    @overload
+    def get_table_name(self, table_id: CoercibleToFeatureKey) -> str: ...
 
     def get_table_name(
         self,
-        key: FeatureKey,
+        table_id: TableIdentifier | CoercibleToFeatureKey,
     ) -> str:
-        """Generate the storage table name for a feature or system table.
+        """Generate the storage table name for a table identifier.
 
-        Applies the configured table_prefix (if any) to the feature key's table name.
+        Applies the configured table_prefix (if any) to the identifier's table name.
         Subclasses can override this method to implement custom naming logic.
 
         Args:
-            key: Feature key to convert to storage table name.
+            table_id: Storage-layer table identifier or feature reference.
 
         Returns:
             Storage table name with optional prefix applied.
         """
-        base_name = key.table_name
+        if not isinstance(table_id, TableIdentifier):
+            table_id = self._resolve_table_id(table_id)
 
-        return f"{self._table_prefix}{base_name}" if self._table_prefix else base_name
+        return f"{self._table_prefix}{table_id.table_name}" if self._table_prefix else table_id.table_name
 
     def _get_default_hash_algorithm(self) -> HashAlgorithm:
         """Get default hash algorithm for Ibis stores.
@@ -267,7 +273,7 @@ class IbisMetadataStore(MetadataStore, ABC):
             )
 
     @property
-    def conn(self) -> "SQLBackend":
+    def conn(self) -> SQLBackend:
         """Get Ibis backend connection.
 
         Returns:
@@ -330,7 +336,7 @@ class IbisMetadataStore(MetadataStore, ABC):
 
     def _write_feature(
         self,
-        feature_key: FeatureKey,
+        table_id: TableIdentifier,
         df: Frame,
         **kwargs: Any,
     ) -> None:
[…]
