### 2026-04-01 899d15e
fix: improve `Map` datatype support in `MetadataStore.resolve_update` (#1139)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
