### 2026-04-01 f3df442
fix: remove missing lock file warning when `metaxy_lock_path` is not configured (#1143)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
