### 2026-04-01 7b29910
chore: enable Tach pytest plugin
