+++ src/metaxy/ext/metadata_stores/iceberg.py
@@ -19,13 +19,13 @@ from packaging.version import Version
 from pydantic import Field
 
 from metaxy._decorators import public
-from metaxy._utils import collect_to_polars
 from metaxy.config import MetaxyConfig
 from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
 from metaxy.metadata_store.types import AccessMode
 from metaxy.metadata_store.utils import is_local_path
 from metaxy.models.plan import FeaturePlan
 from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.utils import collect_to_polars
 from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
 
