+++ tests/versioning/test_polars_map.py
@@ -11,8 +11,8 @@ import pyarrow as pa
 import pytest
 from polars_map import Map
 
-from metaxy._utils import lazy_frame_to_polars
 from metaxy.config import MetaxyConfig
+from metaxy.utils.dataframes import lazy_frame_to_polars
 from metaxy.versioning._arrow_map import (
     convert_extension_maps_to_native,
     convert_maps_to_polars_map,
