### 2026-04-01 1922134
fix: resolve CI lint and typecheck errors with collect_to_arrow import
