+++ src/metaxy/cli/metadata.py
@@ -3,7 +3,8 @@
 from __future__ import annotations
 
 import json
-from typing import TYPE_CHECKING, Annotated, Any, Literal
+import sys
+from typing import TYPE_CHECKING, Annotated, Any, Literal, TypeAlias
 
 import cyclopts
 import narwhals as nw
@@ -12,8 +13,10 @@ from rich.table import Table
 from metaxy.cli.console import console, data_console, error_console
 from metaxy.cli.utils import FeatureSelector, FilterArgs, GlobalFilterArgs, OutputFormat, StalenessPredicateArgs
 
+WriteFormat: TypeAlias = Literal["csv", "json", "parquet", "markdown"]
+
 if TYPE_CHECKING:
-    import polars as pl
+    import pyarrow as pa
 
     from metaxy import FeatureDefinition
     from metaxy.graph.status import FeatureMetadataStatusWithIncrement
@@ -113,20 +116,14 @@ def status(
         selector.resolve(format, graph=graph, error_missing=assert_in_sync)
         missing_keys = selector.missing_keys
 
-        # Handle empty result
+        # Handle empty result for --all-features
+        if selector.all_features and not selector:
+            _output_no_features_warning(format, project_version)
+            return
+
+        # If no valid features remain
         if not selector:
-            if selector.all_features:
-                _output_no_features_warning(format, project_version)
-            else:
-                if format == "json":
-                    data_console.print_json(
-                        data={
-                            "warning": "No valid features to check",
-                            "features": {},
-                        }
-                    )
-                else:
-                    data_console.print("[yellow]Warning:[/yellow] No valid features to check.")
+            _output_no_features_warning(format, project_version)
             return
 
         # Print header for plain format only when using a snapshot
@@ -635,7 +632,7 @@ def rebase(
     if dry_run:
         console.print(f"[bold cyan]Dry run[/bold cyan]: {rows_affected} row(s) would be rebased")
     else:
-        console.print(f"[green]SUCCESS:[/green] Rebased {rows_affected} row(s)")
+        console.print(f"[green]✓[/green] Rebased {rows_affected} row(s)")
 
 
 @app.command()
@@ -712,7 +709,7 @@ def copy(
     dest_store = context.get_store(to_store)
 
     # Resolve feature keys
-    selector.resolve("plain")
+    selector.resolve("plain", error_missing=False)
 
     # Handle no valid features
     if not selector:
@@ -737,7 +734,7 @@ def copy(
             )
 
     data_console.print(
-        f"[green]SUCCESS:[/green] Copy complete: {stats['features_copied']} feature(s), {stats['rows_copied']} row(s) copied"
+        f"[green]✓[/green] Copy complete: {stats['features_copied']} feature(s), {stats['rows_copied']} row(s) copied"
     )
 
 
@@ -760,13 +757,6 @@ def read(
             help="Columns to select. Can be repeated.",
         ),
     ] = None,
-    format: Annotated[
-        Literal["csv", "json", "parquet", "markdown"],
-        cyclopts.Parameter(
-            name=["--format", "-f"],
-            help="Output format: markdown, json, csv, parquet.",
-        ),
-    ] = "markdown",
     query: Annotated[
         str | None,
         cyclopts.Parameter(
@@ -785,6 +775,12 @@ def read(
             help="Output file path (default: print to stdout).",
         ),
     ] = None,
+    format: Annotated[
+        WriteFormat,
+        cyclopts.Parameter(
+            name=["--format", "-f"],
+        ),
+    ] = "markdown",
 ) -> None:
     """Read feature metadata and optionally execute arbitrary SQL against it.
 
@@ -801,11 +797,9 @@ def read(
         $ metaxy metadata read my/feature --query "SELECT * FROM my__feature WHERE status = 'active'" -f csv
         ```
     """
[…]
