### 2026-04-01 0291722
ci: fix tach setup in `main` branch
