### 2026-04-01 44a840e
Refactor metadata read CLI and tests following maintainer feedback (no Polars, parameterized tests)
