+++ tests/cli/test_cli_metadata_read.py
@@ -1,14 +1,13 @@
-"""Tests for the metadata read CLI command."""
+"""Refactored tests for the metadata read CLI command."""
 
 import json
 
-import polars as pl
 import pytest
 from metaxy_testing import TempMetaxyProject
-from polars.testing import assert_frame_equal
 
 
 def _define_features():
+    """Define test features for the metadata read command."""
     from metaxy_testing.models import SampleFeatureSpec
 
     from metaxy import BaseFeature, FeatureKey, FieldKey, FieldSpec
@@ -31,564 +30,133 @@ def _define_features():
     ):
         pass
 
-    class FilesMarkdown(
-        BaseFeature,
-        spec=SampleFeatureSpec(
-            key=FeatureKey(["files_markdown"]),
-            fields=[FieldSpec(key=FieldKey(["default"]), code_version="1")],
+
+@pytest.fixture
+def project_with_data(metaxy_project: TempMetaxyProject):
+    """Fixture that initializes a project with sample metadata written to the 'dev' store."""
+    with metaxy_project.with_features(_define_features):
+        from metaxy.models.types import FeatureKey
+
+        graph = metaxy_project.graph
+        store = metaxy_project.stores["dev"]
+
+        import ibis
+
+        # Create sample data using Ibis (avoid Polars as requested)
+        data = ibis.memtable(
+            {
+                "sample_uid": [1, 2, 3],
+                "value": ["val_1", "val_2", "val_3"],
+                "category": ["A", "B", "A"],
+                "metaxy_provenance_by_field": [
+                    {"default": "hash1"},
+                    {"default": "hash2"},
+                    {"default": "hash3"},
+                ],
+            }
+        )
+
+        with graph.use(), store.open("w"):
+            store.write(FeatureKey(["files_root"]), data)
+            store.write(FeatureKey(["files_parquet"]), data)
+
+        yield metaxy_project
+
+
+@pytest.mark.parametrize(
+    "fmt, check_fn",
+    [
+        ("markdown", lambda out: "sample_uid" in out),
+        ("json", lambda out: "sample_uid" in out and '"sample_uid":1' in out.replace(" ", "")),
+        ("csv", lambda out: "sample_uid" in out and "value" in out and "category" in out),
+    ],
+)
+def test_metadata_read_formats(project_with_data, capsys, fmt, check_fn):
+    """Test reading metadata in different text formats."""
+    result = project_with_data.run_cli(["metadata", "read", "files_root", "-f", fmt], capsys=capsys)
+    assert result.returncode == 0
+    assert check_fn(result.stdout)
+
+
+def test_metadata_read_parquet(project_with_data, tmp_path, capsys):
+    """Test outputting metadata to a parquet file."""
+    output_file = tmp_path / "output.parquet"
+    result = project_with_data.run_cli(
+        ["metadata", "read", "files_root", "-f", "parquet", "-o", str(output_file)],
+        capsys=capsys
+    )
+    assert result.returncode == 0
+    assert output_file.exists()
+    
+    # Verify content using PyArrow (avoid Polars)
+    import pyarrow.parquet as pq
+    table = pq.read_table(output_file)
+    assert table.num_rows == 3
+    assert "value" in table.column_names
+
+
+@pytest.mark.parametrize(
+    "options, expected_count, check_fn",
+    [
+        (
+            ["--select", "sample_uid", "--select", "value", "--filter", "category = 'A'"],
+            2,
+            lambda out: "val_1" in out and "val_3" in out and "val_2" not in out,
         ),
-    ):
-        pass
-
-    class FilesJSON(
-        BaseFeature,
-        spec=SampleFeatureSpec(
-            key=FeatureKey(["files_json"]),
-            fields=[FieldSpec(key=FieldKey(["default"]), code_version="1")],
+        (
[…]
