### 2026-04-01 8379d4a
test: fix coverage for release PRs by testing `metaxy.__version__` (#1150)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
