### 2026-04-01 e7d9d51
Fix CLI AttributeError by wrapping Ibis sql result in nw.from_native before calling collect_to_arrow under DuckDB mode
