### 2026-04-01 9610dc4
Deployed 7c974bfb to 0.1.8 with MkDocs 1.6.1 and mike 2.1.3
