+++ 0.1.8/slides/2026-introducing-metaxy/dist/assets/chunk-YJGJQOYZ-z4wX3TcS.js
@@ -1,0 +1,1 @@
+import{a as s}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";var t,e=(t=class{constructor(i){this.init=i,this.records=this.init()}reset(){this.records=this.init()}},s(t,"ImperativeState"),t);export{e as I};
