+++ 0.1.8/slides/2026-introducing-metaxy/dist/assets/md-BS4GpBhh.js
@@ -1,0 +1,1 @@
+import{b as a,o as i,w as c,g as e,D as t,v as l,x as p,z as o}from"./modules/vue-D2_2azlL.js";import{I as d}from"./slidev/default-DMaQI0-d.js";import{u as m,f as u}from"./slidev/context-CXJvKEzn.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";const f=""+new URL("../img/anatomy.svg",import.meta.url).href,C={__name:"introducing-metaxy.md__slidev_5",setup(g){const{$clicksContext:s,$frontmatter:n}=m();return s.setup(),(_,r)=>(i(),a(d,l(p(o(u)(o(n),4))),{default:c(()=>[...r[0]||(r[0]=[e("h1",null,"Field-Level Dependencies",-1),e("p",null,[t("Features declare which "),e("strong",null,"specific upstream fields"),t(" they depend on.")],-1),e("div",{class:"flex justify-center mt-2"},[e("img",{src:f,class:"h-80"})],-1),e("div",{"mt-2":"","text-center":""},[e("p",null,[e("code",null,"SpeechToText.transcription"),t(" depends on "),e("code",null,"Video.audio"),t(" only. Changing frames does not trigger transcription reprocessing.")])],-1)])]),_:1},16))}};export{C as default};

[…]
