### 2026-04-01 ee47b74
style: fix ruff format issues
