+++ .github/workflows/main.yml
@@ -31,6 +31,8 @@ jobs:
       python-versions: ${{ github.event.inputs.full-matrix == 'true' && '["3.10", "3.11", "3.12", "3.13", "3.14"]' || '["3.10"]' }}
       run-windows: ${{ github.event.inputs.full-matrix == 'true' }}
       run-macos: ${{ github.event.inputs.full-matrix == 'true' }}
+      run-all-checks: ${{ github.event.inputs.full-matrix == 'true' }}
+      use-tach: false
       dependency-matrix: ${{ github.event.inputs.full-matrix == 'true' }}
     secrets:
       CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
