+++ examples/example-expansion/README.md
@@ -1,3 +1,3 @@
 # Expansion
 
-View this example in [Metaxy docs](https://docs.metaxy.io/latest/examples/expansion/).
+View this example in [Metaxy docs](https://docs.metaxy.io/stable/examples/expansion/).
