+++ pyproject.toml
@@ -114,7 +114,7 @@ mx = "metaxy.cli.app:main"
 Documentation = "https://docs.metaxy.io"
 Repository = "https://github.com/anam-org/metaxy"
 Issues = "https://github.com/anam-org/metaxy/issues"
-Changelog = "https://docs.metaxy.io/latest/changelog"
+Changelog = "https://docs.metaxy.io/stable/changelog"
 
 [build-system]
 requires = ["uv_build>=0.8.14,<0.9.0"]
