+++ .pre-commit-config.yaml
@@ -56,6 +56,7 @@ repos:
       language: system
       pass_filenames: false
       files: ^(src/|docs/|pyproject\.toml|uv\.lock|mkdocs\.yml|\.pre-commit-config\.yaml)
+
     - id: tach
       name: Check API Boundaries with Tach
       entry: uv run tach check
