### 2026-04-02 010ee23
refactor: move metadata stores to `metaxy.ext.<store>` (#1153)
## Summary

Resolve #1154

<!-- Brief description of the changes -->

## Changelog

Metadata stores have been moved from `metaxy.ext.metadata_stores.<store>` to `metaxy.ext.<store>`. DeltaLake and Iceberg stores are additionally nested under `metaxy.ext.polars.handlers.`. Old locations have been kept for backwards compatibility but will be removed in `0.2.0`.

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was 
[…]
