+++ src/metaxy/ext/lancedb/metadata_store.py
@@ -1,0 +1,376 @@
+"""LanceDB metadata store implementation."""
+
+from __future__ import annotations
+
+import logging
+from collections.abc import Sequence
+from contextlib import contextmanager
+from pathlib import Path
+from typing import Any
+
+import narwhals as nw
+from narwhals.typing import Frame
+from pydantic import Field
+
+from metaxy._decorators import public
+from metaxy.metadata_store.base import MetadataStore, MetadataStoreConfig
+from metaxy.metadata_store.types import AccessMode
+from metaxy.metadata_store.utils import is_local_path, sanitize_uri
+from metaxy.models.types import CoercibleToFeatureKey, FeatureKey
+from metaxy.utils import collect_to_polars
+from metaxy.versioning.polars import PolarsVersioningEngine
+from metaxy.versioning.types import HashAlgorithm
+
+logger = logging.getLogger(__name__)
+
+
+@public
+class LanceDBMetadataStoreConfig(MetadataStoreConfig):
+    """Configuration for LanceDBMetadataStore.
+
+    Example:
+        ```toml title="metaxy.toml"
+        [stores.dev]
+        type = "metaxy.ext.lancedb.LanceDBMetadataStore"
+
+        [stores.dev.config]
+        uri = "/path/to/featuregraph"
+
+        [stores.dev.config.connect_kwargs]
+        api_key = "your-api-key"
+        ```
+    """
+
+    uri: str | Path = Field(
+        description="Directory path or URI for LanceDB tables.",
+    )
+    connect_kwargs: dict[str, Any] | None = Field(
+        default=None,
+        description="Extra keyword arguments passed to lancedb.connect().",
+    )
+
+
+@public
+class LanceDBMetadataStore(MetadataStore):
+    """
+    [LanceDB](https://lancedb.github.io/lancedb/) metadata store for vector and structured data.
+
+    LanceDB is a columnar database optimized for vector search and multimodal data.
+    Each feature is stored in its own Lance table within the database directory.
+    Uses Polars components for data processing (no native SQL execution).
+
+    Storage layout:
+
+    - Each feature gets its own table: `{namespace}__{feature_name}`
+
+    - Tables are stored as Lance format in the directory specified by the URI
+
+    - LanceDB handles schema evolution, transactions, and compaction automatically
+
+    Example: Local Directory
+        ```py
+        from pathlib import Path
+        from metaxy.ext.lancedb import LanceDBMetadataStore
+
+        # Local filesystem
+        store = LanceDBMetadataStore(Path("/path/to/featuregraph"))
+        ```
+
+    Example: Object Storage (S3, GCS, Azure)
+        ```py
+        # object store (requires credentials)
+        store = LanceDBMetadataStore("s3:///path/to/featuregraph")
+        ```
+
+    Example: LanceDB Cloud
+        ```py
+        import os
+
+        # Option 1: Environment variable
+        os.environ["LANCEDB_API_KEY"] = "your-api-key"
+        store = LanceDBMetadataStore("db://my-database")
+
+        # Option 2: Explicit credentials
+        store = LanceDBMetadataStore(
+            "db://my-database", connect_kwargs={"api_key": "your-api-key", "region": "us-east-1"}
+        )
+        ```
+    """
+
+    _should_warn_auto_create_tables = False
+    versioning_engine_cls = PolarsVersioningEngine
+
+    def __init__(
+        self,
+        uri: str | Path,
+        *,
+        fallback_stores: list[MetadataStore] | None = None,
+        connect_kwargs: dict[str, Any] | None = None,
+        **kwargs: Any,
+    ):
+        """
+        Initialize [LanceDB](https://lancedb.com/docs/) metadata store.
+
[…]
