+++ src/metaxy/ext/metadata_stores/bigquery.py
@@ -1,349 +1,8 @@
-"""BigQuery metadata store - thin wrapper around IbisMetadataStore."""
+"""Deprecated: use metaxy.ext.bigquery.metadata_store instead."""
 
-from typing import TYPE_CHECKING, Any
+from metaxy._warnings import _warn_deprecated_module
+from metaxy.ext.bigquery import BigQueryMetadataStore, BigQueryMetadataStoreConfig
 
-if TYPE_CHECKING:
-    from metaxy.metadata_store.base import MetadataStore
+_warn_deprecated_module("metaxy.ext.metadata_stores.bigquery", "metaxy.ext.bigquery")
 
-from pydantic import Field
-
-from metaxy._decorators import public
-from metaxy.metadata_store.ibis import IbisMetadataStore, IbisMetadataStoreConfig
-from metaxy.versioning.types import HashAlgorithm
-
-
-@public
-class BigQueryMetadataStoreConfig(IbisMetadataStoreConfig):
-    """Configuration for BigQueryMetadataStore.
-
-    Example:
-        ```toml title="metaxy.toml"
-        [stores.dev]
-        type = "metaxy.ext.metadata_stores.bigquery.BigQueryMetadataStore"
-
-        [stores.dev.config]
-        project_id = "my-project"
-        dataset_id = "my_dataset"
-        credentials_path = "/path/to/service-account.json"
-        ```
-    """
-
-    project_id: str | None = Field(default=None, description="Google Cloud project ID containing the dataset.")
-    dataset_id: str | None = Field(default=None, description="BigQuery dataset name for storing metadata tables.")
-    credentials_path: str | None = Field(default=None, description="Path to service account JSON file.")
-    credentials: Any | None = Field(default=None, description="Google Cloud credentials object.")
-    location: str | None = Field(
-        default=None,
-        description="Default location for BigQuery resources (e.g., 'US', 'EU').",
-    )
-
-
-@public
-class BigQueryMetadataStore(IbisMetadataStore):
-    """
-    [BigQuery](https://cloud.google.com/bigquery) metadata store using [Ibis](https://ibis-project.org/) backend.
-
-    Warning:
-        It's on the user to set up infrastructure for Metaxy correctly.
-        Make sure to have large tables partitioned as appropriate for your use case.
-
-    Note:
-        BigQuery automatically optimizes queries on partitioned tables.
-        When tables are partitioned (e.g., by date or ingestion time with _PARTITIONTIME), BigQuery will
-        automatically prune partitions based on WHERE clauses in queries, without needing
-        explicit configuration in the metadata store.
-        Make sure to use appropriate `filters` when calling [BigQueryMetadataStore.read][metaxy.ext.metadata_stores.bigquery.BigQueryMetadataStore.read].
-
-    Example: Basic Connection
-        <!-- skip next -->
-        ```py
-        store = BigQueryMetadataStore(
-            project_id="my-project",
-            dataset_id="my_dataset",
-        )
-        ```
-
-    Example: With Service Account
-        <!-- skip next -->
-        ```py
-        store = BigQueryMetadataStore(
-            project_id="my-project",
-            dataset_id="my_dataset",
-            credentials_path="/path/to/service-account.json",
-        )
-        ```
-
-    Example: With Location Configuration
-        <!-- skip next -->
-        ```py
-        store = BigQueryMetadataStore(
-            project_id="my-project",
-            dataset_id="my_dataset",
-            location="EU",  # Specify data location
-        )
-        ```
-
-    Example: With Custom Hash Algorithm
-        <!-- skip next -->
-        ```py
-        store = BigQueryMetadataStore(
-            project_id="my-project",
-            dataset_id="my_dataset",
[…]
