### 2026-04-02 798393a
test: migrate tests to use new metadata store imports (#1155)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->

`MovedMetadataStoreWarning` has been blacklisted in tests
