+++ tests/metaxy_testing/src/metaxy_testing/metaxy_project.py
@@ -944,13 +944,13 @@ store = "dev"
 auto_create_tables = true
 
 [stores.dev]
-type = "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore"
+type = "metaxy.ext.duckdb.DuckDBMetadataStore"
 
 [stores.dev.config]
 database = "{dev_db_path}"
 
 [stores.staging]
-type = "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore"
+type = "metaxy.ext.duckdb.DuckDBMetadataStore"
 
 [stores.staging.config]
 database = "{staging_db_path}"
