+++ tests/test_toml_source.py
@@ -12,7 +12,7 @@ class _DummySettings(BaseSettings):
     pass
 
 
-STORE_TYPE = "metaxy.ext.metadata_stores.delta.DeltaMetadataStore"
+STORE_TYPE = "metaxy.ext.polars.handlers.delta.DeltaMetadataStore"
 
 
 # --- Basic loading ---
