+++ tests/metadata_stores/core/test_utils.py
@@ -9,7 +9,7 @@ from ibis.common.collections import FrozenOrderedDict
 from ibis.common.temporal import IntervalUnit
 
 from metaxy.ext.duckdb import DuckDBMetadataStore
-from metaxy.metadata_store.ibis import IbisMetadataStore
+from metaxy.ext.ibis.metadata_store import IbisMetadataStore
 from metaxy.metadata_store.utils import is_local_path
 
 
