### 2026-04-02 0e00d0f
refactor: move `PolarsVersioningEngine` to `metaxy.ext.polars.versioning` (#1159)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
