+++ tests/ext/postgres/test_native.py
@@ -9,6 +9,7 @@ import polars as pl
 import polars.testing
 import pytest
 
+from metaxy.ext.polars.versioning import PolarsVersioningEngine
 from metaxy.ext.postgresql import PostgreSQLMetadataStore
 from metaxy.metadata_store import MetadataStore
 from metaxy.metadata_store.exceptions import HashAlgorithmNotSupportedError
@@ -16,7 +17,6 @@ from metaxy.models.constants import METAXY_DATA_VERSION_BY_FIELD, METAXY_PROVENA
 from metaxy.models.feature import FeatureGraph, current_graph
 from metaxy.models.types import FeatureKey
 from metaxy.utils import collect_to_polars
-from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
 from tests.ext.postgres.conftest import _with_search_path
 from tests.metadata_stores.shared import (
