+++ tests/provenance/test_polars_tracker.py
@@ -7,8 +7,8 @@ import polars as pl
 import pytest
 from metaxy_testing.models import SampleFeature
 
+from metaxy.ext.polars.versioning import PolarsVersioningEngine
 from metaxy.models.feature import FeatureGraph
-from metaxy.versioning.polars import PolarsVersioningEngine
 from metaxy.versioning.types import HashAlgorithm
 
 
