+++ tests/ext/sqlalchemy/test_native.py
@@ -31,7 +31,7 @@ def config_with_features():
         store="test_store",
         stores={  # ty: ignore[invalid-argument-type]
             "test_store": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                 "config": {"database": ":memory:"},
             }
         },
@@ -90,7 +90,7 @@ def test_get_metaxy_system_metadata():
         store="test_store",
         stores={  # ty: ignore[invalid-argument-type]
             "test_store": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                 "config": {"database": ":memory:"},
             }
         },
@@ -113,7 +113,7 @@ def test_feature_versions_table_columns_match_polars_schema():
         store="test_store",
         stores={  # ty: ignore[invalid-argument-type]
             "test_store": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                 "config": {"database": ":memory:"},
             }
         },
@@ -177,7 +177,7 @@ def test_get_store_sqlalchemy_url_default_store():
         store="test_store",
         stores={  # ty: ignore[invalid-argument-type]
             "test_store": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                 "config": {"database": "test.db"},
             }
         },
@@ -195,11 +195,11 @@ def test_get_store_sqlalchemy_url_named_store():
         project="test",
         stores={  # ty: ignore[invalid-argument-type]
             "store1": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                 "config": {"database": "store1.db"},
             },
             "store2": {
-                "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                 "config": {"database": "store2.db"},
             },
         },
@@ -224,7 +224,7 @@ class TestProtocolParameter:
             store="test_store",
             stores={  # ty: ignore[invalid-argument-type]
                 "test_store": {
-                    "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                    "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                     "config": {"database": "test.db"},
                 }
             },
@@ -246,7 +246,7 @@ class TestProtocolParameter:
             store="test_store",
             stores={  # ty: ignore[invalid-argument-type]
                 "test_store": {
-                    "type": "metaxy.ext.metadata_stores.duckdb.DuckDBMetadataStore",
+                    "type": "metaxy.ext.duckdb.DuckDBMetadataStore",
                     "config": {"database": "mydb.db"},
                 }
             },
