### 2026-04-02 736ec09
test: fix `source_roots` and remove utility modules in  `tach.toml` (#1162)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
