+++ tach.toml
@@ -7,162 +7,13 @@ exclude = [
   "**/venv",
 ]
 source_roots = ["src", "."]
-
-[[modules]]
-path = "metaxy.entrypoints"
-depends_on = ["metaxy.models"]
-
-[[modules]]
-path = "metaxy.ext.dagster"
-depends_on = [
-  "metaxy",
-  "metaxy._version",
-  "metaxy.models",
-  "metaxy.metadata_store.exceptions",
-  "metaxy._decorators",
-  "metaxy.models.constants",
-  "metaxy.utils",
-  "metaxy.models.types",
+layers = [
+  "cli",
+  "ext",
+  "core",
 ]
 
-[[modules]]
-path = "metaxy._hashing"
-depends_on = ["metaxy"]
-
-[[modules]]
-path = "metaxy._warnings"
-depends_on = []
-
-[[modules]]
-path = "metaxy.models"
-depends_on = [
-  "metaxy._hashing",
-  "metaxy._packaging",
-  "metaxy.models.constants",
-  "metaxy._decorators",
-  "metaxy.config",
-  "metaxy.models.types",
-  "metaxy.versioning",
-  "metaxy.utils",
-]
-
-[[modules]]
-path = "metaxy.ext.sqlalchemy"
-depends_on = [
-  "metaxy.metadata_store",
-  "metaxy.config",
-  "metaxy.models.constants",
-  "metaxy.models",
-  "metaxy._decorators",
-]
-
-[[modules]]
-path = "metaxy.ext.sqlmodel"
-depends_on = [
-  "metaxy.ext.sqlalchemy",
-  "metaxy",
-  "metaxy._decorators",
-  "metaxy.models",
-  "metaxy.models.constants",
-  "metaxy.models.types",
-  "metaxy.config",
-]
-
-[[modules]]
-path = "metaxy.versioning"
-depends_on = [
-  "metaxy.models",
-  "metaxy.utils",
-  "metaxy.config",
-  "metaxy.models.constants",
-  "metaxy._decorators",
-  "metaxy.models.types",
-  "metaxy._hashing",
-]
-
-[[modules]]
-path = "metaxy._exceptions"
-depends_on = ["metaxy.utils"]
-
-[[modules]]
-path = "metaxy.metadata_store"
-depends_on = [
-  "metaxy.models",
-  "metaxy.config",
-  "metaxy._version",
-  "metaxy.metadata_store.base",
-  "metaxy.metadata_store.exceptions",
-  "metaxy.metadata_store.types",
-  "metaxy.models.constants",
-  "metaxy._warnings",
-  "metaxy.models.types",
-]
-
-[[modules]]
-path = "metaxy.cli"
-depends_on = [
-  "metaxy.ext.mcp",
-  "metaxy.metadata_store",
-  "metaxy",
-  "metaxy.models.constants",
-  "metaxy.models",
-  "metaxy.models.types",
-  "metaxy.config",
-  "metaxy._version",
-  "metaxy.graph",
-  "metaxy.metadata_store.exceptions",
-  "metaxy.utils",
-]
-
-[[modules]]
-path = "metaxy.utils"
-depends_on = [
-  "metaxy.metadata_store",
-  "metaxy",
-  "metaxy.versioning",
-  "metaxy._exceptions",
-  "metaxy._warnings",
-  "metaxy.metadata_store.exceptions",
-  "metaxy._decorators",
-  "metaxy.models.types",
-  "metaxy.config",
-  "metaxy.models",
-]
-
-[[modules]]
-path = "metaxy.ext.ray"
-depends_on = ["metaxy", "metaxy.utils"]
-
-[[modules]]
-path = "metaxy.ext.mcp"
-depends_on = ["metaxy", "metaxy.models"]
-
[…]
