### 2026-04-03 9428e7a
feat(map datatype): use `narwhals-map` when `Map` datatype is enabled (#1167)
## Summary

Implement backend-agnostic `Map` datatype support with [`narwhals-map`](https://pypi.org/project/narwhals-map/).

## Changelog

[`narwhals-map`](https://pypi.org/project/narwhals-map/) is now used for `Map` datatype operations in Narwhals. This changes the Narwhals datatype from `Unknown` to `narwhals_map.Map` and exposes `nw.Expr.map` namespace for these columns.

## Test Plan

Existing tests + new tests for I/O with `narwhals_map.Map` columns.
