+++ tests/metadata_stores/shared/ibis_map.py
@@ -138,3 +138,52 @@ class IbisMapTests:
 
         assert table.num_rows == 2
         assert pa.types.is_map(table.schema.field("metaxy_provenance_by_field").type)
+
+    # ── narwhals-map: write narwhals Ibis-backed frames with Map columns ──
+
+    def test_ibis_table_with_map_columns(
+        self,
+        polars_map_config: MetaxyConfig,
+        store: MetadataStore,
+        test_features: dict[str, FeatureDefinition],
+    ) -> None:
+        """An Ibis table with native Map columns survives write→read."""
+        feature = test_features["UpstreamFeatureA"]
+
+        ibis_table = ibis.memtable(
+            pa.table(
+                {
+                    "sample_uid": pa.array([1, 2]),
+                    "metaxy_provenance_by_field": pa.array(
+                        [
+                            [("frames", "f1"), ("audio", "a1")],
+                            [("frames", "f2"), ("audio", "a2")],
+                        ],
+                        type=pa.map_(pa.string(), pa.string()),
+                    ),
+                    "scores": pa.array(
+                        [
+                            [(1, 0.95), (2, 0.87)],
+                            [(1, 0.91), (2, 0.82)],
+                        ],
+                        type=pa.map_(pa.int32(), pa.float32()),
+                    ),
+                }
+            )
+        )
+        assert ibis_table.scores.type().is_map()
+
+        with store.open("w") as s:
+            s.write(feature, ibis_table)
+
+        with store.open("r") as s:
+            result = s.read(feature)
+            assert result is not None
+            table = collect_to_arrow(result).sort_by("sample_uid")
+
+        assert pa.types.is_map(table.schema.field("metaxy_provenance_by_field").type)
+        scores_type = table.schema.field("scores").type
+        assert pa.types.is_map(scores_type)
+        assert scores_type.key_type == pa.int32()
+        assert scores_type.item_type == pa.float32()
+        assert dict(table.column("scores")[0].as_py()) == {1: pytest.approx(0.95), 2: pytest.approx(0.87)}
