+++ conftest.py
@@ -1,2 +1,4 @@
 # Root conftest.py - intentionally minimal
 # Sybil docstring testing is configured in src/metaxy/conftest.py
+
+import narwhals_map  # noqa: F401  # registers Map dtype and .map namespace on narwhals
