### 2026-04-03 1877577
docs(map datatype): document `narwhals-map` usage for `Map` datatype (#1171)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
