+++ 0.1.9/slides/2026-introducing-metaxy/dist/assets/infoDiagram-DRJXUV6F-DIjhDQof.js
@@ -1,0 +1,2 @@
+import{p as i}from"./treemap-KMMF4GRG-YVX6V4UE-DGcgu5Gs.js";import{a as e,l as o,a7 as n,f as p,a8 as g}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";import"./chunk-ZZTYOBSU-CHxtuk5s.js";import"./index-B2pMgQ-r.js";import"./modules/vue-D2_2azlL.js";import"./modules/shiki-bZSEl0vS.js";import"./lz-string-BCxBxReP.js";import"./modules/file-saver-B7oFTzqn.js";var m={parse:e(async r=>{const a=await i("info",r);o.debug(a)},"parse")},v={version:g.version+""},d=e(()=>v.version,"getVersion"),c={getVersion:d},f=e((r,a,s)=>{o.debug(`rendering info diagram
+`+r);const t=n(a);p(t,100,400,!0),t.append("g").append("text").attr("x",100).attr("y",40).attr("class","version").attr("font-size",32).style("text-anchor","middle").text(`v${s}`)},"draw"),l={draw:f},E={parser:m,db:c,renderer:l};export{E as diagram};
