+++ 0.1.9/slides/2026-introducing-metaxy/dist/assets/chunk-SCLVSU6P-VHFMvI2b.js
@@ -1,0 +1,15 @@
+import{a as e}from"./Mermaid.vue_vue_type_script_setup_true_lang-DK1r9W0A.js";var l=e(()=>`
+  /* Font Awesome icon styling - consolidated */
+  .label-icon {
+    display: inline-block;
+    height: 1em;
+    overflow: visible;
+    vertical-align: -0.125em;
+  }
+  
+  .node .label-icon path {
+    fill: currentColor;
+    stroke: revert;
+    stroke-width: revert;
+  }
+`,"getIconStyles");export{l as g};
