### 2026-04-08 f1cdd5d
chore(publications): paper SAO Workshop (#1174)
## Summary

add publication for SAO https://bauplanlabs.github.io/SAO-workshop/#call-for-papers
[paper.pdf](https://github.com/user-attachments/files/26504749/paper.pdf)

we have to submit it before 30th April
