+++ publications/2026-metaxy-field-level-versioning/paper.tex
@@ -1,0 +1,160 @@
+\documentclass[sigconf]{acmart}
+
+%% Remove copyright / ACM reference for workshop submission draft
+\setcopyright{none}
+\settopmatter{printacmref=false}
+\renewcommand\footnotetextcopyrightpermission[1]{}
+\pagestyle{plain}
+
+%% Workshop venue
+\acmConference[SAO @ CAIS 2026]{Workshop on Supporting Our AI Overlords}{May 26, 2026}{Berkeley, CA, USA}
+
+\usepackage{graphicx}
+
+\usepackage{listings}
+\lstset{
+  language=Python,
+  basicstyle=\ttfamily\scriptsize,
+  keywordstyle=\bfseries,
+  commentstyle=\itshape,
+  columns=flexible,
+  breaklines=true,
+  aboveskip=4pt,
+  belowskip=4pt,
+  xleftmargin=2pt,
+  frame=none,
+}
+
+\begin{document}
+
+\title{Metaxy: Field-Level Metadata Management\\for Incremental Multimodal ML Pipelines}
+
+\author{Daniel Gafni}
+\affiliation{\institution{Anam}\country{United Kingdom}}
+\email{danielgafni16@gmail.com}
+
+\author{Georg Heiler}
+\affiliation{\institution{Complexity Science Hub Vienna (CSH)}\country{Austria}}
+\affiliation{\institution{Austrian Supply Chain Intelligence Institute (ASCII)}\country{Austria}}
+\email{heiler@csh.ac.at}
+
+\begin{abstract}
+Multimodal ML pipelines incur substantial GPU costs, yet existing orchestrators invalidate entire downstream datasets when any processing step changes---even for unaffected fields.
+We present \textbf{Metaxy}, an open-source Python library providing \emph{field-level dependency tracking at record granularity}.
+Feature definitions form a directed acyclic graph; for each record, hierarchical version hashes propagate changes along field-level edges.
+A \texttt{resolve\_update} operation returns exactly those records that are new, stale, or orphaned, enabling selective recomputation.
+In production across two organizations processing millions of multimodal samples, Metaxy eliminates redundant GPU work while preserving complete per-record lineage for reproducibility.
+\end{abstract}
+
+\begin{CCSXML}
+<ccs2012>
+   <concept>
+       <concept_id>10002951.10003317.10003347.10003356</concept_id>
+       <concept_desc>Information systems~Data provenance</concept_desc>
+       <concept_significance>500</concept_significance>
+   </concept>
+   <concept>
+       <concept_id>10011007.10011006.10011039</concept_id>
+       <concept_desc>Software and its engineering~Software libraries and repositories</concept_desc>
+       <concept_significance>300</concept_significance>
+   </concept>
+</ccs2012>
+\end{CCSXML}
+
+\ccsdesc[500]{Information systems~Data provenance}
+\ccsdesc[300]{Software and its engineering~Software libraries and repositories}
+
+\keywords{metadata management, incremental processing, feature engineering, multimodal ML, data provenance, reproducibility}
+
+\maketitle
+
+\section{Introduction}
+
+Modern ML increasingly relies on multimodal data---video, audio, images, and text---processed through GPU-accelerated pipelines.
+Unlike tabular ETL, where re-execution is cheap, multimodal steps such as speech-to-text transcription or embedding generation cost orders of magnitude more per record.
+Data orchestrators like Dagster~\cite{dagster} model pipelines as graphs of \emph{assets} (tables or datasets) and track freshness at the asset level: when a processing step changes, every downstream record is recomputed---even those whose inputs are unchanged.
+
+Consider a video pipeline extracting both audio transcripts and face detections (Figure~\ref{fig:field-deps}).
[…]
