+++ publications/2026-metaxy-field-level-versioning/references.bib
@@ -1,0 +1,48 @@
+@misc{dagster,
+  author = {{Dagster Labs}},
+  title = {Dagster: Data Orchestration Platform},
+  year = {2025},
+  howpublished = {\url{https://dagster.io}}
+}
+
+@misc{ibis,
+  author = {{Ibis Project}},
+  title = {Ibis: Portable DataFrames for Different Backends},
+  year = {2025},
+  howpublished = {\url{https://ibis-project.org}}
+}
+
+@misc{narwhals,
+  author = {Marco Gorelli and contributors},
+  title = {Narwhals: Write Once, Run on Many DataFrames},
+  year = {2025},
+  howpublished = {\url{https://narwhals.readthedocs.io}}
+}
+
+@inproceedings{ray,
+  author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Paul, Melanie and Jordan, Michael I. and Stoica, Ion},
+  title = {Ray: A Distributed Framework for Emerging {AI} Applications},
+  booktitle = {Proceedings of the 13th USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
+  year = {2018}
+}
+
+@software{dvc,
+  author = {Kuprieiev, Ruslan and Petrov, Dmitry and others},
+  title = {{DVC}: Data Version Control},
+  year = {2025},
+  howpublished = {\url{https://dvc.org}}
+}
+
+@misc{datachain,
+  author = {{Iterative, Inc.}},
+  title = {DataChain: {AI} Data Warehouse for Multimodal Data},
+  year = {2025},
+  howpublished = {\url{https://datachain.ai}}
+}
+
+@misc{dagsterslurm,
+  author = {Heiler, Georg and Picatto, Hernan and He\ss{}, Maximilian and Pfister, Martin},
+  title = {dagster-slurm: Reproducible Data Orchestration on {HPC}},
+  year = {2025},
+  howpublished = {\url{https://github.com/ascii-supply-networks/dagster-slurm}}
+}
