### 2026-04-09 ddcf38a
feat(map datatype): use Arrow conversion utils from `polars-map` (#1175)
## Summary

Use the new [polars-map Arrow utilities](https://github.com/hafaio/polars-map/pull/6) when doing conversion between Polars and PyArrow.

This PR also bumps `narwhals-map `that has been [updated](https://github.com/danielgafni/narwhals-map/pull/3) to make use of the same conversion utilities.

The slight change is that Polars frames converted to Arrow tables using the new method may have `large_string` keys instead of `string`, but it shouldn't be important.

<!-- Brief description of the changes -->

## Change
[…]
