### 2026-04-09 4199b11
fix: conditionally import DeltaLake and Iceberg stores in `metaxy.ext.polars`  (#1180)
## Summary

This PR fixes import errors for `deltalake`/`pyiceberg` when trying to import the other store from `metaxy.ext.polars`. They should't have been placed there in the first place, but removing them right now would be a breaking change, hence the `__globals__` hack.

- Move `deltalake` import in `src/metaxy/ext/polars/handlers/delta.py` from module-level to `TYPE_CHECKING` + local runtime imports
- Guard Delta/Iceberg handler re-exports in `src/metaxy/ext/polars/__init__.py` behind `TYPE_CHECKING`
- 
[…]
