### 2026-04-09 59e94bd
chore(deltalake, iceberg): deprecate importing DeltaLake and Iceberg stores from `metaxy.ext.polars` (#1181)
## Summary

This PR deprecates importing DeltaLake and Iceberg stores from `metaxy.ext.polars`.

Instead, they should be imported from `metaxy.ext.polars.handlers.<store>`.

<!-- Brief description of the changes -->

## Changelog

- chore(deltalake): the `metaxy.ext.polars.DeltaMetadataStore` import path has been deprecated in favor of `metaxy.ext.polars.handlers.delta.DeltaMetadataStore`
- chore(iceberg): the `metaxy.ext.iceberg.IcebergMetadataStore` import path has been deprecated in 
[…]
