+++ .claude-plugin/skills/metaxy/examples/testing.md
@@ -37,7 +37,7 @@ import metaxy as mx
 def test_with_custom_config(tmp_path):
     with mx.MetaxyConfig(
         stores={"test": mx.StoreConfig(
-            type="metaxy.ext.polars.DeltaMetadataStore",
+            type="metaxy.ext.polars.handlers.delta.DeltaMetadataStore",
             config={"root_path": str(tmp_path / "delta_test")},
         )}
     ).use() as config:
@@ -57,7 +57,7 @@ def metaxy_env(tmp_path):
     with mx.FeatureGraph().use():
         with mx.MetaxyConfig(
             stores={"test": mx.StoreConfig(
-                type="metaxy.ext.polars.DeltaMetadataStore",
+                type="metaxy.ext.polars.handlers.delta.DeltaMetadataStore",
                 config={"root_path": str(tmp_path / "delta_test")},
             )}
         ).use() as config:

[…]
