+++ cliff.toml
@@ -19,6 +19,8 @@ commit_preprocessors = [
   { pattern = '^:robot: ', replace = "ci: " },
   { pattern = '^:bomb: ', replace = "refactor!: " },
   { pattern = '^fix: update devenv.lock', replace = "chore: update devenv.lock" },
+  # Strip markdown list markers from ## Changelog lines in commit bodies
+  { pattern = '(?m)^- ', replace = "" },
   # Strip PR references from messages (handled in template with author links)
   { pattern = ' \(#\d+\)', replace = "" },
 ]
@@ -104,16 +106,35 @@ body = """
 {%- endif -%}
 {%- endfor -%}
 {%- endif -%}
+{%- set_global has_scoped = false -%}
 {%- if changelog_lines | length > 0 -%}
-{%- for line in changelog_lines | trim | split(pat="\n") %}
+{%- set_global cl_scope = "" -%}
+{%- for line in changelog_lines | trim | split(pat="\n") -%}
+{%- if line is containing("(") and line is containing("): ") -%}
+{%- set line_parts = line | split(pat="(") -%}
+{%- set new_scope = line_parts.1 | split(pat=")") | first -%}
+{%- set cl_msg = line | replace(from=line_parts.0 ~ "(" ~ new_scope ~ "): ", to="") -%}
+{%- set_global has_scoped = true -%}
+{%- if new_scope != cl_scope %}
+
+#### {{ new_scope | upper_first }}
+{%- set_global cl_scope = new_scope -%}
+{%- endif %}
+- {{ cl_msg }}
+{%- if commit.remote.pr_number %} ([#{{ commit.remote.pr_number }}](https://github.com/anam-org/metaxy/pull/{{ commit.remote.pr_number }}){% if commit.remote.username %} by [@{{ commit.remote.username }}](https://github.com/{{ commit.remote.username }}){% endif %})
+{%- endif -%}
+{%- else %}
 - {{ line }}
+{%- endif -%}
 {%- endfor -%}
 {%- else %}
 - {{ commit.message | split(pat="\n") | first | trim }}
 {%- endif %}
+{%- if not has_scoped -%}
 {%- if commit.remote.pr_number %} ([#{{ commit.remote.pr_number }}](https://github.com/anam-org/metaxy/pull/{{ commit.remote.pr_number }}){% if commit.remote.username %} by [@{{ commit.remote.username }}](https://github.com/{{ commit.remote.username }}){% endif %})
 {%- elif commit.remote.username %} (by [@{{ commit.remote.username }}](https://github.com/{{ commit.remote.username }}))
-{%- endif %}
+{%- endif -%}
+{%- endif -%}
 {%- endfor %}
 {% endfor %}
 {% endfor %}\
@@ -139,6 +160,10 @@ postprocessors = [
   { pattern = 'referrring', replace = "referring" },
   # deltalake Map support was bundled with the iceberg PR; add a separate changelog section
   { pattern = '(#### Iceberg\n\n- `Map` type support)', replace = "#### Deltalake\n\n- `Map` type support (experimental) ([#1108](https://github.com/anam-org/metaxy/pull/1108) by [@danielgafni](https://github.com/danielgafni))\n\n$1" },
+  # Remove empty scope headers left when per-line scopes replace the commit scope
+  { pattern = '#### [^\n]+\n+\n(#### )', replace = "$1" },
+  # Strip git tag comment lines from tag messages
+  { pattern = '(?m)^#\n|^# Write a message[^\n]*\n?|^#   v[^\n]*\n?|^# Lines starting[^\n]*\n?', replace = "" },
 ]
 footer = """
 
