+++ tach.toml
@@ -17,7 +17,20 @@ layers = [
 
 [[modules]]
 path = "metaxy"
-depends_on = []
+depends_on = [
+  "metaxy.utils",
+  "metaxy._version",
+  "metaxy.models",
+  "metaxy._decorators",
+  "metaxy.config",
+  "metaxy.versioning",
+  "metaxy.entrypoints",
+  "metaxy._exceptions",
+  "metaxy._hashing",
+  "metaxy.metadata_store",
+  "metaxy.models.types",
+  "metaxy._warnings",
+]
 
 [[modules]]
 path = "metaxy._decorators"
@@ -25,11 +38,11 @@ depends_on = []
 
 [[modules]]
 path = "metaxy._exceptions"
-depends_on = []
+depends_on = ["metaxy.utils"]
 
 [[modules]]
 path = "metaxy._hashing"
-depends_on = []
+depends_on = ["metaxy.config", "metaxy._decorators"]
 
 [[modules]]
 path = "metaxy._packaging"
@@ -48,106 +61,146 @@ depends_on = []
 [[modules]]
 path = "metaxy.cli"
 layer = "cli"
-depends_on = []
+depends_on = ["metaxy", "metaxy._version"]
 
 # --- Ext layer ---
 
 [[modules]]
 path = "metaxy.ext.bigquery"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy.ext.ibis", "metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.clickhouse"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy._decorators", "metaxy.ext.ibis"]
 
 [[modules]]
 path = "metaxy.ext.dagster"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy._version", "metaxy._decorators", "metaxy"]
 
 [[modules]]
 path = "metaxy.ext.duckdb"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy.ext.ibis", "metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.ibis"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.lancedb"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.metadata_stores"
 layer = "ext"
-depends_on = []
+depends_on = [
+  "metaxy.ext.lancedb",
+  "metaxy.ext.polars",
+  "metaxy.ext.postgresql",
+  "metaxy.ext.duckdb",
+  "metaxy._warnings",
+  "metaxy.ext.bigquery",
+  "metaxy.ext.clickhouse",
+]
 
 [[modules]]
 path = "metaxy.ext.mcp"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy"]
 
 [[modules]]
 path = "metaxy.ext.polars"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy._warnings", "metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.postgresql"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy.ext.ibis", "metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.ray"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy"]
 
 [[modules]]
 path = "metaxy.ext.sqlalchemy"
 layer = "ext"
-depends_on = []
+depends_on = ["metaxy._decorators"]
 
 [[modules]]
 path = "metaxy.ext.sqlmodel"
 layer = "ext"
-depends_on = []
[…]
