+++ 0.1.10/slides/2026-introducing-metaxy/dist/assets/md-j1mY04cH.js
@@ -1,0 +1,2 @@
+import{_ as l}from"./slidev/CodeBlockWrapper.vue_vue_type_script_setup_true_lang-FpMJoXbT.js";import{b as d,o as m,w as a,g as e,D as t,e as f,m as p,v as u,x as c,z as r}from"./modules/vue-D2_2azlL.js";import{I as g}from"./slidev/default-DMaQI0-d.js";import{u as v,f as k}from"./slidev/context-CXJvKEzn.js";import"./modules/unplugin-icons-BaqmoMqW.js";import"./index-B2pMgQ-r.js";import"./modules/shiki-bZSEl0vS.js";import"./slidev/TitleIcon.vue_vue_type_script_setup_true_lang-CyxFj2c-.js";const _=""+new URL("../img/feature.svg",import.meta.url).href,x={"mt-4":""},V={__name:"introducing-metaxy.md__slidev_4",setup(h){const{$clicksContext:o,$frontmatter:i}=v();return o.setup(),(b,s)=>{const n=l;return m(),d(g,u(c(r(k)(r(i),3))),{default:a(()=>[s[1]||(s[1]=e("h1",null,"The Key Insight: Data Fields",-1)),s[2]||(s[2]=e("p",null,[t("Instead of one version per table, Metaxy tracks "),e("strong",null,"separate versions for each field of every record"),t(".")],-1)),s[3]||(s[3]=e("div",{class:"flex justify-center mt-4"},[e("img",{src:_,class:"h-72"})],-1)),e("div",x,[f(n,p({},{title:"",ranges:[]}),{default:a(()=>[...s[0]||(s[0]=[e("pre",{class:"shiki shiki-themes vitesse-dark vitesse-light slidev-code",style:{"--shiki-dark":"#dbd7caee","--shiki-light":"#393a34","--shiki-dark-bg":"#121212","--shiki-light-bg":"#ffffff"}},[e("code",{class:"language-text"},[e("span",{class:"line"},[e("span",null,'video_001: {"audio": "a7f3c2d8", "frames": "b9e1f4a2"}')]),t(`
+`),e("span",{class:"line"},[e("span",null,'video_002: {"audio": "c1d5e9f3", "frames": "f7a2b8c4"}')])])],-1)])]),_:1},16)]),s[4]||(s[4]=e("blockquote",null,[e("p",null,"Each sample carries a versioning dictionary, not a single version string.")],-1))]),_:1},16)}}};export{V as default};
