+++ 0.1.10/slides/2026-introducing-metaxy/dist/assets/modules/unplugin-icons-BaqmoMqW.js
@@ -1,0 +1,1 @@
+import{y as l,f as o,o as t,g as e}from"./vue-D2_2azlL.js";const r={class:"slidev-icon",viewBox:"0 0 32 32",width:"1.2em",height:"1.2em"};function i(a,n){return t(),o("svg",r,[...n[0]||(n[0]=[e("path",{fill:"currentColor","fill-rule":"evenodd",d:"M16 2a14 14 0 0 0-4.43 27.28c.7.13 1-.3 1-.67v-2.38c-3.89.84-4.71-1.88-4.71-1.88a3.7 3.7 0 0 0-1.62-2.05c-1.27-.86.1-.85.1-.85a2.94 2.94 0 0 1 2.14 1.45a3 3 0 0 0 4.08 1.16a2.93 2.93 0 0 1 .88-1.87c-3.1-.36-6.37-1.56-6.37-6.92a5.4 5.4 0 0 1 1.44-3.76a5 5 0 0 1 .14-3.7s1.17-.38 3.85 1.43a13.3 13.3 0 0 1 7 0c2.67-1.81 3.84-1.43 3.84-1.43a5 5 0 0 1 .14 3.7a5.4 5.4 0 0 1 1.44 3.76c0 5.38-3.27 6.56-6.39 6.91a3.33 3.33 0 0 1 .95 2.59v3.84c0 .46.25.81 1 .67A14 14 0 0 0 16 2"},null,-1)])])}const B=l({name:"carbon-logo-github",render:i}),s={class:"slidev-icon",viewBox:"0 0 32 32",width:"1.2em",height:"1.2em"};function c(a,n){return t(),o("svg",s,[...n[0]||(n[0]=[e("path",{fill:"currentColor",d:"M2 26h28v2H2zM25.4 9c.8-.8.8-2 0-2.8l-3.6-3.6c-.8-.8-2-.8-2.8 0l-15 15V24h6.4zm-5-5L24 7.6l-3 3L17.4 7zM6 22v-3.6l10-10l3.6 3.6l-10 10z"},null,-1)])])}const C=l({name:"carbon-edit",render:c}),d={class:"slidev-icon",viewBox:"0 0 32 32",width:"1.2em",height:"1.2em"};function h(a,n){return t(),o("svg",d,[...n[0]||(n[0]=[e("path",{fill:"currentColor",d:"m18 6l-1.43 1.393L24.15 15H4v2h20.15l-7.58 7.573L18 26l10-10z"},null,-1)])])}const b=l({name:"carbon-arrow-right",render:h}),u={class:"slidev-icon",viewBox:"0 0 256 256",width:"1.2em",height:"1.2em"};function _(a,n){return t(),o("svg",u,[...n[0]||(n[0]=[e("path",{fill:"currentColor",d:"M200 32h-36.26a47.92 47.92 0 0 0-71.48 0H56a16 16 0 0 0-16 16v168a16 16 0 0 0 16 16h144a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16m-72 0a32 32 0 0 1 32 32H96a32 32 0 0 1 32-32m72 184H56V48h26.75A47.9 47.9 0 0 0 80 64v8a8 8 0 0 0 8 8h80a8 8 0 0 0 8-8v-8a47.9 47.9 0 0 0-2.75-16H200Z"},null,-1)])])}const A=l({name:"ph-clipboard",render:_}),m={class:"slidev-icon",viewBox:"0 0 256 256",width:"1.2em",height:"1.2em"};function v(a,n){return t(),o("svg",m,[...n[0]||(n[0]=[e("path",{fill:"currentColor",d:"M173.66 98.34a8 8 0 0 1 0 11.32l-56 56a8 8 0 0 1-
[…]
