+++ 0.1.10/ruff.toml
@@ -1,0 +1,19 @@
+extend = "../ruff.toml"
+
+# docs often benefit from shorter lines
+line-length = 88
+
+[lint]
+ignore = [
+  "E402", # examples often have imports in the middle
+]
+
+[lint.flake8-import-conventions]
+# Declare the banned `from` imports.
+banned-from = ["dagster", "narwhals", "polars", "metaxy"]
+
+[lint.flake8-import-conventions.extend-aliases]
+"metaxy" = "mx"
+"dagster" = "dg"
+"polars" = "pl"
+"narwhals" = "nw"

[…]
