+++ tests/metadata_stores/core/test_resolve_update_multi_dep_samples.py
@@ -1,0 +1,255 @@
+"""Tests for sample-scoped resolve_update on multi-dependency features."""
+
+from __future__ import annotations
+
+from typing import Any
+
+import polars as pl
+from metaxy import FeatureDep, FeatureGraph
+from metaxy.metadata_store.base import MetadataStore
+from metaxy.models.constants import (
+    METAXY_DATA_VERSION_BY_FIELD,
+    METAXY_PROVENANCE_BY_FIELD,
+)
+from metaxy.models.field import FieldDep, FieldSpec
+from metaxy.models.types import FieldKey
+from metaxy_testing.models import SampleFeature, SampleFeatureSpec
+from pytest_cases import parametrize_with_cases
+
+from metaxy_testing import add_metaxy_provenance_column
+from tests.metadata_stores.conftest import BasicStoreCases
+
+
+def _provider_metadata(field_name: str, versions: dict[str, str]) -> pl.DataFrame:
+    sample_ids = list(versions)
+    field_versions = list(versions.values())
+    return pl.DataFrame(
+        {
+            "sample_uid": sample_ids,
+            METAXY_PROVENANCE_BY_FIELD: [{field_name: version} for version in field_versions],
+            METAXY_DATA_VERSION_BY_FIELD: [{field_name: version} for version in field_versions],
+        }
+    )
+
+
+def _write_provider(
+    store: MetadataStore,
+    feature: type[SampleFeature],
+    *,
+    field_name: str,
+    versions: dict[str, str],
+) -> None:
+    provider_df = _provider_metadata(field_name, versions)
+    store.write(feature, add_metaxy_provenance_column(provider_df, feature))
+
+
+def _create_multi_dep_features(
+    graph: FeatureGraph,
+    *,
+    suffix: str,
+) -> tuple[type[SampleFeature], type[SampleFeature], type[SampleFeature]]:
+    with graph.use():
+
+        class ProviderA(
+            SampleFeature,
+            spec=SampleFeatureSpec(
+                key=f"provider_a_{suffix}",
+                fields=[FieldSpec(key=FieldKey(["a_payload"]), code_version="1")],
+            ),
+        ):
+            pass
+
+        class ProviderB(
+            SampleFeature,
+            spec=SampleFeatureSpec(
+                key=f"provider_b_{suffix}",
+                fields=[FieldSpec(key=FieldKey(["b_payload"]), code_version="1")],
+            ),
+        ):
+            pass
+
+        class FusedFeature(
+            SampleFeature,
+            spec=SampleFeatureSpec(
+                key=f"fused_{suffix}",
+                deps=[
+                    FeatureDep(feature=ProviderA),
+                    FeatureDep(feature=ProviderB),
+                ],
+                fields=[
+                    FieldSpec(
+                        key=FieldKey(["fused_payload"]),
+                        code_version="1",
+                        deps=[
+                            FieldDep(feature=ProviderA.spec().key, fields=[FieldKey(["a_payload"])]),
+                            FieldDep(feature=ProviderB.spec().key, fields=[FieldKey(["b_payload"])]),
+                        ],
+                    )
+                ],
+            ),
+        ):
+            pass
+
+    return ProviderA, ProviderB, FusedFeature
+
+
+class TestResolveUpdateMultiDepSamples:
+    """Tests for non-root resolve_update calls that provide samples."""
+
+    @parametrize_with_cases("store_config", cases=BasicStoreCases)
+    def test_samples_detect_staleness_across_all_upstreams(
+        self,
+        store_config: tuple[type[MetadataStore], dict[str, Any]],
+    ) -> None:
+        """A scoped sample still detects changes in every declared dependency."""
+        store_type, config = store_config
[…]

[…]
