### 2026-04-19 #10141 @abhishektiwari
@aringadre76 @KaushalprajapatiKP @benjaminade would you be open to help us with review of this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Metaxy: Record-Level Feature Metadata Management for GPU-Accelerated ML Pipeline".

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines, https://joss.readthedocs.io/en/latest/review_criteria.html.

Thanks
Abhi
