### 2026-04-20 #10449 @editorialbot
[REVIEW]: Metaxy: Record-Level Feature Metadata Management for GPU-Accelerated ML Pipelines
**Submitting author:** @geoheil (<a href="http://orcid.org/0000-0002-8684-1163">Georg Heiler</a>)
**Repository:** https://github.com/anam-org/metaxy/
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1
