### 2026-04-21 anam-org/metaxy#1191 @thc1006
JOSS review: paper presentation and metadata
Opened as part of the JOSS review at openjournals/joss-reviews#10449. Grouping smaller paper text and metadata items here so the review thread stays clean. Separate issues will follow for state of the field, claim substantiation, and references.

Line numbers below refer to `publications/2026-introducing-metaxy/paper.md` on main.

### Small fixes

- **L13 first author ORCID.** Daniel's author block has no `orcid:` field while Georg's does (L16). JOSS recommends ORCIDs for all authors, so worth adding if available.
- **L129 grammar.** "Large parts of the documentation *was* written by hand" should be `were written` (plural subject).
- **L98 capitalization.** "Postgresql" → **PostgreSQL**, "Ducklake" → **DuckLake** (matching how those projects spell themselves).
- **L133 acknowledgements.** The paper thanks Ibis, Narwhals, Pydantic, and Dagster communities. Ray is cited at L42 and L80, and there's a real `metaxy/ext/ray` integration, so adding Ray would keep this consistent.

### Hamilton reference

The `@hamilton` bib entry points at `hamilton.dagworks.io`. The project has since moved to the Apache Software Foundation and the canonical home is now `hamilton.apache.org` (the old URL still redirects). Suggest updating the URL and possibly referring to it as "Apache Hamilton" at first mention in L76.

Hamilton also has a SciPy 2024 proceedings paper with a DOI (`10.25080/gerudo-f2bc6f59-014`) that could replace the URL-only citation if preferred.

### Title and tags framing (soft suggestion)

More a question than a fix. The paper's title is "GPU-Accelerated ML Pipelines" and the tags include `GPU` and `AI`, but the README description, `CLAUDE.md`, and the GitHub topics all describe Metaxy around *multimodal incremental pipelines*. Metaxy itself doesn't do GPU work; the GPU part is downstream.

Something like "for Multimodal ML Pipelines" or "for Compute-Intensive ML Pipelines" would match the software's actual scope more closely. Same thought for the `GPU` / `AI` tags: `data lineage` / `incremental computation` / `caching` are already repo topics and might be better descriptors.

Not a blocker, just flagging for consideration.
