### 2026-04-21 anam-org/metaxy#1192 @thc1006
JOSS review: claim substantiation and software design details
Follow-up to #1191, covering the more substantive items. A third issue on state of the field and references will follow. All line numbers refer to `publications/2026-introducing-metaxy/paper.md` on main. Treating all of these as suggestions rather than blockers, happy to discuss any of them.

### Quantitative claims

**L33**: "GPUs that cost 10 to 100 times more per hour than standard processors." The range seems reasonable overall, though the ratio varies significantly with instance type. A short footnote anchoring it (e.g. "AWS on-demand pricing, YYYY-MM") would help readers calibrate. For reference, current AWS on-demand pricing lands roughly between 12x (t3.medium vs g4dn.xlarge) and ~290x (c6i.2xlarge vs p5.48xlarge), so the 10-100x framing holds for many common configurations.

**L114**: "Metaxy has been running in production at Anam for processing millions of training samples across multimodal video pipelines." This is the main anchor for Research Impact (evidence of realized impact), so it would be worth making the claim as specific as is comfortable to share publicly. Some possible directions:

- Approximate scale, even as order of magnitude rather than "millions"
- Duration in production
- One headline effect attributable to Metaxy (skip rate on edits, GPU hours saved, or rough cost comparison)

Completely understand if some of these aren't shareable. Even a single concrete number would help ("at Anam, Metaxy skips recomputation for roughly X% of records on a typical edit").

### Software design

**L88-91 hash function.** The paper uses "hierarchical hashing" and "deterministic hashing" without naming the function. Glancing at the shipped code, `hashlib.sha256` appears to be the default (used in `metaxy/models/{feature, feature_spec, feature_definition}.py`), with `XXHASH64` available as an alternative in `_hashing.py`. A one-sentence note on the default plus the configurable option would help readers reason about collision properties at the record-level scales the paper targets.

**L89 user-specified code versions.** "Field code versions are user-specified strings that mark algorithmic changes" is a meaningful design choice that affects correctness: if a developer edits an algorithm but forgets to bump the version string, Metaxy cannot detect the change, and downstream records keep stale versions. A brief acknowledgement of this limitation, plus a pointer to any safeguards the project provides (CI checks, pre-commit hooks, documentation guidance), would set reader expectations nicely.

**L115 "append-only lineage design" placement.** The phrase appears once in Research Impact, but append-only is a fairly foundational architectural decision. It might read more naturally if introduced in Software Design, where its downstream effects in Research Impact could then build on it.

### Citation policy (optional)

Small consistency observation: L41-42 cites integration libraries (Ibis, Narwhals, Dagster, Ray), while the storage backends at L98 (DuckDB, ClickHouse, BigQuery, Delta Lake, DuckLake, LanceDB, PostgreSQL) currently go uncited. Either direction works. Glad to share suggested references for the backends if that would help, or a light rephrase at L41-42 could also resolve the asymmetry.

Happy to iterate on any of these if it helps.
