### 2026-04-21 anam-org/metaxy#1193 @thc1006
JOSS review: reference archive DOIs
Third and final in the JOSS review series (#1191 and #1192 cover the earlier items). This one is intentionally short: two small reference-archive notes for `paper.bib` if you'd like to upgrade from URL-only citations. Both are entirely optional, no pressure either way.

### Ray (`@ray`)

The OSDI 2018 Ray paper doesn't have a Crossref DOI (USENIX proceedings don't mint them), so the commonly-cited canonical identifier is the arXiv preprint: **arXiv:1712.05889**. If formalizing the citation sounds useful, an `eprint = {1712.05889}` field or a link to the arXiv record would do it.

### Pydantic (`@pydantic`)

Pydantic has a Zenodo archive by Samuel Colvin et al. with DOI **10.5281/zenodo.19665399** that could work as a drop-in for the URL-only citation if a formal DOI is preferred. Fully optional.

(Hamilton's SciPy 2024 DOI was mentioned in #1191 alongside the Apache rename, so not repeated here.)

That wraps up the review issues from me. Thanks for bearing with the three-issue split, and looking forward to reading the next revision whenever it's ready.
