### 2026-04-21 #10449 @thc1006
First pass of review is up. Filed the following on the target repo to keep the thread clean:

- anam-org/metaxy#1191: paper presentation and metadata (small fixes + soft title/tags suggestion)
- anam-org/metaxy#1192: claim substantiation and software design details
- anam-org/metaxy#1193: reference archive DOIs (optional)

Will check off the corresponding checklist items once the author has had a chance to respond. cc @abhishektiwari
