### 2026-04-21 anam-org/metaxy#1191 @danielgafni
Thanks for the feedback, I think everything is spot-on, except this one:

> L133 acknowledgements. The paper thanks Ibis, Narwhals, Pydantic, and Dagster communities. Ray is cited at L42 and L80, and there's a real metaxy/ext/ray integration, so adding Ray would keep this consistent.

Metaxy isn't dependent on Ray, it just can be used with Ray. That's why it's not acknowledged directly.

On the other hand, Narwhals and Ibis were used to build the core versioning engine in Metaxy. Metaxy probably couldn't exist without them.
