### 2026-04-21 anam-org/metaxy#1192 @danielgafni
Thanks for the feedback!

> One headline effect attributable to Metaxy (skip rate on edits, GPU hours saved, or rough cost comparison)

This one would be pretty hard to estimate, I am not sure a metric like this would make sense. It's not like we would be doing a lot of extra compute without Metaxy (probably a little here and there?); we'd just had to go through a lot of extra hops, manual metadata edits, hacky overrides, and so on, to achieve the same effect. Metaxy automates and standardizes these decision, that's the main benefit here.

> Glancing at the shipped code, hashlib.sha256 appears to be the default (used in metaxy/models/{feature, feature_spec, feature_definition}.py), with XXHASH64 available as an alternative in _hashing.py.

These hash functions are used for different purposes: `sha256` is only used in the high level **Python computations** that run at the **graph level**: e.g. `"my_field": "v1"` is turned into `njkasdniojb12hiuysadbij`. This part of Metaxy is not configurable and the exact hashing function doesn't really matter.

On the other hand, `xxhash32` and other hash functions are applied at **sample level** - so for every row in the dataset. It typically runs in the metadata store (database) and therefore must be supported there. That's why these are configurable and default to `xxhash32` (if supported) to optimize for execution speed, but users should be able to use any hash function supported by their metadata store backend.

>  if a developer edits an algorithm but forgets to bump the version string, Metaxy cannot detect the change, and downstream records keep stale versions. A brief acknowledgement of this limitation, plus a pointer to any safeguards the project provides (CI checks, pre-commit hooks, documentation guidance), would set reader expectations nicely.

This isn't a Metaxy limitation, it is impossible to detect such changes in principle. It's impossible to distinguish e.g. code refactors from actual algorithmic changes (false-positives lead to recomputation and are too expensive), so that's not something we are even attempting to do. The user is responsible for specifying the right code version for each field, and that's it. We can emphasize on this in the paper.

> Small consistency observation: L41-42 cites integration libraries (Ibis, Narwhals, Dagster, Ray), while the storage backends at L98 (DuckDB, ClickHouse, BigQuery, Delta Lake, DuckLake, LanceDB, PostgreSQL) currently go uncited.

Honestly, I'd only keep citations for Narwhals and Ibis as direct building blocks for Metaxy, and Dagster as source of inspiration and reference for some design choices made in Metaxy (e.g. the `FeatureSpec` API). Do you think integrations also deserve a citation? I'm not sure about it since they are just an optional target for some glue code on our side.
