### 2026-04-21 anam-org/metaxy#1192 @thc1006
Thanks Daniel, this helps clarify a lot. Quick follow-ups:

### L114 / production impact

That framing of the value prop (automation and standardization of metadata work, rather than raw compute savings) actually fits the paper's claims better than the current L115 wording ("eliminates redundant computation, making expensive GPU workloads financially viable"). Something along the lines of "before Metaxy, this kind of selective recomputation required manual metadata edits, hacky overrides, and custom bookkeeping; Metaxy standardizes this into a declarative model" might land closer to the real value.

### Hash function

Thanks for the correction (I'd only spotted `XXHASH64` in the enum and missed that xxhash32 is the actual default on DuckDB). The two-level split (graph-level `sha256` fixed, sample-level pluggable with backend-appropriate defaults) is a useful design detail that would be worth stating in Software Design. Readers currently can't tell there are two hashing layers.

### Code version

Agree it's design-by-choice, not a limitation. Making the principle explicit (user-controlled correctness over false-positive recomputation) in the paper would land well.

### Citations

Just to make sure I understand your proposal: Narwhals, Ibis, Pydantic (L57), and Dagster stay cited; Ray and the storage backends stay uncited. That scope works for me. One mechanical note: `@ray` currently appears at both L42 ("Dagster (@dagster) and Ray (@ray) consume...") and L80 ("such as Ray (@ray) or Dagster (@dagster)..."), so both sentences would need a small rephrase.

### Two items left open

You didn't touch on the L33 "10 to 100 times" citation ask, or the append-only placement note (which I mistakenly labeled L115 in the issue, it's actually L116). Both are fine to handle during the revision itself rather than here in the thread, just flagging they're still open.

Happy to iterate on any of these.
