+++ publications/2026-introducing-metaxy/paper.bib
@@ -12,11 +12,12 @@
   howpublished = {\url{https://narwhals.readthedocs.io}}
 }
 
-@misc{pydantic,
+@software{pydantic,
   author = {Samuel Colvin and Pydantic contributors},
-  title = {Pydantic},
-  year = {2025},
-  howpublished = {\url{https://pydantic.dev}}
+  title  = {Pydantic},
+  year   = {2025},
+  doi    = {10.5281/zenodo.19665399},
+  url    = {https://pydantic.dev}
 }
 
 @misc{dagster,
@@ -47,11 +48,13 @@
   howpublished = {\url{https://feast.dev}}
 }
 
-@misc{hamilton,
-  author = {Krawczyk, Stefan and Gilani, Elijah},
-  title = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
-  year = {2025},
-  howpublished = {\url{https://hamilton.dagworks.io}}
+@inproceedings{hamilton,
+  author    = {Krawczyk, Stefan and Gilani, Elijah},
+  title     = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
+  booktitle = {Proceedings of the 23rd Python in Science Conference (SciPy 2024)},
+  year      = {2024},
+  doi       = {10.25080/gerudo-f2bc6f59-014},
+  url       = {https://hamilton.apache.org}
 }
 
 @misc{datachain,
