+++ publications/2026-introducing-metaxy/paper.md
@@ -1,13 +1,14 @@
 ---
-title: 'Metaxy: Record-Level Feature Metadata Management for GPU-Accelerated ML Pipelines'
+title: 'Metaxy: Record-Level Feature Metadata Management for Multimodal ML Pipelines'
 tags:
   - Python
   - machine learning
   - metadata
   - feature engineering
   - reproducibility
-  - GPU
-  - AI
+  - data lineage
+  - incremental computation
+  - caching
   - Multimodal
 authors:
   - name: Daniel Gafni
@@ -30,8 +31,10 @@ bibliography: paper.bib
 
 Software that processes large datasets often repeats expensive computations when any part of the input data or processing logic changes.
 Metaxy is about perfecting the art of doing nothing: only compute what changed, save time and money, and accelerate exploration.
-In machine learning pipelines that handle video, audio, and images, these computations run on Graphics Processing Units (GPUs) that cost 10 to 100 times more per hour than standard processors.
+In machine learning pipelines that handle video, audio, and images, these computations run on Graphics Processing Units (GPUs) that cost 10 to 100 times more per hour than standard processors[^gpu-cost].
 A small change to one processing step can trigger unnecessary recomputation of unrelated steps, wasting both time and money.
+
+[^gpu-cost]: AWS on-demand pricing, 2026; the exact ratio depends on the instance families compared, spanning roughly 12× (general-purpose vs entry GPU) to several hundred-fold (commodity CPU vs top-end accelerator).
 
 Metaxy is a Python library that tracks which specific data records need reprocessing after a change, rather than rerunning entire datasets.
 It builds a dependency graph that connects individual data fields across processing steps.
@@ -39,7 +42,7 @@ When a researcher modifies one step, Metaxy identifies exactly which records are
 This selective approach eliminates redundant GPU work while preserving complete lineage for reproducibility.
 
 The library integrates with various backends through Ibis (@ibis) and supports multiple dataframe engines via Narwhals (@narwhals).
-Orchestration platforms such as Dagster (@dagster) and Ray (@ray) consume Metaxy's record-level diffs to schedule only the necessary GPU workloads.
+Orchestration platforms such as Dagster (@dagster) and Ray consume Metaxy's record-level diffs to schedule only the necessary GPU workloads.
 
 # Statement of Need
 
@@ -73,11 +76,11 @@ The target audience includes ML engineers working with multimodal data, MLOps te
 Several tools address aspects of feature management, but none, to our knowledge, provide field-level dependency tracking at record granularity as a standalone metadata layer.
 DVC (@dvc) versions datasets at the file level, treating each file as an opaque artifact without tracking individual records or fields within it.
 Feast (@feast) focuses on feature definition, materialization, and online serving. Recent Feast releases include DAG-based feature computation, but Feast does not expose field-level provenance for propagating version changes and deciding which downstream records require recomputation after an upstream modification.
-Hamilton (@hamilton) builds dataflows from Python functions and can report lineage over the resulting DAG, typically at the level of nodes, columns, and dataframe outputs. It does not maintain per-record, per-field provenance for selective downstream invalidation.
+Apache Hamilton (@hamilton) builds dataflows from Python functions and can report lineage over the resulting DAG, typically at the level of nodes, columns, and dataframe outputs. It does not maintain per-record, per-field provenance for selective downstream invalidation.
 DataChain (@datachain) combines metadata management with a Python-native data processing framework and supports delta processing over new or changed records. However, its incremental model is tied to dataset processing within that framework rather than to a standalone field-level dependency graph that can drive recomputation across multiple compute backends.
 
 Metaxy fills the gap by separating metadata from compute: it tracks field-level dependencies at record granularity and propagates version changes topologically through the dependency graph.
[…]
