### 2026-04-21 anam-org/metaxy#1193 @geoHeil
check https://github.com/anam-org/metaxy/pull/1194 should be tackled now
