+++ docs/publications/2026-introducing-metaxy/paper.bib
@@ -1,62 +1,0 @@
-@misc{ibis,
-  author = {{Ibis Project}},
-  title = {Ibis: Portable DataFrames for Different Backends},
-  year = {2025},
-  howpublished = {\url{https://ibis-project.org}}
-}
-
-@misc{narwhals,
-  author = {Marco Gorelli and contributors},
-  title = {Narwhals: Write Once, Run on Many DataFrames},
-  year = {2025},
-  howpublished = {\url{https://narwhals.readthedocs.io}}
-}
-
-@misc{pydantic,
-  author = {Samuel Colvin and Pydantic contributors},
-  title = {Pydantic},
-  year = {2025},
-  howpublished = {\url{https://pydantic.dev}}
-}
-
-@misc{dagster,
-  author = {{Dagster Labs}},
-  title = {Dagster: Data Orchestration Platform},
-  year = {2025},
-  howpublished = {\url{https://dagster.io}}
-}
-
-@inproceedings{ray,
-  author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Paul, Melanie and Jordan, Michael I. and Stoica, Ion},
-  title = {Ray: A Distributed Framework for Emerging AI Applications},
-  booktitle = {Proceedings of the 13th USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
-  year = {2018}
-}
-
-@software{dvc,
-  author = {Kuprieiev, Ruslan and Petrov, Dmitry and others},
-  title = {DVC: Data Version Control},
-  year = {2025},
-  howpublished = {\url{https://dvc.org}}
-}
-
-@misc{feast,
-  author = {{Feast Community}},
-  title = {Feast: Feature Store for Machine Learning},
-  year = {2025},
-  howpublished = {\url{https://feast.dev}}
-}
-
-@misc{hamilton,
-  author = {Krawczyk, Stefan and Gilani, Elijah},
-  title = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
-  year = {2025},
-  howpublished = {\url{https://hamilton.dagworks.io}}
-}
-
-@misc{datachain,
-  author = {{Iterative, Inc.}},
-  title = {DataChain: AI Data Warehouse for Multimodal Data},
-  year = {2025},
-  howpublished = {\url{https://datachain.ai}}
-}
