+++ publications/2026-introducing-metaxy/paper.bib
@@ -31,7 +31,11 @@
   author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Paul, Melanie and Jordan, Michael I. and Stoica, Ion},
   title = {Ray: A Distributed Framework for Emerging AI Applications},
   booktitle = {Proceedings of the 13th USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
-  year = {2018}
+  year = {2018},
+  eprint = {1712.05889},
+  archivePrefix = {arXiv},
+  primaryClass = {cs.DC},
+  url = {https://arxiv.org/abs/1712.05889}
 }
 
 @software{dvc,
