+++ main/publications/2026-introducing-metaxy/paper.bib
@@ -12,11 +12,12 @@
   howpublished = {\url{https://narwhals.readthedocs.io}}
 }
 
-@misc{pydantic,
+@software{pydantic,
   author = {Samuel Colvin and Pydantic contributors},
-  title = {Pydantic},
-  year = {2025},
-  howpublished = {\url{https://pydantic.dev}}
+  title  = {Pydantic},
+  year   = {2025},
+  doi    = {10.5281/zenodo.19665399},
+  url    = {https://pydantic.dev}
 }
 
 @misc{dagster,
@@ -30,7 +31,11 @@
   author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Paul, Melanie and Jordan, Michael I. and Stoica, Ion},
   title = {Ray: A Distributed Framework for Emerging AI Applications},
   booktitle = {Proceedings of the 13th USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
-  year = {2018}
+  year = {2018},
+  eprint = {1712.05889},
+  archivePrefix = {arXiv},
+  primaryClass = {cs.DC},
+  url = {https://arxiv.org/abs/1712.05889}
 }
 
 @software{dvc,
@@ -47,11 +52,13 @@
   howpublished = {\url{https://feast.dev}}
 }
 
-@misc{hamilton,
-  author = {Krawczyk, Stefan and Gilani, Elijah},
-  title = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
-  year = {2025},
-  howpublished = {\url{https://hamilton.dagworks.io}}
+@inproceedings{hamilton,
+  author    = {Krawczyk, Stefan and Gilani, Elijah},
+  title     = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
+  booktitle = {Proceedings of the 23rd Python in Science Conference (SciPy 2024)},
+  year      = {2024},
+  doi       = {10.25080/gerudo-f2bc6f59-014},
+  url       = {https://hamilton.apache.org}
 }
 
 @misc{datachain,
