+++ publications/2026-introducing-metaxy/paper.md
@@ -12,6 +12,7 @@ tags:
   - multimodal
 authors:
   - name: Daniel Gafni
+    orcid: 0000-0003-1237-6876
     affiliation: "1"
   - name: Georg Heiler
     orcid: 0000-0002-8684-1163
