+++ publications/2026-scipy-virtual-poster-session/poster.tex
@@ -1,0 +1,362 @@
+% Metaxy — SciPy 2026 Virtual Poster (LaTeX / tikzposter)
+% A0 landscape. Build:  make   (or: tectonic poster.tex after rendering SVGs)
+
+\documentclass[a0paper, landscape, 25pt, margin=0mm, innermargin=7mm,
+  blockverticalspace=2mm, colspace=10mm, subcolspace=6mm]{tikzposter}
+
+% ─── packages ─────────────────────────────────────────────────
+\usepackage{iftex}
+\ifPDFTeX
+  \usepackage[utf8]{inputenc}
+  \usepackage[T1]{fontenc}
+\fi
+\usepackage{graphicx}
+\usepackage{amsmath}
+\usepackage{xcolor}
+\usepackage{listings}
+\usepackage{url}
+
+% Font stack — engine-detecting packages, no [default] option to avoid
+% fontspec collisions under pdflatex.
+\usepackage{ebgaramond}
+\usepackage{sourcesanspro}
+\usepackage{sourcecodepro}
+\renewcommand{\familydefault}{\sfdefault}
+% Use the right family-selection syntax per engine. Under pdflatex the
+% NFSS name is "EBGaramond-LF"; under xelatex/luatex fontspec registers
+% EB Garamond as the default \rmfamily via the ebgaramond package.
+\ifPDFTeX
+  \newcommand{\garamond}{\fontfamily{EBGaramond-LF}\selectfont}
+\else
+  \newcommand{\garamond}{\rmfamily}
+\fi
+
+% Pre-rendered SVG → PDF (see Makefile). Tectonic does not support the
+% svg package's \write18 pipe, so rendering runs ahead of the build.
+\newcommand{\svgimg}[2][]{\includegraphics[#1]{#2.svg.pdf}}
+
+% Stub \href so tikzposter + TL2026 pdflatex (which chokes on hyperref's
+% \AtBeginDocument hook) can still compile the same source.
+\providecommand{\href}[2]{#2}
+
+\graphicspath{{assets/}}
+
+% Turn off the LaTeX/TikZposter watermark on the bottom-right of the poster.
+\tikzposterlatexaffectionproofoff
+
+% ─── design tokens ────────────────────────────────────────────
+\definecolor{navy}{HTML}{08111D}
+\definecolor{deep}{HTML}{0F172A}
+\definecolor{cream}{HTML}{F6F4EF}
+\definecolor{paper}{HTML}{FFFFFF}
+\definecolor{teal}{HTML}{0F766E}
+\definecolor{tealdark}{HTML}{14B8A6}
+\definecolor{tealsoft}{HTML}{F0FDF9}
+\definecolor{ink}{HTML}{0A0A0A}
+\definecolor{ink60}{HTML}{404040}
+\definecolor{ink40}{HTML}{737373}
+\definecolor{rulecolor}{HTML}{D4D4D4}
+
+% Python syntax highlighting palette (readable on the dark code background).
+\definecolor{codekw}{HTML}{5EEAD4}
+\definecolor{codestr}{HTML}{FCD34D}
+\definecolor{codecom}{HTML}{94A3B8}
+\definecolor{codenum}{HTML}{F0ABFC}
+\definecolor{codeid}{HTML}{E2E8F0}
+
+% ─── tikzposter styles ────────────────────────────────────────
+\definebackgroundstyle{metaxyBg}{
+  \fill[color=cream] (bottomleft) rectangle (topright);
+}
+\usebackgroundstyle{metaxyBg}
+
+\definetitlestyle{metaxyTitle}{
+  width=\paperwidth, roundedcorners=0, linewidth=0pt, innersep=3pt,
[…]
