+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ repos:
         exclude: '^uv\.lock$'
       - id: check-yaml
       - id: check-added-large-files
-        exclude: '^uv\.lock|examples/example-dagster/uv\.lock$'
+        exclude: '^(uv\.lock|examples/example-dagster/uv\.lock|docs/assets/slides/2026-introducing-metaxy/georg\.jpg)$'
 
   - repo: local
     hooks:
