+++ main/slides/2026-introducing-metaxy/dist/assets/metaxy-DEkS6NCl.js
@@ -1,0 +1,1 @@
+const s="/slides/2026-introducing-metaxy/dist/img/metaxy.svg";export{s as _};
