+++ main/slides/2026-introducing-metaxy/dist/assets/md-DKEdWZXc.js
@@ -1,0 +1,4 @@
+import{b as l,o as n,w as i,i as o,g as e,D as d,v as c,x as m,A as a}from"./modules/vue-BtB2w7oG.js";import{_ as p}from"./white-PAwb7_2b.js";import{u,f as g}from"./slidev/context-xkorsVSV.js";import"./index-DAQQqyPx.js";import"./modules/shiki-qexTPG9f.js";const x="/slides/2026-introducing-metaxy/dist/img/bauplan-scenario.jpg",k={__name:"introducing-metaxy.md__slidev_3",setup(f){const{$clicksContext:r,$frontmatter:s}=u();return r.setup(),(v,t)=>(n(),l(p,c(m(a(g)(a(s),2))),{default:i(()=>[o(` ──────────────────────────────────────────────────────
+     SLIDE 3 · WHY NOW  (white · compute has changed)
+     Set up the GPU economics problem with the bigger picture.
+────────────────────────────────────────────────────── `),t[0]||(t[0]=e("div",{class:"h-full max-w-6xl mx-auto px-16 py-6 flex flex-col justify-center gap-4"},[e("div",{class:"max-w-5xl space-y-2"},[e("div",{class:"eyebrow"},"Why Now"),e("h1",{class:"slide-heading",style:{"font-size":"2.14rem"}},"Modern data workflows branch and iterate"),e("p",{class:"lead max-w-5xl",style:{"font-size":"0.96rem","line-height":"1.4"}},"LLM calls and GPU-heavy stages make coarse reruns expensive. Teams increasingly need to compare alternative paths without recomputing everything.")]),e("div",{class:"grid gap-4 items-start",style:{"grid-template-columns":"0.94fr 1.06fr"}},[e("div",{class:"space-y-0"},[e("div",{class:"grid gap-4 border-t border-neutral-300 py-2.5",style:{"grid-template-columns":"110px 1fr"}},[e("div",{class:"text-sm font-semibold uppercase tracking-[0.22em] text-teal-700"},"Cost"),e("div",{class:"text-sm leading-relaxed text-neutral-700"},"LLM calls and GPU hours make full reruns expensive.")]),e("div",{class:"grid gap-4 border-t border-neutral-300 py-2.5",style:{"grid-template-columns":"110px 1fr"}},[e("div",{class:"text-sm font-semibold uppercase tracking-[0.22em] text-teal-700"},"Legacy"),e("div",{class:"text-sm leading-relaxed text-neutral-700"},"Many data platforms assume one shared state and coarse invalidation.")]),e("div",{class:"grid gap-4 border-t border-neutral-300 py-2.5",style:{"grid-template-columns":"110px 1fr"}},[e("div",{class:"text-sm font-semibold uppercase tracking-[0.22em] text-teal-700"},"Agents"),e("div",{class:"text-sm leading-relaxed text-neutral-700"},"Exploratory and agent-driven workflows branch, compare alternatives, and merge after review.")]),e("div",{class:"grid gap-4 border-t border-neutral-300 py-2.5",style:{"grid-template-columns":"110px 1fr"}},[e("div",{class:"text-sm font-semibold uppercase tracking-[0.22em] text-teal-700"},"Need"),e("div",{class:"text-sm leading-relaxed text-neutral-700"},"Provenance, incremental recompute, and branch-aware state.")])]),e("div",{class:"rounded-lg border border-neutral-200 bg-neutral-50 p-3"},[e("img",{src:x,alt:"Bauplan branching scenario",class:"w-full rounded-md border border-neutral-200 object-contain"}),e("p",{class:"mt-2 text-xs leading-snug text-neutral-600"},[d(" Bauplan: branching analytics can fork shared state in
[…]
