+++ main/slides/2026-introducing-metaxy/dist/assets/slidev/404-BCHbFmur.js
@@ -1,1 +1,0 @@
-import{d as f,a3 as m,B as _,f as x,g as o,D as a,t as r,z as v,b as u,i,w as c,a as g,o as n}from"../modules/vue-D2_2azlL.js";import{n as k,_ as h}from"../index-B2pMgQ-r.js";import"../modules/shiki-bZSEl0vS.js";const B={class:"grid justify-center text-center pt-15% gap-5"},N={class:"text-2xl"},w={class:"op-60"},y={class:"mt-3 flex flex-col gap-2 max-w-xs mx-auto w-full"},C=f({__name:"404",setup(R){const{currentRoute:l}=m(),{total:d}=k(),s=_(()=>{const t=l.value.path.match(/\d+/);if(t){const e=+t[0];if(e>0&&e<=d.value)return e}return null});return(p,t)=>{const e=g("RouterLink");return n(),x("div",B,[o("div",null,[t[2]||(t[2]=o("h1",{class:"text-9xl font-light"}," 404 ",-1)),o("p",N,[t[0]||(t[0]=a(" Page ",-1)),o("code",w,r(v(l).path),1),t[1]||(t[1]=a(" not found ",-1))])]),o("div",y,[s.value!==1?(n(),u(e,{key:0,to:"/",class:"page-link"},{default:c(()=>[...t[3]||(t[3]=[a(" Go Home ",-1)])]),_:1})):i("v-if",!0),s.value?(n(),u(e,{key:1,to:`/${s.value}`,class:"page-link"},{default:c(()=>[a(" Go to Slide "+r(s.value),1)]),_:1},8,["to"])):i("v-if",!0)])])}}}),G=h(C,[["__scopeId","data-v-2af184e6"]]);export{G as default};

[…]
