+++ publications/2026-scipy-virtual-poster-session/poster.tex
@@ -175,8 +175,9 @@
     {\ttfamily\color{tealdark}\fontsize{30}{34}\selectfont%
     docs.metaxy.io \quad\textbullet\quad github.com/anam-org/metaxy \quad\textbullet\quad uv add metaxy}\\[1mm]%
     {\sffamily\color{white!85!black}\fontsize{30}{34}\selectfont%
-    \href{https://gafni.dev/}{Daniel Gafni}\, \textbullet\, \href{https://georgheiler.com/}{Georg Heiler}%
-    \quad\color{white!65!black}\normalsize ex-ML engineer at Anam\, \textbullet\, Jubust / ASCII / CSH}%
+    \href{https://gafni.dev/}{Daniel Gafni}\, \textbullet\, \href{https://georgheiler.com/}{Georg Heiler}}\\[0.5mm]%
+    {\sffamily\color{white!65!black}\fontsize{24}{28}\selectfont%
+    Daniel: ex-ML engineer at Anam \quad\textbullet\quad Georg: Jubust / ASCII / CSH}%
   \end{minipage}\hfill%
   \begin{minipage}[c]{0.18\textwidth}\raggedleft%
     \svgimg[height=13cm]{metaxy}%
