+++ main/slides/2026-introducing-metaxy/dist/assets/slidev/entry-B0cKiJ3O.js
@@ -1,0 +1,1 @@
+import{f as n,e as t,w as o,a as r,o as i,g as l,D as a}from"../modules/vue-BtB2w7oG.js";import{_ as d}from"../index-CsGYBKJ1.js";import"../modules/shiki-qexTPG9f.js";const c={},f={class:"h-full w-full flex items-center justify-center gap-5 lt-md:flex-col"};function p(u,e){const s=r("RouterLink");return i(),n("div",f,[t(s,{to:"/",class:"page-link"},{default:o(()=>[...e[0]||(e[0]=[l("div",{class:"i-carbon:presentation-file"},null,-1),a(" Slides ",-1)])]),_:1}),t(s,{to:"/presenter",class:"page-link"},{default:o(()=>[...e[1]||(e[1]=[l("div",{class:"i-carbon:user-speaker"},null,-1),a(" Presenter ",-1)])]),_:1}),t(s,{to:"/notes",class:"page-link"},{default:o(()=>[...e[2]||(e[2]=[l("div",{class:"i-carbon:catalog"},null,-1),a(" Notes ",-1)])]),_:1}),t(s,{to:"/overview",class:"page-link"},{default:o(()=>[...e[3]||(e[3]=[l("div",{class:"i-carbon:list-boxes"},null,-1),a(" Overview ",-1)])]),_:1})])}const k=d(c,[["render",p],["__scopeId","data-v-d5062f01"]]);export{k as default};
