+++ main/slides/2026-introducing-metaxy/dist/assets/metaxy-CcMTLh2-.js
@@ -1,0 +1,1 @@
+const s="./img/metaxy.svg";export{s as _};
