### 2026-04-21 anam-org/metaxy#1198 @Amorfati123
JOSS review: Unsubstantiated Performance Claims (repo issue)
Opened as part of the JOSS review at [https://github.com/openjournals/joss-reviews/issues/10449](url).

The paper states GPUs "cost 10 to 100 times more per hour than standard processors" (line 11–12) without a citation. More importantly, the docs and paper repeatedly reference processing "millions of training samples" and claim that "millions of cache keys can be calculated in under a second," but no benchmarks, timing data, or reproducible performance experiments are provided.
JOSS asks: "If there are any performance claims of the software, have they been confirmed?"

Recommendation: Either (a) include a minimal reproducible benchmark (even just a script in the examples/ folder that demonstrates key resolution time at scale), or (b) soften the performance language in the paper to avoid making claims that aren't backed by evidence a reviewer can verify. A simple timing table, e.g., "resolve_update on N records across K features completes in X seconds on commodity hardware" would suffice.
