### 2026-04-21 anam-org/metaxy#1198 @geoHeil
See the PR: https://github.com/anam-org/metaxy/pull/1199 is this loking fine for you @Amorfati123? So if merged, could we resolve this issue?
