### 2026-04-21 anam-org/metaxy#1198 @geoHeil
The claims sem from using a MPP-style/distributed system (one of the more powerful metaxy catalog backend implementations) in production for way larger datasets.
