### 2026-04-21 anam-org/metaxy#1198 @Amorfati123
This is fixed. Thanks to the authors.
