### 2026-04-21 anam-org/metaxy#1200 @Amorfati123
JOSS review: Software Design section needs trade-off discussion
Opened as part of the JOSS review at [https://github.com/openjournals/joss-reviews/issues/10449](url)

The Software Design section covers the architecture well (Pydantic models, DAG, hierarchical hashing, pluggable backends) but doesn't deeply explore the why behind these decisions.

The JOSS checklist asks: "Does the paper explain the trade-offs weighed, the design/architecture chosen, and why it matters?"

Some questions that would strengthen this section:

- Why Pydantic for feature definitions over simpler approaches (dataclasses, plain dicts)?

- Why a standalone metadata layer decoupled from compute, rather than embedding versioning into orchestrator plugins? What are the trade-offs?

- The paper mentions pushing "version computation into the metadata store through SQL", what's the alternative (client-side computation?), and what does SQL pushdown buy you?

- Why hierarchical hashing rather than content-addressable storage or other approaches?

Even 3-5 sentences addressing the most important trade-off (probably the decoupled metadata layer decision) would satisfy this requirement.
