### 2026-04-21 anam-org/metaxy#1201 @Amorfati123
JOSS Review: Figure 1 readability and Section heading "Research Impact" should be "Research Impact Statement"
Figure 1 in the rendered paper PDF has some readability issues, the version hash strings and "Feature Version" / "Field Version" labels overlap with box borders and other text, making it a bit hard to parse at first glance.

Paper section heading "Research Impact" should match JOSS template name.
