### 2026-04-21 anam-org/metaxy#1202 @Amorfati123
JOSS review: Reference list should include academic citations
Opened as part of the JOSS review at [[https://github.com/openjournals/joss-reviews/issues/10449](https://github.com/anam-org/metaxy/issues/url)]

The current reference list are all software project URLs or a single conference paper (Ray). The underlying concepts, dependency-based build systems, incremental computation, data lineage/provenance, have relevant academic literature that would strengthen the paper's scholarly context.

This isn't about padding the reference list, it's about situating Metaxy within the broader intellectual context. A paper about record-level dependency tracking should acknowledge the CS foundations it builds on.
