### 2026-04-21 #10449 @Amorfati123
Metaxy tackles a real and underserved problem in the ML tooling ecosystem. The field-level dependency tracking at record granularity is a genuinely novel contribution, and the decoupled metadata-layer approach is architecturally sound. The codebase is well-engineered, good test coverage, pluggable backends, proper CI/CD, comprehensive docs. The paper is generally well-written and clearly communicates the software's value proposition.

One concern I have is the first public release (v0.1.0) was February 13, 2026, and the paper was submitted on March 4, 2026, roughly three weeks later. The JOSS guidelines explicitly state:

"Projects developed privately are not eligible until there is a public record of open development: at least six months of public history prior to submission, with evidence of releases, public issues and pull requests."

Now, the repository does show good number of commits and substantial internal development, and the docs mention dogfooding since December 2025. But the public development history at the time of submission was well under six months. The repo has since accumulated more public history (releases v0.1.0 through v0.1.10), but I'd like the editor to weigh in on whether the timeline meets the spirit of the requirement. I'd recommend the authors clarify the project's development timeline, when was the repo first made public, and was there any earlier public-facing development (e.g., an earlier repo, public prototypes)?
