### 2026-04-21 #10449 @Amorfati123
Other reviews:

1. https://github.com/anam-org/metaxy/issues/1198 : Addressed and closed.
2. https://github.com/anam-org/metaxy/issues/1200
3. https://github.com/anam-org/metaxy/issues/1201
4. https://github.com/anam-org/metaxy/issues/1202
