### 2026-04-21 anam-org/metaxy#1200 @danielgafni
Thanks, I'll address your points tomorrow

Cc @geoHeil
