### 2026-04-21 anam-org/metaxy#1198 @danielgafni
I'd clarify that cache keys are calculated within a DB, at Anam we use a single-node Clickhouse deployment, there ain't anything distributed about this setup. DuckDB would yield very similar performance. Just OLAP being OLAP.
