### 2026-04-21 #10449 @danielgafni
@Amorfati123 the first commit it Metaxy was around October 2025. This covers 6 months of public development (it was open-source from the beginning, way before the first GitHub release).
