### 2026-04-22 #10449 @abhishektiwari
Thanks @Amorfati123 - when looking at "Was the software developed openly from early stages? For projects with recent public repositories, is there at least six months of public development history with evidence of releases, public issues/pull requests, and ideally external engagement?" criteria - We should consider totality of evidence. E.g, at least 6 months old, multiple developers, external engagement in terms of PR/watch/stars/issues.

In this case 6 months, multiple developers, 100+ watch/stars.
