### 2026-04-22 anam-org/metaxy#1191 @thc1006
Looks like PR #1194 addressed all the items here: ORCID, grammar fix, PostgreSQL/DuckLake, Apache Hamilton rename, and the title + tags reframing. Closing as resolved. Thanks for the fast turnaround.
