### 2026-04-22 anam-org/metaxy#1192 @thc1006
PR #1194 and PR #1199 addressed most of this thoroughly: footnote for the 10-100x anchor, two-layer hash description, code version framed as a deliberate design choice, append-only moved to Software Design, citation scope made consistent, and L114 reframed to match the automation value prop. The benchmark table in PR #1199 is especially useful for the Research Impact claim.

The one item not taken up is the State of Field comparison with record-level change-tracking systems like LakeFS, Iceberg, Hudi, or Delta Lake CDF. I'm reading that as a scope judgment (feature metadata vs data lake change tracking are different categories) and am happy to defer to it. Closing as resolved.
