### 2026-04-22 anam-org/metaxy#1193 @thc1006
All three suggestions were picked up in PR #1194: `@ray` has the arXiv eprint, `@hamilton` uses the SciPy 2024 DOI, and `@pydantic` has the Zenodo DOI. Closing as resolved. Thanks.
