### 2026-04-22 anam-org/metaxy#1200 @geoHeil
@Amorfati123 how about https://github.com/anam-org/metaxy/pull/1203 would this be fine for you?
