### 2026-04-22 anam-org/metaxy#1201 @geoHeil
@Amorfati123 is https://github.com/anam-org/metaxy/pull/1203 fine for you?
