### 2026-04-22 anam-org/metaxy#1200 @danielgafni
### Why Pydantic for feature definitions over simpler approaches (dataclasses, plain dicts)?

I'll give some more context about feature definitions (more info [here](https://docs.metaxy.io/stable/guide/concepts/definitions/features/#feature-definitions)). What Metaxy exactly calls a _feature definition_ is an instance of `metaxy.FeatureDefinition` class. These instances often need to be serialized/deserialized: Metaxy supports writing them to lock files and pushing them to the metadata store. `FeatureSpec` is one of the building blocks for `FeatureDefinition`.

Both `FeatureSpec` and `FeatureDefinition` have non-trivial validation rules which are best encoded with Pydantic (and the need to ser/de them speaks for itself).

Now, the most common _user interface_ for creating a new `FeatureDefinition` is to create a new child class of `metaxy.BaseFeature`. Under the hood this registers a new `FeatureDefinition` on Metaxy's feature graph.

Now I believe your original question must have been about `metaxy.BaseFeature`, since it's *also* a Pydantic model. Now this part isn't actually strictly necessary and was done for some nice-to-haves:
- to provide better support to some integrations (e.g. SQLModel, LanceDB) which are built on top of Pydantic
- because individual metadata samples are naturally represented by _instances of `BaseFeature` subclasses_ and are often ser/de-d between parts of data processing systems
- to benefit from some of the Pydantic features around schema generation

but having `metaxy.BaseFeature` a Pydantic model is not actually a very strong requirement. It may be lifted in the future. I won't consider it an inherit property of Metaxy. I will also [explore alternative interfaces ](https://github.com/anam-org/metaxy/issues/800)to `FeatureDefinition`, including creating them directly.

To summarize, Pydantic is used internally Metaxy needs to ser/de `FeatureDefinition` instances, but I would consider it an implementation detail and not a user-facing thing.

### Why a standalone metadata layer decoupled from compute, rather than embedding versioning into orchestrator plugins? What are the trade-offs?

I don't really understand this question, but the whole point of Metaxy is to be a pluggable library people could use with their own infrastructure.

Orchestrators and data processing frameworks are typically tightly coupled to infrastructure, so that's not something Metaxy is interested in building on top of.

Also, Metaxy can be used outside of data processing pipelines, e.g. to build reporting dashboards. It's just an orthogonal tool sitting at a different layer.

### The paper mentions pushing "version computation into the metadata store through SQL", what's the alternative (client-side computation?), and what does SQL pushdown buy you?

Pushing computations into the metadata store (database) is usually beneficial because:
1. The Python process does way less I/O during `MedataStore.resolve_update` (and also `MetadataStore.read`), it only needs to read the calculated increment instead of fetching entire full metadata tables
2. The Python process doesn't need to run any computations, which allows calling `MetadataStore.resolve_update` from lightweight environments, e.g. dashboards, laptops, and so on, while the database handles the compute (and for ex can auto-scale if needed).

### Why hierarchical hashing rather than content-addressable storage or other approaches?

Content-addressable storage *cannot be used in principle* in this situation because we want to acquire the increment _before any computations have been performed_. With content-addressable storage we would have to first compute the new samples only to realize they didn't have to be computed in the first place. In other words, there is no *content* for Metaxy to even run on, and Metaxy doesn't touch this content at all because it is decoupled from compute infrastructure (it doesn't rely on e.g. Object Storage e-tags or similar signatures for the same reason). Metaxy just never sees the data at all and doesn't need to.

However, Metaxy supports [user-defined data versions](https://docs.metaxy.io/stable/guide/concepts/versioning/#data-version), which can be used to override provenance once the computation ends, and this is where users could use content hashes, for example for some additional deduplication.
