### 2026-04-22 anam-org/metaxy#1200 @geoHeil
@Amorfati123 please see the updated version https://github.com/anam-org/metaxy/pull/1203
