+++ publications/2026-introducing-metaxy/paper.bib
@@ -12,14 +12,6 @@
   howpublished = {\url{https://narwhals.readthedocs.io}}
 }
 
-@software{pydantic,
-  author = {Samuel Colvin and Pydantic contributors},
-  title  = {Pydantic},
-  year   = {2025},
-  doi    = {10.5281/zenodo.19665399},
-  url    = {https://pydantic.dev}
-}
-
 @misc{dagster,
   author = {{Dagster Labs}},
   title = {Dagster: Data Orchestration Platform},
@@ -66,4 +58,66 @@
   title = {DataChain: AI Data Warehouse for Multimodal Data},
   year = {2025},
   howpublished = {\url{https://datachain.ai}}
+}
+
+@article{mokhov2018build,
+  author    = {Mokhov, Andrey and Mitchell, Neil and Peyton Jones, Simon},
+  title     = {Build Systems {\`a} la Carte},
+  journal   = {Proceedings of the ACM on Programming Languages},
+  volume    = {2},
+  number    = {ICFP},
+  year      = {2018},
+  pages     = {79:1--79:29},
+  doi       = {10.1145/3236774}
+}
+
+@inproceedings{acar2006adaptive,
+  author    = {Acar, Umut A. and Blelloch, Guy E. and Harper, Robert},
+  title     = {Adaptive Functional Programming},
+  booktitle = {ACM Transactions on Programming Languages and Systems (TOPLAS)},
+  volume    = {28},
+  number    = {6},
+  year      = {2006},
+  pages     = {990--1034},
+  doi       = {10.1145/1186632.1186634}
+}
+
+@article{cui2003lineage,
+  author    = {Cui, Yingwei and Widom, Jennifer},
+  title     = {Lineage Tracing for General Data Warehouse Transformations},
+  journal   = {The VLDB Journal},
+  volume    = {12},
+  number    = {1},
+  year      = {2003},
+  pages     = {41--58},
+  doi       = {10.1007/s00778-002-0083-8}
+}
+
+@inproceedings{buneman2001why,
+  author    = {Buneman, Peter and Khanna, Sanjeev and Tan, Wang-Chiew},
+  title     = {Why and Where: A Characterization of Data Provenance},
+  booktitle = {Proceedings of the 8th International Conference on Database Theory (ICDT)},
+  year      = {2001},
+  pages     = {316--330},
+  doi       = {10.1007/3-540-44503-X_20}
+}
+
+@article{cheney2009provenance,
+  author    = {Cheney, James and Chiticariu, Laura and Tan, Wang-Chiew},
+  title     = {Provenance in Databases: Why, How, and Where},
+  journal   = {Foundations and Trends in Databases},
+  volume    = {1},
+  number    = {4},
+  year      = {2009},
+  pages     = {379--474},
+  doi       = {10.1561/1900000006}
+}
+
+@techreport{moreau2013prov,
+  author      = {Moreau, Luc and Missier, Paolo and Belhajjame, Khalid and B'Far, Reza and Cheney, James and Coppens, Sam and Cresswell, Stephen and Gil, Yolanda and Groth, Paul and Klyne, Graham and Lebo, Timothy and McCusker, Jim and Miles, Simon and Myers, James and Sahoo, Satya and Tilmes, Curt},
+  title       = {{PROV-DM: The PROV Data Model}},
+  institution = {World Wide Web Consortium (W3C)},
+  type        = {W3C Recommendation},
+  year        = {2013},
+  url         = {https://www.w3.org/TR/prov-dm/}
 }
