+++ publications/2026-introducing-metaxy/paper.md
@@ -42,8 +42,8 @@ It builds a dependency graph that connects individual data fields across process
 When a researcher modifies one step, Metaxy identifies exactly which records are affected and which can be skipped.
 This selective approach lets downstream systems avoid redundant GPU work when a change does not affect a record, while preserving complete lineage for reproducibility.
 
-The library integrates with various backends through Ibis (@ibis) and supports multiple dataframe engines via Narwhals (@narwhals).
-Orchestration platforms such as Dagster (@dagster) and Ray consume Metaxy's record-level diffs to schedule only the necessary GPU workloads.
+The library treats the *metadata store* (which persists version entries), the *compute engine* (which evaluates feature code over dataframes), and the *orchestrator* (which schedules execution) as pluggable abstractions.
+Concrete integrations for the widely used Python ecosystem ship with the package (@ibis; @narwhals; @dagster; @ray), so that any compatible orchestrator can consume Metaxy's record-level diffs and schedule only the necessary GPU workloads.
 
 # Statement of Need
 
@@ -58,7 +58,7 @@ Existing orchestrators operate at table granularity, treating feature tables as 
 They cannot distinguish which fields within a record changed, forcing unnecessary recomputation of all downstream steps.
 
 Metaxy resolves this problem through field-level dependency tracking at record granularity.
-Researchers declare features as Python classes (@pydantic) that specify which upstream fields each computation depends on.
+Researchers declare features as Python classes that specify which upstream fields each computation depends on.
 The system constructs a directed acyclic graph where nodes represent feature fields and edges represent data flow (\autoref{fig:anatomy}).
 For each data record, Metaxy computes version hashes that combine code versions with upstream record versions, propagating changes along graph edges.
 When resolving incremental updates, the system returns only those records whose upstream dependencies have actually changed.
@@ -81,11 +81,11 @@ Apache Hamilton (@hamilton) builds dataflows from Python functions and can repor
 DataChain (@datachain) combines metadata management with a Python-native data processing framework and supports delta processing over new or changed records. However, its incremental model is tied to dataset processing within that framework rather than to a standalone field-level dependency graph that can drive recomputation across multiple compute backends.
 
 Metaxy fills the gap by separating metadata from compute: it tracks field-level dependencies at record granularity and propagates version changes topologically through the dependency graph.
-This separation allows teams to integrate Metaxy with any compute framework, such as Ray or Dagster, while retaining precise control over which records require reprocessing.
+This separation allows teams to integrate Metaxy with any orchestrator and compute engine, while retaining precise control over which records require reprocessing.
 
 # Software Design
 
-Metaxy represents features as declarative Pydantic models organized into a directed acyclic graph.
+Metaxy represents features as declarative models organized into a directed acyclic graph.
 Each feature declares its identifier columns, computed fields, and dependencies on upstream feature fields.
 The system constructs a global dependency graph at initialization, enabling downstream version propagation before any data processing occurs.
 
@@ -97,9 +97,19 @@ This record-level granularity is what distinguishes Metaxy from table-level orch
 When an upstream field changes for a subset of records, only those downstream records that transitively depend on the changed field are marked for recomputation.
 Records with unchanged dependencies are skipped.
 
-Metadata persistence uses pluggable, append-only storage backends: version entries are never overwritten, preserving the lineage needed for retrospective audits.
-DuckDB provides embedded storage for prototyping; ClickHouse, BigQuery, Delta Lake, DuckLake, LanceDB and PostgreSQL scale to production workloads.
-Narwhals supplies a backend-agnostic dataframe interface, allowing users to work with Pandas, Polars, or Ibis interchangeably.
[…]
