+++ publications/2026-introducing-metaxy/paper.md
@@ -114,18 +114,26 @@ Third, Metaxy hashes provenance signatures (code version plus upstream record ve
 # Performance
 
 Per-record hashing and the downstream diff are executed as SQL inside the metadata store, so their throughput is bounded by the vectorised hash kernels of the backend rather than by Python.
-The benchmark script `publications/2026-introducing-metaxy/benchmark.py` materialises a two-feature graph (a root with fields `audio` and `frames`, and a downstream leaf) against a fresh DuckDB store, measures `resolve_update` for the initial materialisation (`resolve_new`), then bumps the upstream `audio` provenance for 10% of records and measures the incremental diff (`resolve_stale`).
-Medians over three runs on an Apple M2 Max (64 GiB RAM, Python 3.10.19, DuckDB 1.4.3, `xxhash64`) are reported in \autoref{tab:benchmark}; throughput scales linearly with `N` and saturates near $10^7$ records per second once the hash kernel dominates fixed setup costs.
+The benchmark in `publications/2026-introducing-metaxy/` is implemented as a [`pytest-benchmark`](https://pytest-benchmark.readthedocs.io) suite that exercises two feature graphs: a `simple` graph (one root with two fields feeding a single-field leaf) and a `wide` graph (two roots with four fields each feeding a two-field leaf, so the leaf must join two upstream tables and aggregate eight field versions per record).
+For each graph and record count $N$, the benchmark measures `resolve_update` in two phases: the initial materialisation (`resolve_new`) and, after bumping the upstream `audio` provenance for 10% of records, the incremental diff (`resolve_stale`).
 
-: Median `resolve_update` wall-clock time on DuckDB for `N` records across a 2-feature graph (3 runs, 10% change fraction). \label{tab:benchmark}
+To isolate caching effects between rounds, each measurement runs in `pytest-benchmark`'s pedantic mode with a `setup` callback that allocates a fresh DuckDB file in its own temporary directory and reseeds it from scratch. The DuckDB buffer pool, temp-table state, and file-system page cache for the new inode are therefore cold at the start of every round, so no round can inherit state from a previous one. The configuration under test sets `enable_map_datatype = true` in `metaxy.toml`, reflecting the upcoming default; the provenance column is stored and exchanged as a native `Map`. We report ten independent rounds per cell on an Apple M2 Max (64 GiB RAM, Python 3.10.19, DuckDB 1.4.3, `xxhash64`): the median as a robust central estimate, the 25th--75th percentile interquartile range (IQR) as a non-parametric confidence band, and mean $\pm$ standard deviation for readers preferring a parametric view (\autoref{tab:benchmark}).
+End-to-end wall-clock scales near-linearly with $N$ on both graphs (\autoref{tab:benchmark}, \autoref{fig:benchmark}); the constant-factor gap between `simple` and `wide` at any given $N$ reflects the additional upstream join and the richer provenance struct rather than super-linear slowdown, confirming that the SQL pushdown strategy absorbs join width well.
 
-|          N |  resolve_new (s) | resolve_stale (s) |   rows/s (new) |
-|-----------:|-----------------:|------------------:|---------------:|
-|     10,000 |            0.089 |             0.128 |        112,938 |
-|    100,000 |            0.101 |             0.188 |        985,968 |
-|  1,000,000 |            0.169 |             0.439 |      5,931,115 |
[…]
