+++ pyproject.toml
@@ -147,6 +147,8 @@ dev = [
   "ty==0.0.25",
   "datafusion>=51.0.0",
   "pytest-asyncio>=0.26.0",
+  "pytest-benchmark>=5.0.0",
+  "matplotlib>=3.8.0",
   "pandas>=2.3.3",
   "numpy>=2.2.6",
   "sybil>=9.3.0",
