+++ main/publications/2026-introducing-metaxy/benchmark.py
@@ -1,232 +1,0 @@
-"""Reproducible benchmark for the JOSS paper.
-
-Measures the wall-clock time of ``resolve_update`` on the embedded DuckDB
-metadata store for a simple two-feature graph, across a range of record
-counts. The benchmark exercises the full pipeline: SQL-side per-record
-hashing, join with existing target metadata, and diff computation.
-
-Run:
-
-    uv run python publications/2026-introducing-metaxy/benchmark.py
-
-Optional flags:
-
-    --records 1_000,100_000,1_000_000   comma-separated list of N values
-    --repeats 3                         runs per N (median is reported)
-    --database :memory:                 DuckDB path (default: tempfile)
-    --hash xxhash64                     hash algorithm
-"""
-
-from __future__ import annotations
-
-import argparse
-import json
-import random
-import statistics
-import tempfile
-import time
-import warnings
-from dataclasses import dataclass
-from pathlib import Path
-
-warnings.filterwarnings(
-    "ignore",
-    message="Global Metaxy configuration not initialized.*",
-)
-warnings.filterwarnings("ignore", message="AUTO_CREATE_TABLES is enabled.*")
-
-import polars as pl
-
-import metaxy as mx
-from metaxy.config import MetaxyConfig
-from metaxy.ext.duckdb import DuckDBMetadataStore
-from metaxy.models.constants import METAXY_PROVENANCE_BY_FIELD
-from metaxy.versioning.types import HashAlgorithm
-from metaxy_testing import add_metaxy_provenance_column
-
-MetaxyConfig.set(MetaxyConfig())
-
-
-# ---------------------------------------------------------------------------
-# Feature graph
-# ---------------------------------------------------------------------------
-
-
-def build_graph() -> tuple[type[mx.BaseFeature], type[mx.BaseFeature], mx.FeatureGraph]:
-    graph = mx.FeatureGraph()
-    with graph.use():
-
-        class Root(
-            mx.BaseFeature,
-            spec=mx.FeatureSpec(
-                key="bench/root",
-                fields=[
-                    mx.FieldSpec(key="audio", code_version="1"),
-                    mx.FieldSpec(key="frames", code_version="1"),
-                ],
-                id_columns=("sample_uid",),
-            ),
-        ):
-            pass
-
-        class Leaf(
-            mx.BaseFeature,
-            spec=mx.FeatureSpec(
-                key="bench/leaf",
-                deps=[Root],
-                fields=[mx.FieldSpec(key="prediction", code_version="1")],
-                id_columns=("sample_uid",),
-            ),
-        ):
-            pass
-
-    return Root, Leaf, graph
-
-
-# ---------------------------------------------------------------------------
-# Synthetic data
-# ---------------------------------------------------------------------------
-
-
-def make_upstream_df(n: int, seed: int, root_feature: type[mx.BaseFeature]) -> pl.DataFrame:
-    rng = random.Random(seed)
-    audio_versions = [f"a{rng.randint(0, 2**20)}" for _ in range(n)]
-    frames_versions = [f"f{rng.randint(0, 2**20)}" for _ in range(n)]
-    df = pl.DataFrame(
-        {
-            "sample_uid": list(range(n)),
-            METAXY_PROVENANCE_BY_FIELD: [
-                {"audio": a, "frames": f} for a, f in zip(audio_versions, frames_versions, strict=True)
-            ],
-        },
-        schema={
-            "sample_uid": pl.UInt64,
-            METAXY_PROVENANCE_BY_FIELD: pl.Struct({"audio": pl.Utf8, "frames": pl.Utf8}),
-        },
-    )
[…]
