### 2026-04-27 anam-org/metaxy#1202 @Amorfati123
This has been addressed. Closing this issue. Thanks
