### 2026-04-27 #10449 @thc1006
Review complete. The items I raised in anam-org/metaxy#1191, #1192 and #1193 are all addressed in the v1.1 paper, and the new Performance section with the benchmark table is a real improvement over what I first read. Recommending acceptance.

Two small heads-ups for editor discretion, not blockers:

1. The "Research Impact Statement" section was renamed to "Ongoing Research Projects". Content-wise it's the same (Anam production, reproducibility), but the substring `Research impact` no longer appears, so `@editorialbot generate pdf` may flag the section check next time it's re-run.
2. Word count is now around 2,200 (mostly the benchmark table). Soft guideline is 1,750. Up to the editor whether trimming is wanted.

Thanks to @danielgafni and @geoHeil for turning everything around so fast, and to @abhishektiwari for handling.
