### 2026-04-27 #10449 @danielgafni
May we do another pass on the paper? This could trim the word count.

I'd be down to do it this week.
