### 2026-05-04 effe454
Fix Dagster table preview for Map columns with non-primitive value types
