### 2026-05-04 4c0f3c4
fix(dagster): simplify table formatting in Python
