### 2026-05-14 d07a7fd
chore(publication): address reviewer feedback for SAO (#1207)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->

## Addressing reviewer feedback:
### reviewer 1
Paper summary

The paper presents Metaxy, an open-source metadata layer for field-level, record-level dependency tracking in multimodal ML pipelines. The core idea is practical and relevant: instead of invalidating entire downstr
[…]
