+++ publications/2026-metaxy-field-level-versioning/paper.tex
@@ -1,3 +1,4 @@
+%% ACM Primary Article Template, SIGCONF proceedings format.
 \documentclass[sigconf]{acmart}
 
 %% Remove copyright / ACM reference for workshop submission draft
@@ -10,20 +11,25 @@
 \acmConference[SAO @ CAIS 2026]{Workshop on Supporting Our AI Overlords}{May 26, 2026}{Berkeley, CA, USA}
 
 \usepackage{graphicx}
+\usepackage{booktabs}
 
 \usepackage{listings}
 \lstset{
   language=Python,
-  basicstyle=\ttfamily\scriptsize,
+  basicstyle=\ttfamily\footnotesize,
   keywordstyle=\bfseries,
   commentstyle=\itshape,
-  columns=flexible,
+  columns=fullflexible,
   breaklines=true,
+  breakatwhitespace=true,
   aboveskip=4pt,
   belowskip=4pt,
-  xleftmargin=2pt,
+  xleftmargin=0pt,
   frame=none,
 }
+
+%% Give slightly more freedom on hyphenation and line breaks for tight columns.
+\sloppy
 
 \begin{document}
 
@@ -39,7 +45,7 @@
 \email{heiler@csh.ac.at}
 
 \begin{abstract}
-Multimodal ML pipelines incur substantial GPU costs, yet existing orchestrators invalidate entire downstream datasets when any processing step changes---even for unaffected fields.
+Multimodal ML pipelines incur high GPU costs, yet existing orchestrators invalidate entire downstream datasets when any processing step changes---even for unaffected fields.
 We present \textbf{Metaxy}, an open-source Python library providing \emph{field-level dependency tracking at record granularity}.
 Feature definitions form a directed acyclic graph; for each record, hierarchical version hashes propagate changes along field-level edges.
 A \texttt{resolve\_update} operation returns exactly those records that are new, stale, or orphaned, enabling selective recomputation.
@@ -71,73 +77,169 @@ In production across two organizations processing millions of multimodal samples
 \section{Introduction}
 
 Modern ML increasingly relies on multimodal data---video, audio, images, and text---processed through GPU-accelerated pipelines.
-Unlike tabular ETL, where re-execution is cheap, multimodal steps such as speech-to-text transcription or embedding generation cost orders of magnitude more per record.
+Unlike tabular ETL, where re-execution is cheap, multimodal steps such as speech-to-text transcription or embedding generation can be expensive per record~\cite{anyscale2024gpu}.
 Data orchestrators like Dagster~\cite{dagster} model pipelines as graphs of \emph{assets} (tables or datasets) and track freshness at the asset level: when a processing step changes, every downstream record is recomputed---even those whose inputs are unchanged.
 
-Consider a video pipeline extracting both audio transcripts and face detections (Figure~\ref{fig:field-deps}).
-Improving only the audio denoising algorithm triggers recomputation of face detections that depend solely on video frames---a costly false positive.
-This problem compounds as pipelines grow and is especially acute for iterative, agent-driven ML development, where rapid experimentation on feature pipelines demands that each iteration recomputes only what is strictly necessary.
+Consider a video pipeline that extracts audio transcripts and face detections (Figure~\ref{fig:field-deps}).
+Improving only the audio denoising algorithm should not trigger recomputation of face detections that depend solely on video frames---yet at table granularity, that recomputation fires anyway.
+This problem compounds as pipelines grow, and is especially acute for iterative, agent-driven ML development, where rapid experimentation demands that each iteration recomputes only what is strictly necessary.
 
[…]
