+++ publications/2026-metaxy-field-level-versioning/references.bib
@@ -40,9 +40,103 @@
   howpublished = {\url{https://datachain.ai}}
 }
 
+@misc{feast,
+  author = {{Feast Community}},
+  title = {Feast: Feature Store for Machine Learning},
+  year = {2025},
+  howpublished = {\url{https://feast.dev}}
+}
+
+@inproceedings{hamilton,
+  author    = {Krawczyk, Stefan and Gilani, Elijah},
+  title     = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
+  booktitle = {Proceedings of the 23rd Python in Science Conference (SciPy 2024)},
+  year      = {2024},
+  doi       = {10.25080/gerudo-f2bc6f59-014},
+  url       = {https://hamilton.apache.org}
+}
+
 @misc{dagsterslurm,
   author = {Heiler, Georg and Picatto, Hernan and He\ss{}, Maximilian and Pfister, Martin},
   title = {dagster-slurm: Reproducible Data Orchestration on {HPC}},
   year = {2025},
   howpublished = {\url{https://github.com/ascii-supply-networks/dagster-slurm}}
+}
+
+@article{mokhov2018build,
+  author    = {Mokhov, Andrey and Mitchell, Neil and Peyton Jones, Simon},
+  title     = {Build Systems {\`a} la Carte},
+  journal   = {Proceedings of the ACM on Programming Languages},
+  volume    = {2},
+  number    = {ICFP},
+  year      = {2018},
+  pages     = {79:1--79:29},
+  doi       = {10.1145/3236774}
+}
+
+@article{acar2006adaptive,
+  author    = {Acar, Umut A. and Blelloch, Guy E. and Harper, Robert},
+  title     = {Adaptive Functional Programming},
+  journal   = {ACM Transactions on Programming Languages and Systems},
+  volume    = {28},
+  number    = {6},
+  year      = {2006},
+  pages     = {990--1034},
+  doi       = {10.1145/1186632.1186634}
+}
+
+@article{cui2003lineage,
+  author    = {Cui, Yingwei and Widom, Jennifer},
+  title     = {Lineage Tracing for General Data Warehouse Transformations},
+  journal   = {The VLDB Journal},
+  volume    = {12},
+  number    = {1},
+  year      = {2003},
+  pages     = {41--58},
+  doi       = {10.1007/s00778-002-0083-8}
+}
+
+@inproceedings{buneman2001why,
+  author    = {Buneman, Peter and Khanna, Sanjeev and Tan, Wang-Chiew},
+  title     = {Why and Where: A Characterization of Data Provenance},
+  booktitle = {Proceedings of the 8th International Conference on Database Theory (ICDT)},
+  year      = {2001},
+  pages     = {316--330},
+  doi       = {10.1007/3-540-44503-X_20}
+}
+
+@article{cheney2009provenance,
+  author    = {Cheney, James and Chiticariu, Laura and Tan, Wang-Chiew},
+  title     = {Provenance in Databases: Why, How, and Where},
+  journal   = {Foundations and Trends in Databases},
+  volume    = {1},
+  number    = {4},
+  year      = {2009},
+  pages     = {379--474},
+  doi       = {10.1561/1900000006}
+}
+
+@techreport{moreau2013prov,
+  author      = {Moreau, Luc and Missier, Paolo and Belhajjame, Khalid and Cheney, James and Coppens, Sam and Cresswell, Stephen and Gil, Yolanda and Groth, Paul and Klyne, Graham and Lebo, Timothy and McCusker, Jim and Miles, Simon and Myers, James and Sahoo, Satya and Tilmes, Curt},
+  title       = {{PROV-DM: The PROV Data Model}},
[…]
