### 2026-05-14 7d173fd
chore(slides): add VDSG metaxy-applications
