+++ mkdocs.yml
@@ -310,6 +310,7 @@ nav:
       - Introducing Metaxy (2026): publications/2026-introducing-metaxy/paper.md
     - Slides:
       - Introducing Metaxy (2026): slides/2026-introducing-metaxy/intro.md
+      - Metaxy Applications (2026): https://georgheiler.com/event/optimizing-multimodal-ai-pipelines-with-metaxy/
   - 📝 Changelog: changelog.md
 
 extra:
