### 2026-05-15 a931992
refactor: use native `narwhals.struct` expression in the versioning engine (#1209)
## Summary

resolves: https://github.com/anam-org/metaxy/issues/1188

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
