+++ src/metaxy/ext/polars/versioning.py
@@ -85,35 +85,7 @@ class PolarsVersioningEngine(VersioningEngine):
         return cast(FrameT, nw.from_native(df_pl))
 
     @staticmethod
-    def build_struct_column(
-        df: FrameT,
-        struct_name: str,
-        field_columns: dict[str, str],
-    ) -> FrameT:
-        """Build a metadata column (Struct or Map) from existing columns.
-
-        When ``enable_map_datatype`` is set, produces a ``polars_map.Map(String, String)``
-        column. Otherwise produces a Struct column.
-        """
-        from metaxy.config import MetaxyConfig
-
-        if MetaxyConfig.get().enable_map_datatype:
-            return PolarsVersioningEngine._build_map_column(df, struct_name, field_columns)
-        return PolarsVersioningEngine._build_polars_struct_column(df, struct_name, field_columns)
-
-    @staticmethod
-    def _build_polars_struct_column(
-        df: FrameT,
-        col_name: str,
-        field_columns: dict[str, str],
-    ) -> FrameT:
-        """Build a Polars Struct column from existing columns."""
-        df_pl = cast(pl.DataFrame | pl.LazyFrame, df.to_native())  # ty: ignore[invalid-argument-type]
-        struct_expr = pl.struct([pl.col(src_col).alias(field_name) for field_name, src_col in field_columns.items()])
-        return cast(FrameT, nw.from_native(df_pl.with_columns(struct_expr.alias(col_name))))
-
-    @staticmethod
-    def _build_map_column(
+    def build_map_column(
         df: FrameT,
         col_name: str,
         field_columns: dict[str, str],
@@ -121,6 +93,7 @@ class PolarsVersioningEngine(VersioningEngine):
         """Build a polars_map.Map(String, String) column from existing columns."""
         import polars_map  # noqa: F401  # registers .map accessor
 
+        assert df.implementation == nw.Implementation.POLARS, "Only Polars DataFrames are accepted"
         df_pl = cast(pl.DataFrame | pl.LazyFrame, df.to_native())  # ty: ignore[invalid-argument-type]
         kv_pairs = [
             pl.struct(
