+++ pyproject.toml
@@ -28,7 +28,7 @@ classifiers = [
 ]
 dependencies = [
   "cyclopts>=4.4.0",
-  "narwhals>=2.18.1",
+  "narwhals>=2.20.0",
   "typing-extensions>=4.0.0",
   "pydantic-core>=2.0.0",
   "polars>=1.35.0",
