### 2026-06-03 c5cd9aa
test: verify Ibis path in metadata read; drop history comments and redundant import
