### 2026-06-04 d7edd33
fix(cli): update IbisMetadataStore import to relocated ext.ibis module
The MetadataStore refactor on main moved metaxy.metadata_store.ibis to
metaxy.ext.ibis.metadata_store. Update the metadata read command's import
accordingly to resolve the typecheck failure.
