+++ src/metaxy/_hashing.py
@@ -116,7 +116,7 @@ def ensure_hash_compatibility(hash1: str, hash2: str) -> bool:
 
     # Check if one is a prefix of the other (truncation)
     shorter, longer = sorted([hash1, hash2], key=len)
-    return longer.startswith(shorter)  # ty: ignore[unresolved-attribute]
+    return longer.startswith(shorter)
 
 
 @nw.narwhalify
