+++ src/metaxy/ext/ray/datasource.py
@@ -166,7 +166,7 @@ class MetaxyDatasource(Datasource):
         with self.store:
             return self._read_lazy().select(nw.len()).collect().item()
 
-    def get_read_tasks(
+    def get_read_tasks(  # ty: ignore[invalid-method-override]
         self, parallelism: int, per_task_row_limit: int | None = None, **kwargs: object
     ) -> list[ReadTask]:
         """Return read tasks for the feature metadata.
