+++ tests/metaxy_testing/src/metaxy_testing/pytest_helpers.py
@@ -40,8 +40,8 @@ def add_metaxy_system_columns(df: IntoFrameT) -> IntoFrameT:
         DataFrame with all required metaxy system columns (same type as input)
     """
     # Convert to narwhals if needed (works with both eager and lazy frames)
-    df_nw = nw.from_native(df)
-    columns = df_nw.collect_schema().names()  # ty: ignore[unresolved-attribute]
+    df_nw = nw.from_native(df)  # ty: ignore[no-matching-overload]
+    columns = df_nw.collect_schema().names()
 
     columns_to_add: list[nw.Expr] = []
 
@@ -70,7 +70,7 @@ def add_metaxy_system_columns(df: IntoFrameT) -> IntoFrameT:
                 columns_to_add.append(nw.lit("test_data_version").alias(METAXY_DATA_VERSION))
 
     if columns_to_add:
-        df_nw = df_nw.with_columns(columns_to_add)  # ty: ignore[unresolved-attribute]
+        df_nw = df_nw.with_columns(columns_to_add)
 
     return df_nw.to_native()
 
