### 2026-06-04 261bec5
fix(cli): produce real markdown without fallback; drop dead read_all checks
