### 2026-06-04 ddff9f4
fix(cli): render markdown natively without pandas
Build the markdown table directly from the DuckDB Arrow result instead
of pandas to_markdown(), per maintainer feedback to avoid pandas.
