### 2026-06-05 f9dc595
fix(clickhouse): make structs non-nullable before writing (for ClickHouse 26.4)
