+++ main/publications/2026-introducing-metaxy/paper.bib
@@ -6,7 +6,7 @@
 }
 
 @misc{narwhals,
-  author = {Marco Gorelli and contributors},
+  author = {Gorelli, Marco and others},
   title = {Narwhals: Write Once, Run on Many DataFrames},
   year = {2025},
   howpublished = {\url{https://narwhals.readthedocs.io}}
