+++ src/metaxy/ext/starrocks/__init__.py
@@ -1,0 +1,15 @@
+"""StarRocks metadata store extension."""
+
+from metaxy.ext.starrocks.metadata_store import (
+    StarRocksMetadataStore,
+    StarRocksMetadataStoreConfig,
+    StarRocksMySQLMetadataStore,
+)
+from metaxy.ext.starrocks.versioning import StarRocksVersioningEngine
+
+__all__ = [
+    "StarRocksMetadataStore",
+    "StarRocksMetadataStoreConfig",
+    "StarRocksMySQLMetadataStore",
+    "StarRocksVersioningEngine",
+]
