### 2026-06-07 10d1546
feat(versioning): add xxh3_64 hash support
