### 2026-06-08 6456fb5
feat: add `xxh3_64` hash support (#1220)
Add support for the faster `xxh3_64` hash function
