+++ .github/workflows/_qa.yml
@@ -5,6 +5,8 @@ permissions:
 
 env:
   FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: "true"
+  # Silence the Material for MkDocs "MkDocs 2.0" upgrade notice.
+  NO_MKDOCS_2_WARNING: "1"
 
 on:
   workflow_call:
