### 2026-06-09 98b47c7
Deployed 6f0867d4 to 0.1.11 with MkDocs 1.6.1 and mike 2.1.3
