+++ 0.1.11/publications/2026-introducing-metaxy/paper.bib
@@ -1,0 +1,123 @@
+@misc{ibis,
+  author = {{Ibis Project}},
+  title = {Ibis: Portable DataFrames for Different Backends},
+  year = {2025},
+  howpublished = {\url{https://ibis-project.org}}
+}
+
+@misc{narwhals,
+  author = {Gorelli, Marco and others},
+  title = {Narwhals: Write Once, Run on Many DataFrames},
+  year = {2025},
+  howpublished = {\url{https://narwhals.readthedocs.io}}
+}
+
+@misc{dagster,
+  author = {{Dagster Labs}},
+  title = {Dagster: Data Orchestration Platform},
+  year = {2025},
+  howpublished = {\url{https://dagster.io}}
+}
+
+@inproceedings{ray,
+  author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Paul, Melanie and Jordan, Michael I. and Stoica, Ion},
+  title = {Ray: A Distributed Framework for Emerging AI Applications},
+  booktitle = {Proceedings of the 13th USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
+  year = {2018},
+  eprint = {1712.05889},
+  archivePrefix = {arXiv},
+  primaryClass = {cs.DC},
+  url = {https://arxiv.org/abs/1712.05889}
+}
+
+@software{dvc,
+  author = {Kuprieiev, Ruslan and Petrov, Dmitry and others},
+  title = {DVC: Data Version Control},
+  year = {2025},
+  howpublished = {\url{https://dvc.org}}
+}
+
+@misc{feast,
+  author = {{Feast Community}},
+  title = {Feast: Feature Store for Machine Learning},
+  year = {2025},
+  howpublished = {\url{https://feast.dev}}
+}
+
+@inproceedings{hamilton,
+  author    = {Krawczyk, Stefan and Gilani, Elijah},
+  title     = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
+  booktitle = {Proceedings of the 23rd Python in Science Conference (SciPy 2024)},
+  year      = {2024},
+  doi       = {10.25080/gerudo-f2bc6f59-014},
+  url       = {https://hamilton.apache.org}
+}
+
+@misc{datachain,
+  author = {{Iterative, Inc.}},
+  title = {DataChain: AI Data Warehouse for Multimodal Data},
+  year = {2025},
+  howpublished = {\url{https://datachain.ai}}
+}
+
+@article{mokhov2018build,
+  author    = {Mokhov, Andrey and Mitchell, Neil and Peyton Jones, Simon},
+  title     = {Build Systems {\`a} la Carte},
+  journal   = {Proceedings of the ACM on Programming Languages},
+  volume    = {2},
+  number    = {ICFP},
+  year      = {2018},
+  pages     = {79:1--79:29},
+  doi       = {10.1145/3236774}
+}
+
+@inproceedings{acar2006adaptive,
+  author    = {Acar, Umut A. and Blelloch, Guy E. and Harper, Robert},
+  title     = {Adaptive Functional Programming},
+  booktitle = {ACM Transactions on Programming Languages and Systems (TOPLAS)},
+  volume    = {28},
+  number    = {6},
+  year      = {2006},
+  pages     = {990--1034},
+  doi       = {10.1145/1186632.1186634}
+}
+
+@article{cui2003lineage,
+  author    = {Cui, Yingwei and Widom, Jennifer},
+  title     = {Lineage Tracing for General Data Warehouse Transformations},
+  journal   = {The VLDB Journal},
+  volume    = {12},
+  number    = {1},
+  year      = {2003},
[…]
