+++ 0.1.11/slides/2026-introducing-metaxy/dist/assets/slidev/TitleIcon.vue_vue_type_script_setup_true_lang-D0-vSwJ2.js
@@ -1,0 +1,1 @@
+import{d as p,f as d,i as f,o as v,n as y}from"../modules/vue-BtB2w7oG.js";const a=p({__name:"TitleIcon",props:{title:{}},setup(o){const c={pnpm:"i-vscode-icons:file-type-light-pnpm",npm:"i-vscode-icons:file-type-npm",yarn:"i-vscode-icons:file-type-yarn",bun:"i-vscode-icons:file-type-bun",deno:"i-vscode-icons:file-type-deno",vue:"i-vscode-icons:file-type-vue",svelte:"i-vscode-icons:file-type-svelte",angular:"i-vscode-icons:file-type-angular",react:"i-vscode-icons:file-type-reactjs",next:"i-vscode-icons:file-type-light-next",nuxt:"i-vscode-icons:file-type-nuxt",solid:"logos:solidjs-icon",astro:"i-vscode-icons:file-type-light-astro",rollup:"i-vscode-icons:file-type-rollup",webpack:"i-vscode-icons:file-type-webpack",vite:"i-vscode-icons:file-type-vite",esbuild:"i-vscode-icons:file-type-esbuild","package.json":"i-vscode-icons:file-type-node","tsconfig.json":"i-vscode-icons:file-type-tsconfig",".npmrc":"i-vscode-icons:file-type-npm",".editorconfig":"i-vscode-icons:file-type-editorconfig",".eslintrc":"i-vscode-icons:file-type-eslint",".eslintignore":"i-vscode-icons:file-type-eslint","eslint.config":"i-vscode-icons:file-type-eslint",".gitignore":"i-vscode-icons:file-type-git",".gitattributes":"i-vscode-icons:file-type-git",".env":"i-vscode-icons:file-type-dotenv",".env.example":"i-vscode-icons:file-type-dotenv",".vscode":"i-vscode-icons:file-type-vscode","tailwind.config":"vscode-icons:file-type-tailwind","uno.config":"i-vscode-icons:file-type-unocss","unocss.config":"i-vscode-icons:file-type-unocss",".oxlintrc":"i-vscode-icons:file-type-oxlint","vue.config":"i-vscode-icons:file-type-vueconfig",".mts":"i-vscode-icons:file-type-typescript",".cts":"i-vscode-icons:file-type-typescript",".ts":"i-vscode-icons:file-type-typescript",".tsx":"i-vscode-icons:file-type-typescript",".mjs":"i-vscode-icons:file-type-js",".cjs":"i-vscode-icons:file-type-js",".json":"i-vscode-icons:file-type-json",".js":"i-vscode-icons:file-type-js",".jsx":"i-vscode-icons:file-type-js",".md":"i-vscode-icons:file-type-markdown",".py":"i-vscode-icons:file-type-python",".ico":"i-vscode-icons:file-type-favicon",".html":"i-vscode-icons:file-type-html",".css":"i-vscode-icons:file-type-css",".scss":"i-vscode-icons:file-type-scss",".yml":"i-vscode-icons:file-type-light-yaml",".yaml":"i-vscode-icons:file-type-light-yaml",".php":"i-vscode-icons:file-type-php",".svg":"i-vscode-icons:file-type-svg"};function t(s){const i=s.match(/~([^~]+)~/g);if(i&&i.length>0)return i[0].slice(1,-1);const n=Object.keys(c).sort((e,l)=>l.length-e.length);for(const e of n)if(s.toLowerCase().includes(e.toLowerCase()))return c[e];return""}return(s,i)=>t(o.title)?(v(),d("div",{key:0,class:y(`${t(o.title)} w-3.5 h-3.5 relative`)},null,2)):f("v-if",!0)}});export{a as _};
