+++ 0.1.11/slides/2026-introducing-metaxy/dist/assets/deep-Bf6extjE.js
@@ -1,0 +1,1 @@
+import{f as o,o as s,j as t}from"./modules/vue-BtB2w7oG.js";import{u as l}from"./slidev/context-DXplRwSc.js";const r={class:"slide-deep h-full w-full overflow-hidden"},c={__name:"deep",setup(a){return l(),(e,n)=>(s(),o("div",r,[t(e.$slots,"default")]))}};export{c as _};
