+++ 0.1.11/slides/2026-introducing-metaxy/dist/assets/slidev/useSlideInfo-BpiuJ4iL.js
@@ -1,0 +1,1 @@
+import{r as o,z as f,A as u}from"../modules/vue-BtB2w7oG.js";import{g as s}from"../index-CsGYBKJ1.js";function c(n){return{info:o(s(n)?.meta.slide??null),update:async()=>{}}}const l={};function p(n){function t(e){return l[e]??=c(e)}return{info:f({get(){return t(u(n)).info.value},set(e){t(u(n)).info.value=e}}),update:async(e,r)=>{const i=t(r??u(n)),a=await i.update(e);return a&&(i.info.value=a),a}}}export{c as a,p as u};
