+++ 0.1.11/slides/2026-introducing-metaxy/dist/assets/slidev/DrawingLayer-B55-Qu78.js
@@ -1,0 +1,1 @@
+import{d as u,H as l,r as m,P as c,T as i,f as p,o as f,n as d,A as t}from"../modules/vue-BtB2w7oG.js";import{u as _}from"./SlideWrapper-BlZCqid9.js";import{u as v}from"./context-DXplRwSc.js";import"../index-CsGYBKJ1.js";import"../modules/shiki-qexTPG9f.js";const E=u({__name:"DrawingLayer",setup(g){const{drauu:e,drawingEnabled:o,loadCanvas:s}=_(),r=v().$scale,n=m();return l(()=>{e.mount(n.value,n.value.parentElement),c(r,a=>e.options.coordinateScale=1/a,{immediate:!0}),s()}),i(()=>{e.unmount()}),(a,w)=>(f(),p("svg",{ref_key:"svg",ref:n,class:d(["w-full h-full absolute top-0",{"pointer-events-none":!t(o),"touch-none":t(o)}])},null,2))}});export{E as default};

[…]
