### 2026-06-13 df56a14
chore(publication): fix references (#1222)
## Summary

addresses https://github.com/openjournals/joss-reviews/issues/10449#issuecomment-4699296401
