+++ publications/2026-introducing-metaxy/paper.bib
@@ -46,7 +46,7 @@
 
 @inproceedings{hamilton,
   author    = {Krawczyk, Stefan and Gilani, Elijah},
-  title     = {Hamilton: A Micro-Framework for Creating Dataflows from Python Functions},
+  title     = {Hamilton: A Micro-Framework for Creating Dataflows from {Python} Functions},
   booktitle = {Proceedings of the 23rd Python in Science Conference (SciPy 2024)},
   year      = {2024},
   doi       = {10.25080/gerudo-f2bc6f59-014},
