### 2026-06-14 #10449 @geoHeil
> [@geoHeil](https://github.com/geoHeil) Zenodo DOI (zenodo.20683249) you shared currently shows `v0.1.11-r2` as release tag which I don't see corresponding tag in Github.
>
> I will suggest to keep release tag v0.1.11 and Zenodo DOI (zenodo.20623597) as there is no new code/software change - only change in paper.
>
> Let me know your thoughts.

agreed! LGTM
