### 2026-07-17 90d7f6a
docs(duckdb): show preinstalled extensions
