### 2026-07-17 d36450e
feat(core): add metadata query format parser
