### 2026-07-17 8145a84
feat(core): make lifecycle system columns configurable
Allow configuring creation, update, and soft-deletion column names while preserving existing defaults. Thread the configured names through metadata stores, versioning, integrations, and testing helpers.

Closes #866
