### 2026-07-17 437d1e4
feat(core): attach hash settings to feature definitions (#114)
