### 2026-07-21 0ec8945
ci: run integrations with matrix Python
