### 2026-07-21 8d06afe
Address upstream compatibility review feedback
Signed-off-by: Floze <88098863+floze-the-genius@users.noreply.github.com>
