### 2026-07-21 0fc3ca0
ci: test latest upstream dependencies periodically
