+++ .github/workflows/_qa.yml
@@ -31,6 +31,10 @@ on:
         description: "Run dependency version compatibility tests"
         type: boolean
         default: false
+      upgrade-dependencies:
+        description: "Resolve the newest dependency versions allowed by pyproject.toml"
+        type: boolean
+        default: false
       use-tach:
         description: "Use tach to skip unaffected tests"
         type: boolean
@@ -168,6 +172,8 @@ jobs:
         uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
+        with:
+          upgrade-dependencies: ${{ inputs.upgrade-dependencies }}
       - name: Run Tach
         run: uv run tach check
 
@@ -213,6 +219,8 @@ jobs:
         uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
+        with:
+          upgrade-dependencies: ${{ inputs.upgrade-dependencies }}
       - name: Run ty
         run: uv run ty check
 
@@ -241,6 +249,7 @@ jobs:
         uses: ./.github/actions/setup-devenv
         with:
           extras: "duckdb delta"
+          upgrade-dependencies: ${{ inputs.upgrade-dependencies }}
 
       - name: Run Tests with Coverage
         id: pytest
@@ -301,6 +310,7 @@ jobs:
         uses: ./.github/actions/setup-devenv
         with:
           install-deps: "minimal"
+          upgrade-dependencies: ${{ inputs.upgrade-dependencies }}
 
       - name: Run Tests with Coverage
         env:
@@ -378,6 +388,7 @@ jobs:
         uses: ./.github/actions/setup-devenv
         with:
           install-deps: "minimal"
+          upgrade-dependencies: ${{ inputs.upgrade-dependencies }}
       - name: Run ${{ matrix.session }}
         run: uv run nox -s "${{ matrix.session }}"
 
@@ -491,6 +502,8 @@ jobs:
         uses: actions/checkout@v6.0.2
       - name: Setup devenv
         uses: ./.github/actions/setup-devenv
+        with:
+          upgrade-dependencies: ${{ inputs.upgrade-dependencies }}
       - name: Setup Bun
         uses: oven-sh/setup-bun@v2.1.3
         with:
