+++ docs/index.md
@@ -15,12 +15,18 @@ description: "A high level introduction to Metaxy."
   <a href="https://pypi.org/project/metaxy/"><img src="https://img.shields.io/pypi/dm/metaxy.svg?color=4644ad" alt="PyPI downloads"></a>
   <a href="https://github.com/anam-org/metaxy/actions/workflows/main.yml"><img src="https://github.com/anam-org/metaxy/actions/workflows/main.yml/badge.svg" alt="CI"></a>
   <a href="https://codecov.io/gh/anam-org/metaxy"><img src="https://codecov.io/gh/anam-org/metaxy/graph/badge.svg" alt="codecov"></a>
+  <a href="https://joss.theoj.org/papers/10.21105/joss.10449"><img src="https://joss.theoj.org/papers/10.21105/joss.10449/status.svg" alt="JOSS paper"></a>
   <a href="https://docs.astral.sh/ruff/"><img src="https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json" alt="Ruff"></a>
   <a href="https://docs.astral.sh/ty/"><img src="https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ty/main/assets/badge/v0.json" alt="Ty"></a>
   <a href="https://prek.j178.dev"><img src="https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/j178/prek/master/docs/assets/badge-v0.json" alt="prek"></a>
 </p>
 
 ---
+
+!!! success "Published in JOSS"
+
+    Metaxy is published in the [Journal of Open Source Software](https://joss.theoj.org/papers/10.21105/joss.10449).
+    Read the [docs copy of the paper](./publications/2026-introducing-metaxy/paper.md).
 
 ## Metaxy
 
