### 2026-07-21 03a42e3
ci: add labeled latest-deps matrix
