+++ .github/workflows/Pull Request.yml
@@ -43,3 +43,16 @@ jobs:
     secrets:
       CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
       MOTHERDUCK_TOKEN: ${{ secrets.MOTHERDUCK_TOKEN }}
+
+  qa-latest-deps:
+    if: ${{ github.event_name == 'pull_request' && contains(github.event.pull_request.labels.*.name, 'tests:latest-deps') }}
+    uses: ./.github/workflows/_qa.yml
+    with:
+      python-versions: ${{ contains(github.event.pull_request.labels.*.name, 'tests:full') && '["3.10", "3.11", "3.12", "3.13", "3.14"]' || '["3.10"]' }}
+      run-windows: ${{ contains(github.event.pull_request.labels.*.name, 'tests:full') }}
+      run-macos: ${{ contains(github.event.pull_request.labels.*.name, 'tests:full') }}
+      dependency-matrix: ${{ contains(github.event.pull_request.labels.*.name, 'tests:full') }}
+      test-latest-dependencies: true
+    secrets:
+      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      MOTHERDUCK_TOKEN: ${{ secrets.MOTHERDUCK_TOKEN }}
