### 2026-07-22 1e3b54d
chore(ci): unblock latest-deps compatibility checks
Signed-off-by: Floze <88098863+floze-the-genius@users.noreply.github.com>
