### 2026-07-23 2116499
ci: isolate latest-dependency service secrets
