### 2026-07-23 e0cd67c
ci: make compatibility incidents stateful
Signed-off-by: Floze <88098863+floze-the-genius@users.noreply.github.com>
