### 2026-07-23 aab24fd
fix: support latest Polars and pytest
Signed-off-by: Floze <88098863+floze-the-genius@users.noreply.github.com>
