+++ tests/metaxy_testing/src/metaxy_testing/runbook.py
@@ -216,7 +216,7 @@ class Runbook(BaseModel):
     _execution_state: RunbookExecutionState | None = PrivateAttr(default=None)
 
     def set_execution_state(self, state: RunbookExecutionState) -> None:
-        self._execution_state = state
+        self._execution_state = state  # ty: ignore[invalid-assignment]
 
     @property
     def execution_state(self) -> RunbookExecutionState | None:
