### 2026-07-25 e4bf253
chore(dependencies): do not leak pytest-cases to runtime deps (#1234)
## Summary

move pytest-cases to dev dependencies

## Changelog

do not leak pytest-cases to runtime deps

## Test Plan

standard CI pipeline
