### 2026-08-08 da6e365
feat: add metaxy-rivers integration
Adds a Rivers integration mirroring the existing Dagster one:

- metaxify(): decorator wiring Metaxy feature info into a Rivers
  @Asset (name, code_version, metadata). Column schema/lineage
  injection and key_prefix remapping are Dagster-only concepts with
  no confirmed Rivers equivalent yet, so they are not ported.
- MetaxyIOHandler: reads/writes to a Metaxy MetadataStore via
  Rivers' BaseIOHandler (load_input/handle_output). Per-run
  materialization tracking is not reproducible since Rivers'
  Resource.setup() has no run-id argument at setup time.
- M
[…]
