+++ src/metaxy/ext/rivers/__init__.py
@@ -1,0 +1,12 @@
+from metaxy.ext.rivers.io_handler import MetaxyIOHandler, MetaxyOutput
+from metaxy.ext.rivers.metaxify import metaxify
+from metaxy.ext.rivers.resources import MetaxyResource
+from metaxy.ext.rivers.utils import build_feature_info_metadata
+
+__all__ = [
+    "metaxify",
+    "MetaxyResource",
+    "MetaxyIOHandler",
+    "MetaxyOutput",
+    "build_feature_info_metadata",
+]
