### 2026-08-18 4f6a897
feat: enable `Map` datatype by default
