+++ publications/2026-introducing-metaxy/paper.md
@@ -28,9 +28,7 @@ date: 2026-02-16
 bibliography: paper.bib
 ---
 
-!!! "Published in JOSS"
-
-    This paper is published in the [Journal of Open Source Software](https://joss.theoj.org/papers/10.21105/joss.10449).
+This paper is published in the [Journal of Open Source Software](https://joss.theoj.org/papers/10.21105/joss.10449).
 
 # Summary
 
