+++ tests/metadata_stores/core/test_rebase.py
@@ -19,9 +19,13 @@ from metaxy import (
 from metaxy.config import MetaxyConfig
 from metaxy.ext.polars.handlers.delta import DeltaMetadataStore
 from metaxy.metadata_store.system.storage import SystemTableStorage
+from metaxy.utils import collect_to_polars
+from metaxy.utils._arrow_map import convert_structs_to_maps
 from metaxy_testing.models import SampleFeatureSpec
 
-from metaxy_testing import TempFeatureModule, add_metaxy_provenance_column
+from metaxy_testing import TempFeatureModule, add_metaxy_provenance_column, by_field_maps_to_structs
+
+BY_FIELD_COLUMNS = ["metaxy_provenance_by_field", "metaxy_data_version_by_field"]
 
 UPSTREAM_KEY = FeatureKey(["test", "upstream"])
 DOWNSTREAM_KEY = FeatureKey(["test", "downstream"])
@@ -99,7 +103,7 @@ def test_rebase_updates_provenance(store: MetadataStore, make_graph: list[TempFe
         store.write(UPSTREAM_KEY, upstream_data)
 
         increment = store.resolve_update(DOWNSTREAM_KEY)
-        store.write(DOWNSTREAM_KEY, increment.new.to_polars())
+        store.write(DOWNSTREAM_KEY, collect_to_polars(increment.new))
 
         old_feature_version = temp_module_v1.graph.get_feature_version(DOWNSTREAM_KEY)
 
@@ -132,16 +136,14 @@ def test_rebase_updates_provenance(store: MetadataStore, make_graph: list[TempFe
             to_feature_version=new_feature_version,
         )
 
-        store.write(DOWNSTREAM_KEY, rebased.to_native(), preserve_feature_version=True)
+        store.write(DOWNSTREAM_KEY, collect_to_polars(rebased), preserve_feature_version=True)
 
-        result = (
+        result = collect_to_polars(
             store.read(
                 DOWNSTREAM_KEY,
                 with_feature_history=True,
                 filters=[nw.col("metaxy_feature_version") == new_feature_version],
             )
-            .collect()
-            .to_polars()
         )
 
         assert result.height == 3
@@ -152,7 +154,8 @@ def test_rebase_updates_provenance(store: MetadataStore, make_graph: list[TempFe
         assert result["metaxy_project_version"].unique().to_list() == [expected_project_version]
 
         # provenance_by_field should have entries for all downstream fields
-        for row_pbf in result["metaxy_provenance_by_field"].to_list():
+        result_structs = by_field_maps_to_structs(result, ["metaxy_provenance_by_field"])
+        for row_pbf in result_structs["metaxy_provenance_by_field"].to_list():
             for field_key in expected_version_by_field:
                 assert field_key in row_pbf
 
@@ -200,10 +203,10 @@ def test_rebase_recalculates_provenance_by_field(tmp_path: Path, make_graph: lis
         store.write(UPSTREAM_KEY, upstream_data)
 
         increment = store.resolve_update(DOWNSTREAM_KEY)
-        store.write(DOWNSTREAM_KEY, increment.new.to_polars())
+        store.write(DOWNSTREAM_KEY, collect_to_polars(increment.new))
 
         old_feature_version = temp_v1.graph.get_feature_version(DOWNSTREAM_KEY)
-        old_pbf = increment.new.to_polars()["metaxy_provenance_by_field"].to_list()
+        old_pbf = by_field_maps_to_structs(collect_to_polars(increment.new))["metaxy_provenance_by_field"].to_list()
 
         storage = SystemTableStorage(store)
         storage.push_graph_snapshot()
@@ -244,22 +247,20 @@ def test_rebase_recalculates_provenance_by_field(tmp_path: Path, make_graph: lis
         )
 
[…]
